### 2025-12-01 8faec37
DOC: update docstrings (gp, md)
