### 2025-12-01 0691520
MAINT: switch to nb-clean
