+++ doc/conf.py
@@ -40,6 +40,7 @@ html_theme = 'sphinx_book_theme'
 
 # Raise Error (not default Warning) when a notebook execution fails
 # (due to code error, timeout, etc.)
+# nb_execution_mode = "off"
 nb_execution_raise_on_error = False
 nb_execution_show_tb = True
 nb_execution_timeout = 300
