+++ .pre-commit-config.yaml
@@ -1,5 +1,5 @@
 repos:
-  - repo: https://github.com/kynan/nbstripout
-    rev: 0.8.1
+  - repo: https://github.com/srstevenson/nb-clean
+    rev: 4.0.1
     hooks:
-      - id: nbstripout
+      - id: nb-clean
