### 2025-12-01 0c7b341
MAINT: remove notebook metadata
