+++ _sources/api/summary/GaPFlow.md.GoldAlkane.rst
@@ -1,0 +1,36 @@
+﻿GaPFlow.md.GoldAlkane
+=====================
+
+.. currentmodule:: GaPFlow.md
+
+.. autoclass:: GoldAlkane
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~GoldAlkane.__init__
+      ~GoldAlkane.build_input_files
+      ~GoldAlkane.read_output
+      ~GoldAlkane.run
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~GoldAlkane.dtool_basepath
+      ~GoldAlkane.name
+      ~GoldAlkane.params
+      ~GoldAlkane.main_file
+      ~GoldAlkane.num_worker
+      ~GoldAlkane.is_mock
+   
+   
