+++ _sources/api/summary/GaPFlow.models.WallStress.rst
@@ -1,0 +1,69 @@
+﻿GaPFlow.models.WallStress
+=========================
+
+.. currentmodule:: GaPFlow.models
+
+.. autoclass:: WallStress
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~WallStress.__init__
+      ~WallStress.init
+      ~WallStress.init_database
+      ~WallStress.predict
+      ~WallStress.update
+      ~WallStress.write
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~WallStress.Xtest
+      ~WallStress.Xtrain
+      ~WallStress.Yerr
+      ~WallStress.Yscale
+      ~WallStress.Ytrain
+      ~WallStress.cumtime_infer
+      ~WallStress.cumtime_train
+      ~WallStress.database
+      ~WallStress.dp_dx
+      ~WallStress.dp_dy
+      ~WallStress.extra
+      ~WallStress.full
+      ~WallStress.height
+      ~WallStress.height_and_slopes
+      ~WallStress.kernel_lengthscale
+      ~WallStress.kernel_variance
+      ~WallStress.last_fit_train_size
+      ~WallStress.lower
+      ~WallStress.obs_stddev
+      ~WallStress.pressure
+      ~WallStress.solution
+      ~WallStress.trusted
+      ~WallStress.upper
+      ~WallStress.variance
+      ~WallStress.name
+      ~WallStress.is_gp_model
+      ~WallStress.active_dims
+      ~WallStress.use_active_learning
+      ~WallStress.build_gp
+      ~WallStress.rtol
+      ~WallStress.atol
+      ~WallStress.max_steps
+      ~WallStress.pause_steps
+      ~WallStress.params_init
+      ~WallStress.noise
+      ~WallStress.prop
+      ~WallStress.geo
+   
+   
