+++ _sources/api/summary/GaPFlow.topography.ElasticDeformation.rst
@@ -1,0 +1,26 @@
+﻿GaPFlow.topography.ElasticDeformation
+=====================================
+
+.. currentmodule:: GaPFlow.topography
+
+.. autoclass:: ElasticDeformation
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ElasticDeformation.__init__
+      ~ElasticDeformation.get_G_real
+      ~ElasticDeformation.get_G_real_slices
+      ~ElasticDeformation.get_deformation
+      ~ElasticDeformation.get_deformation_underrelax
+   
+   
+
+   
+   
+   
