+++ _sources/api/summary/GaPFlow.md.Mock.rst
@@ -1,0 +1,36 @@
+﻿GaPFlow.md.Mock
+===============
+
+.. currentmodule:: GaPFlow.md
+
+.. autoclass:: Mock
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~Mock.__init__
+      ~Mock.build_input_files
+      ~Mock.read_output
+      ~Mock.run
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~Mock.dtool_basepath
+      ~Mock.name
+      ~Mock.params
+      ~Mock.main_file
+      ~Mock.num_worker
+      ~Mock.is_mock
+   
+   
