+++ _sources/api/summary/GaPFlow.Database.rst
@@ -1,0 +1,45 @@
+﻿GaPFlow.Database
+================
+
+.. currentmodule:: GaPFlow
+
+.. autoclass:: Database
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~Database.__init__
+      ~Database.add_data
+      ~Database.get_readme_list_local
+      ~Database.get_readme_list_remote
+      ~Database.initialize
+      ~Database.set_training_path
+      ~Database.write
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~Database.X_scale
+      ~Database.Xtrain
+      ~Database.Y_scale
+      ~Database.Ytrain
+      ~Database.Ytrain_err
+      ~Database.config
+      ~Database.has_mock_md
+      ~Database.md_config
+      ~Database.num_features
+      ~Database.output_path
+      ~Database.size
+      ~Database.training_path
+   
+   
