+++ _sources/api/summary/GaPFlow.models.Pressure.rst
@@ -1,0 +1,65 @@
+﻿GaPFlow.models.Pressure
+=======================
+
+.. currentmodule:: GaPFlow.models
+
+.. autoclass:: Pressure
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~Pressure.__init__
+      ~Pressure.init
+      ~Pressure.init_database
+      ~Pressure.predict
+      ~Pressure.update
+      ~Pressure.write
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~Pressure.Xtest
+      ~Pressure.Xtrain
+      ~Pressure.Yerr
+      ~Pressure.Yscale
+      ~Pressure.Ytrain
+      ~Pressure.cumtime_infer
+      ~Pressure.cumtime_train
+      ~Pressure.database
+      ~Pressure.extra
+      ~Pressure.height
+      ~Pressure.height_and_slopes
+      ~Pressure.kernel_lengthscale
+      ~Pressure.kernel_variance
+      ~Pressure.last_fit_train_size
+      ~Pressure.name
+      ~Pressure.obs_stddev
+      ~Pressure.pressure
+      ~Pressure.solution
+      ~Pressure.trusted
+      ~Pressure.v_sound
+      ~Pressure.variance
+      ~Pressure.is_gp_model
+      ~Pressure.active_dims
+      ~Pressure.use_active_learning
+      ~Pressure.build_gp
+      ~Pressure.rtol
+      ~Pressure.atol
+      ~Pressure.max_steps
+      ~Pressure.pause_steps
+      ~Pressure.params_init
+      ~Pressure.noise
+      ~Pressure.prop
+      ~Pressure.geo
+   
+   
