+++ _sources/api/summary/GaPFlow.viz.plotting.rst
@@ -1,0 +1,15 @@
+﻿GaPFlow.viz.plotting
+====================
+
+.. automodule:: GaPFlow.viz.plotting
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      plot_frame
+      plot_frames
+      plot_height
+      plot_history
+   
