### 2025-12-02 cd959d2
MAINT: split MD runner and class definitions
