+++ GaPFlow/md/runner.py
@@ -23,37 +23,13 @@
 #
 import os
 import sys
-import abc
-import shutil
 import warnings
-import dtoolcore
 from mpi4py import MPI
-from datetime import datetime, date
-from dateutil.relativedelta import relativedelta
-from ruamel.yaml import YAML
-from getpass import getuser
-from socket import gethostname
-import scipy.constants as sci
-from copy import deepcopy
 
 try:
     from lammps import lammps
 except ImportError:
     warnings.warn('Failed to import lammps. Only Mock MD object available.')
-
-import jax.random as jr
-import jax.numpy as jnp
-import numpy as np
-
-from ..models.pressure import eos_pressure
-from ..models.viscous import stress_bottom, stress_top
-from .moltemplate import write_template, build_template
-from .utils import read_output_files
-from ..utils import bordered_text, make_dumpable
-
-yaml = YAML()
-yaml.explicit_start = True
-yaml.indent(mapping=4, sequence=4, offset=2)
 
 
 def main():
@@ -91,393 +67,3 @@ def run_serial(fname):
 if __name__ == "__main__":
     # main is called by an individual spawned process for parallel MD runs
     main()
-
-
-class MolecularDynamics:
-    """Driver for molecular dynamics simulations.
-
-    Abstract base class for MD setup, running, and reading outputs.
-    Derived classes need to implement methods to write LAMMPS input files
-    into a dtool dataset, and to read the output of this simulation.
-
-    Attributes
-    ----------
-    name : str
-        Name of the MD object
-    params : dict
-        Parameters to control the MD setup, will be written to the dtool metadata.
-    main_file : str
-        File name of the main LAMMPS input file.
-    num_worker : int
-        Number of cores to run the parallel MD simulation.
-    is_mock : bool
-        Whether the subclass is only a mock object, which does not run an actual MD simulation.
-    """
-    __metaclass__ = abc.ABCMeta
-
-    name = str
-    params: dict
-    main_file: str
-    num_worker: int
-    is_mock: bool
-    _dtool_basepath: str = '/tmp/'
-    _readme_template: str = ""
-    _input_names: list[str] = ['ρ', 'jx', 'jy', 'h', '∂h/∂x', '∂h/∂y'] + [f'extra_{i}' for i in range(10)]
-    _ascii_art: str = r"""
-  _        _    __  __ __  __ ____  ____
- | |      / \  |  \/  |  \/  |  _ \/ ___|
- | |     / _ \ | |\/| | |\/| | |_) \___ \
- | |___ / ___ \| |  | | |  | |  __/ ___) |
- |_____/_/   \_\_|  |_|_|  |_|_|   |____/
-
-"""
-
-    @property
-    def dtool_basepath(self):
-        """File location, where dtool datasets are written into (default is '/tmp/')."""
-        return self._dtool_basepath
-
-    @dtool_basepath.setter
-    def dtool_basepath(self, name):
-        self._dtool_basepath = name
-
-    @abc.abstractmethod
-    def build_input_files(self, dataset, location, X) -> None:
-        """Builds LAMMPS input files based on GP inputs
-         and writes them to a dtool dataset.
-
-        Parameters
-        ----------
-        dataset : dtoolcore.proto_dataset
-            A proto_dataset object.
-        location : str
-            Absolute path of the proto dataset.
-        X : Array
-            Input (i.e. density, gap height, ...)
-        """
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def read_output(self):
-        """Read simulation output and returns observations and their standard error."""
-        raise NotImplementedError
-
-    def _pretty_print(self, proto_datapath, X) -> None:
-        """Print header before the start of a LAMMPS simulation.
-
-        Parameters
-        ----------
[…]
