### 2025-12-02 71dd5f5
MAINT: small modifications moltemplate setup
