### 2025-12-10 6b69f53
ENH: FEM restructuring, added temperature profile and wall heat flux
