### 2025-12-11 58d5664
MAINT: turn to LAMMPS default if MPI grid not specified
