### 2025-12-11 e8f3cbb
MAINT: avoid duplicate YAML key
