### 2025-12-11 9f5ff1f
ENH: possibility to use actual MD inputs as training data instead of target
