+++ GaPFlow/md/utils.py
@@ -22,6 +22,8 @@
 # SOFTWARE.
 #
 import numpy as np
+from scipy.signal import find_peaks
+from scipy.constants import N_A
 
 
 def _get_MPI_grid(Natoms, size, max_cpu, atoms_per_core=1000):
@@ -132,6 +134,54 @@ def read_output_files(fname='stress_wall.dat', sf=1.):
         Yerr[11] = np.sqrt(tauxzU_err)
 
     return Y, Yerr
+
+
+def read_output_files_X(X):
+
+    density_fluid_file = 'profile.dens.fluid.dat'
+    density_solid_file = 'profile.dens.solid.dat'
+    vx_fluid_file = 'profile.vx.fluid.dat'
+    vy_fluid_file = 'profile.vy.fluid.dat'
+
+    _, zdf, _, dens_f = np.loadtxt(density_fluid_file, unpack=True, skiprows=4)
+    _, zds, _, dens_s = np.loadtxt(density_solid_file, unpack=True, skiprows=4)
+    _, zvx, _, vx = np.loadtxt(vx_fluid_file, unpack=True, skiprows=4)
+    _, zvy, _, vy = np.loadtxt(vy_fluid_file, unpack=True, skiprows=4)
+
+    # Actual fluid density
+    density = np.mean(dens_f[dens_f > 0.]) * N_A * 1e-24  # from g/cm^3 to g/mol/A^3
+
+    # Actual gap height
+    # Distance outermost fluid maxima
+    fluid_peaks, _ = find_peaks(dens_f)
+    lower_f = fluid_peaks[0]
+    upper_f = fluid_peaks[-1]
+    dist_fluid = zdf[upper_f] - zdf[lower_f]
+
+    # Distance innermost solid maxima
+    nz = len(zds)
+    solid_peaks_l, _ = find_peaks(dens_s[:nz // 2])
+    solid_peaks_u, _ = find_peaks(dens_s[nz // 2:])
+    zds_l = zds[:nz // 2]
+    zds_u = zds[nz // 2:]
+    dist_solid = zds_u[solid_peaks_u[0]] - zds_l[solid_peaks_l[-1]]
+
+    # Combined
+    gap_height = (dist_fluid + dist_solid) / 2.
+
+    # Actual flux
+    vx_integral = np.trapezoid(vx, zvx)
+    vy_integral = np.trapezoid(vy, zvy)
+    flux_x = vx_integral * density / gap_height
+    flux_y = vy_integral * density / gap_height
+
+    # overwrite input data
+    X.at[0].set(density)
+    X.at[1].set(flux_x)
+    X.at[2].set(flux_y)
+    X.at[3].set(gap_height)
+
+    return X
 
 
 def autocorr_func_1d(x):
