### 2025-12-11 88ac979
MAINT: avoid prime number of cpus
