+++ GaPFlow/md/utils.py
@@ -60,6 +60,23 @@ def _get_MPI_grid(Natoms, size, max_cpu, atoms_per_core=1000):
     return (nx, ny, nz)
 
 
+def is_prime(n: int) -> bool:
+    """Return True if n is a prime number, otherwise False."""
+    if n <= 1:
+        return False
+    if n <= 3:
+        return True
+    if n % 2 == 0 or n % 3 == 0:
+        return False
+
+    i = 5
+    while i * 2 <= n:   # equivalent to i <= sqrt(n) but avoids importing math
+        if n % i == 0 or n % (i + 2) == 0:
+            return False
+        i += 6
+    return True
+
+
 def read_output_files(fname='stress_wall.dat', sf=1.):
 
     md_data = np.loadtxt(fname) * sf
