### 2025-12-12 62ae914
MAINT: more sanity checks on processor count
