+++ GaPFlow/md/moltemplate.py
@@ -27,7 +27,7 @@ import subprocess
 import scipy.constants as sci
 from ase.lattice.cubic import FaceCenteredCubic
 
-from .utils import is_prime
+from .utils import sanitize_num_cpus
 
 
 def write_init(preset="TraPPE", **kwargs):
@@ -680,12 +680,8 @@ def write_template(args, template_dir='moltemplate_files', output_dir="moltempla
     if mpi_grid is None:
         # Let LAMMPS decide MPI grid
         mpi_grid = ['*', '*', '*']
-
-        # but avoid large prime numbers
-        num_cpu = min(max_cpu, Natoms // atoms_per_core)
-        while num_cpu > 5 and is_prime(num_cpu):
-            num_cpu -= 1
-
+        # but use reasonable number of cores
+        num_cpu = sanitize_num_cpus(Natoms // atoms_per_core, max_cpu)
     else:
         num_cpu = np.prod(mpi_grid)
         assert num_cpu <= max_cpu
