### 2025-12-12 2f57be8
MAINT: wrong syntax for jax in-place modifications
