### 2025-12-12 40c0acc
MAINT: make z profile sampling independent from stress sampling
