### 2025-12-15 a057bd6
ENH: several clean-ups, get_T_z helper, print_header fix
