+++ GaPFlow/models/energy.py
@@ -59,11 +59,14 @@ class Energy():
         - T_wall: wall temperature (for simple model)
         - h_Robin: effective heat transfer coefficient
         - alpha_wall: wall heat transfer coefficient (for simple model)
+        - T_0: tuple, initial temperature profile specification
+        - E_0: float or None, direct energy specification (alternative to T_0)
     """
 
     def __init__(self,
                  fc: Any,
-                 energy_spec: dict
+                 energy_spec: dict,
+                 grid: dict
                  ) -> None:
 
         self.__field = fc.register_real_field('total_energy')
@@ -78,11 +81,25 @@ class Energy():
         self.cv = energy_spec['cv']
 
         self.wall_flux_model = energy_spec['wall_flux_model']
-        self.T_wall = energy_spec.get('T_wall', 300.0)
+        self.T_wall = energy_spec['T_wall']
         self.h_Robin = energy_spec['h_Robin']
-        self.alpha_wall = energy_spec.get('alpha_wall', 0.0)
+        self.alpha_wall = energy_spec['alpha_wall']
+
+        # Initial condition specification
+        self.T_0_spec = energy_spec['T0']
+
+        # Boundary conditions (x-direction)
+        self.bc_xW = energy_spec.get('bc_xW', 'P')
+        self.bc_xE = energy_spec.get('bc_xE', 'P')
+        self.T_bc_xW = energy_spec.get('T_bc_xW', self.T_wall)
+        self.T_bc_xE = energy_spec.get('T_bc_xE', self.T_wall)
 
         self.__solution = fc.get_real_field('solution')
+        self.__x = fc.get_real_field('x')
+        self.dim = grid['dim']
+        self.Lx = grid['Lx']
+        self.dx = grid['dx']
+        self._flow_periodic_x = all(grid['bc_xW_P']) and all(grid['bc_xE_P'])
 
         # Initialize bulk temperatures with T_wall as default
         self.__Tb_top.p[:] = self.T_wall
@@ -133,37 +150,129 @@ class Energy():
         """Return full solution field."""
         return self.__solution.p
 
+    def initialize(self) -> None:
+        """
+        Initialize energy field from T_0 specification.
+        """
+        # Check dimensionality for non-uniform profiles
+        if self.T_0_spec[0] in ('half_sine', 'block') and self.dim != 1:
+            raise ValueError(
+                f"Temperature profile '{self.T_0_spec[0]}' is only supported for 1D problems. "
+                f"Got dim={self.dim}. Use 'uniform' profile for 2D problems."
+            )
+
+        # Check: flow periodic but energy BC non-periodic
+        if self._flow_periodic_x and (self.bc_xW != 'P' or self.bc_xE != 'P'):
+            raise ValueError(
+                "Flow is periodic in x but energy BC are non-periodic."
+            )
+
+        # Compute initial energy from initial temperature profile
+        ux = self.solution[1] / self.solution[0]
+        uy = self.solution[2] / self.solution[0]
+        kinetic_energy = 0.5 * (ux**2 + uy**2)
+        T_profile = self._get_T_profile(self.__x.p)
+        self.energy[:] = self.solution[0] * (self.cv * T_profile + kinetic_energy)
+
+        # Update temperature field for consistency
+        self.update_temperature()
+
+    def _get_T_profile(self, x: NDArray) -> NDArray:
+        """
+        Compute initial temperature profile based on specification.
+        """
+        if self.T_0_spec[0] == 'uniform':
+            T = self.T_0_spec[1]
[…]
