+++ setup.cfg
@@ -1,4 +1,4 @@
 [flake8]
 max_line_length = 120
-ignore = E731, W503
+ignore = E731
 exclude = lammps

[…]
