### 2025-12-15 c288c36
WIP: introduced domain decomposition, refactoring .p to .pg, pending muFFT compatibility
