+++ GaPFlow/parallel.py
@@ -1,0 +1,336 @@
+#
+# Copyright 2025 Hannes Holey
+#           2025 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+
+from mpi4py import MPI
+import numpy as np
+import numpy.typing as npt
+
+from typing import TYPE_CHECKING, Any
+if TYPE_CHECKING:
+    from .problem import Problem
+
+from muGrid import (
+    CartesianDecomposition,
+    GlobalFieldCollection,
+    Communicator,
+)
+
+
+class DomainDecomposition:
+    """
+    Manages domain decomposition for MPI-parallel simulations.
+
+    Parameters
+    ----------
+    grid : dict
+        Grid configuration containing:
+        - Nx, Ny: number of grid points
+        - bc_xE, bc_xW, bc_yS, bc_yN: boundary conditions per component
+        - nb_ghost: ghost cell width (default: 1)
+    comm : MPI.Comm, optional
+        MPI communicator (default: MPI.COMM_WORLD)
+    """
+
+    def __init__(self, grid: dict):
+
+        self.grid = grid
+        self._mpi_comm = MPI.COMM_WORLD
+        self._comm = Communicator(self._mpi_comm)
+
+        # Grid dimensions
+        self._Nx = grid['Nx']
+        self._Ny = grid['Ny']
+        self._nb_domain_grid_pts = (self._Nx, self._Ny)
+
+        # Ghost cells
+        nb_ghost = 1
+        self._nb_ghost_left = (nb_ghost, nb_ghost)
+        self._nb_ghost_right = (nb_ghost, nb_ghost)
+
+        # Subdivision
+        self._nb_subdivisions = (1, self._comm.size)
+
+        # Create CartesianDecomposition
+        self._decomp = CartesianDecomposition(
+            self._comm,
+            list(self._nb_domain_grid_pts),
+            list(self._nb_subdivisions),
+            list(self._nb_ghost_left),
+            list(self._nb_ghost_right),
+        )
+
+    def get_fc(self) -> GlobalFieldCollection:
+        """
+        Create a GlobalFieldCollection on the domain decomposition.
+
+        Returns
+        -------
+        muGrid.GlobalFieldCollection
+            The global field collection associated with the domain decomposition.
+        """
+        return self._decomp.collection
+
+    # ---------------------------
+    # MPI properties
+    # ---------------------------
+
+    @property
[…]
