### 2025-12-17 #9645 @editorialbot
[REVIEW]: GaPFlow: Gap-averaged flow simulations with Gaussian process regression
**Submitting author:** @hannes-holey (<a href="http://orcid.org/0000-0002-4547-8791">Hannes Holey</a>)
**Repository:** https://github.com/hannes-holey/GaPFlow
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
