### 2025-12-17 ebdc633
MAINT: simplify multi-output shear stress GP
