+++ GaPFlow/models/gp.py
@@ -26,7 +26,7 @@ import warnings
 import numpy as np
 from copy import deepcopy
 from datetime import datetime
-from typing import Tuple, List
+from typing import Tuple
 
 import jax
 import jax.numpy as jnp
@@ -39,55 +39,6 @@ with warnings.catch_warnings():
 from tinygp import GaussianProcess, kernels, transforms
 
 JAXArray = jax.Array
-
-
-class MultiOutputKernel(kernels.Kernel):
-    """
-    Multi-output Gaussian process kernel.
-
-    Combines multiple latent kernels and a projection matrix to handle
-    vector-valued (multi-output) functions.
-
-    Parameters
-    ----------
-    kernels : list of tinygp.kernels.Kernel or tinygp.transforms.Linear
-        List of kernels corresponding to latent GPs.
-    projection : jax.Array, shape (num_outputs, num_latents)
-        Projection matrix mapping latent processes to outputs.
-
-    Methods
-    -------
-    evaluate(X1, X2)
-        Evaluate the kernel covariance between input sets `X1` and `X2`.
-    """
-
-    kernels: List[kernels.Kernel | transforms.Linear]
-    projection: JAXArray  # shape = (num_outputs, num_latents)
-
-    def evaluate(self,
-                 X1: Tuple[JAXArray, JAXArray],
-                 X2: Tuple[JAXArray, JAXArray],
-                 ) -> JAXArray:
-        """
-        Compute the covariance matrix between two sets of inputs.
-
-        Parameters
-        ----------
-        X1 : tuple
-            Tuple `(t1, idx1)` where `t1` is input array and `idx1`
-            indexes the corresponding outputs.
-        X2 : tuple
-            Tuple `(t2, idx2)` analogous to `X1`.
-
-        Returns
-        -------
-        jax.Array
-            Covariance matrix computed via the kernel projections.
-        """
-        t1, idx1 = X1
-        t2, idx2 = X2
-        latents = jnp.stack([k.evaluate(t1, t2) for k in self.kernels])
-        return (self.projection[idx1] * self.projection[idx2]) @ latents
 
 
 class GaussianProcessSurrogate:
@@ -107,7 +58,6 @@ class GaussianProcessSurrogate:
     is_gp_model: bool
     active_dims: list[int]
     use_active_learning: bool
-    build_gp: callable
     rtol: float
     atol: float
     max_steps: int
@@ -279,9 +229,14 @@ class GaussianProcessSurrogate:
             self.extra
         ]).reshape(self._database.num_features, -1).T
 
+    @property
+    def has_multi_output(self):
+        return self.Ytrain.ndim > 1
+
     # ------------------------------------------------------------------
     # Logging and summary
     # ------------------------------------------------------------------
+
     def write(self) -> None:
         """Log current GP hyperparameters and diagnostics."""
         if self.is_gp_model:
@@ -308,6 +263,28 @@ class GaussianProcessSurrogate:
     # ------------------------------------------------------------------
     # Training and Inference
     # ------------------------------------------------------------------
+    def build_gp(self,
+                 params: dict,
+                 X: JAXArray,
+                 yerr: float | JAXArray) -> GaussianProcess:
+        """Default GP build method.
+
+        Parameters
+        ----------
+        params : dict
+            Dictionary with hyperparameters
+        X : jax.Array
+            Input data.
+        yerr : jax.Array
[…]
