+++ GaPFlow/models/stress.py
@@ -27,10 +27,9 @@ import numpy.typing as npt
 import jax.numpy as jnp
 from jax import vmap, grad
 from jax import Array
-from typing import Optional, Tuple, Any
+from typing import Optional, Any
 
 from .gp import GaussianProcessSurrogate
-from .gp import multi_in_single_out, multi_in_multi_out
 from .pressure import eos_pressure
 from .viscous import stress_bottom, stress_top, stress_avg
 from .sound import eos_sound_velocity
@@ -100,7 +99,6 @@ class WallStress(GaussianProcessSurrogate):
             self.pause_steps = gp['pause_steps']
             self.is_gp_model = True
             self.use_active_learning = gp['active_learning']
-            self.build_gp = multi_in_multi_out
         else:
             self.is_gp_model = False
             self.use_active_learning = False
@@ -191,41 +189,14 @@ class WallStress(GaussianProcessSurrogate):
         return np.gradient(self.pressure, self.__y.p[0, :], axis=1)
 
     @property
-    def Xtest(self) -> Tuple[JAXArray, JAXArray]:
-        """
-        Test inputs for GP prediction.
-
-        Returns
-        -------
-        tuple of jax.Array
-            (X, flag) where X contains duplicated and scaled features and flag
-            indicates which output (lower/upper) the sample corresponds to.
-        """
-        X = jnp.concatenate([(self._Xtest / self.database.X_scale)[:, self.active_dims],
-                             (self._Xtest / self.database.X_scale)[:, self.active_dims]])
-
-        flag = jnp.concatenate([jnp.zeros(X.shape[0] // 2, dtype=int),
-                                jnp.ones(X.shape[0] // 2, dtype=int)])
-
-        return X, flag
+    def Xtest(self) -> JAXArray:
+        """Test inputs for shear stress GP (normalized)."""
+        return (self._Xtest / self.database.X_scale)[:, self.active_dims]
 
     @property
-    def Xtrain(self) -> Tuple[JAXArray, JAXArray]:
-        """
-        Training inputs for GP (duplicated to account for two outputs).
-
-        Returns
-        -------
-        tuple of jax.Array
-            (X, flag) where X contains duplicated input rows and flag labels.
-        """
-        X = jnp.concatenate([self.database.Xtrain[:, self.active_dims],
-                             self.database.Xtrain[:, self.active_dims]])
-
-        flag = jnp.concatenate([jnp.zeros(X.shape[0] // 2, dtype=int),
-                                jnp.ones(X.shape[0] // 2, dtype=int)])
-
-        return X, flag
+    def Xtrain(self) -> JAXArray:
+        """Training inputs for shear stress GP (normalized)."""
+        return self.database.Xtrain[:, self.active_dims]
 
     @property
     def _Ytrain(self) -> JAXArray:
@@ -237,8 +208,8 @@ class WallStress(GaussianProcessSurrogate):
         jax.Array
             Concatenated array of training outputs (lower then upper).
         """
-        return jnp.concatenate([self.database._Ytrain[:self.last_fit_train_size, self._out_index + 1],
-                                self.database._Ytrain[:self.last_fit_train_size, self._out_index + 7]])
+        return jnp.vstack([self.database._Ytrain[:self.last_fit_train_size, self._out_index + 1],
+                           self.database._Ytrain[:self.last_fit_train_size, self._out_index + 7]]).T
 
     @property
     def Ytrain(self) -> JAXArray:
@@ -278,20 +249,20 @@ class WallStress(GaussianProcessSurrogate):
             Observation noise standard deviation normalized by Yscale.
         """
 
[…]
