### 2025-12-17 hannes-holey/hans#48 @matthewfeickert
[JOSS Review]: Use build system to match software requirement and consider distribution on conda-forge
Related to https://github.com/openjournals/joss-reviews/issues/9645, for JOSS reviews [it is expected that reviewers should be able to install software in full using language specific package management tools](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions), though they are asked to build the software from source to verify that build instructions are correct.

At the moment (3d1d9432bf8d39f5fccb614e42442f0b611698e2), GaPFlow does not provide software distributions on PyPI that have are full builds of the software and it is noted that

https://github.com/hannes-holey/GaPFlow/blob/3d1d9432bf8d39f5fccb614e42442f0b611698e2/README.md#L18-L24

The project should be able to use [`scikit-build-core`](https://github.com/scikit-build/scikit-build-core) as a `build-system`

```toml
[build-system]
requires = ["scikit-build-core"]
build-backend = "scikit_build_core.build"
```

and provide full distributions on PyPI. Software should be fully portable, so manually setting environment variables for runtime use to work

https://github.com/hannes-holey/GaPFlow/blob/3d1d9432bf8d39f5fccb614e42442f0b611698e2/README.md#L46-L54

should be avoided unless absolutely necessary in niche cases.

I'd highly suggest looking at the [Scientific Python Library Development Guide](https://learn.scientific-python.org/development/) for information on this.

Additionally, given some of the difficulties that can be encountered with wheels, it would be quite useful (but not required) for both developers and end users to have GaPFlow also be built and distributed on [conda-forge](https://conda-forge.org/).

If you have questions RE: conda-forge, submitting recipes, or maintenance questions feel free to ask me about this.
