### 2025-12-23 eabcc22
WIP: FEM 2D, first working case
