+++ GaPFlow/fem/utils2d.py
@@ -1,0 +1,216 @@
+#
+# Copyright 2025 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+import numpy as np
+from muGrid import ConvolutionOperator
+
+from typing import Callable
+import numpy.typing as npt
+
+NDArray = npt.NDArray[np.floating]
+
+
+class NonLinearTerm():
+    def __init__(self,
+                 name: str,
+                 description: str,
+                 res: str,
+                 dep_vars: list[str],
+                 dep_vals: list[str],
+                 fun: Callable,
+                 der_funs: list[Callable],
+                 d_dx_resfun: bool = False,
+                 d_dy_resfun: bool = False,
+                 der_testfun: bool = False):
+        self.name = name
+        self.description = description
+        self.res = res
+        self.dep_vars = dep_vars
+        self.dep_vals = dep_vals
+        self.fun_ = fun
+        self.der_funs_ = der_funs
+        self.d_dx_resfun = d_dx_resfun
+        self.d_dy_resfun = d_dy_resfun
+        self.der_testfun = der_testfun
+        self.built = False
+
+    def build(self, ctx: dict) -> None:
+        self.fun = self.fun_(ctx)
+        self.der_funs = [der_fun_(ctx) for der_fun_ in self.der_funs_]
+        self.built = True
+
+    def evaluate(self, *args) -> NDArray:
+        if not self.built:
+            raise Exception("Term not built")
+        return self.fun(*args)
+
+    def evaluate_deriv(self, dep_var: str, *args) -> NDArray:
+        if not self.built:
+            raise Exception("Term not built")
+        i = self.dep_vars.index(dep_var)
+        return self.der_funs[i](*args)
+
+
+def get_N0_left_test_vals():
+    val_1 = 2/3
+    val_2 = 1/6
+    val_3 = 1/6
+    return np.array([val_1, val_2, val_3])
+
+def get_N1_left_test_vals():
+    val_1 = 1/6
+    val_2 = 1/6
+    val_3 = 2/3
[…]
