### 2025-12-23 21154a6
WIP: cleaned up index mask functions
