+++ GaPFlow/solver_fem_2d.py
@@ -30,7 +30,7 @@ from .fem.utils2d import (
 )
 from .fem.utils2d import *
 
-from functools import cached_property
+from functools import cached_property, lru_cache
 import numpy as np
 import time
 
@@ -87,6 +87,7 @@ class FEMSolver2D:
         self.energy = p.fem_solver['equations']['energy']
         self.dynamic = p.fem_solver['dynamic']
 
+        self.decomp = p.decomp
         self.global_coords = p.decomp.icoordsg  # shape (2, Nx_padded, Ny_padded)
 
         nb_subdomain_pts = p.decomp.nb_subdomain_grid_pts
@@ -106,9 +107,11 @@ class FEMSolver2D:
 
         self.variables = ['rho', 'jx', 'jy']
         self.residuals = ['mass', 'momentum_x', 'momentum_y']
+        self.field_map = {'rho': p.q[0], 'jx': p.q[1], 'jy': p.q[2]}
         if self.energy:
             self.variables.append('E')
             self.residuals.append('energy')
+            self.field_map['E'] = p.energy.energy
 
         self.nb_inner_pts = self.Nx_inner * self.Ny_inner
         self.res_size = len(self.residuals) * self.nb_inner_pts
@@ -116,12 +119,18 @@ class FEMSolver2D:
 
         self.mat_size = (self.res_size, self.var_size)
 
-        assert self.nb_inner_pts == len(self.indices_inner_local)
-        assert self.nb_contributors == len(self.indices_padded_local)
-
         self.sq_per_row = self.Nx_inner + 1
         self.sq_per_col = self.Ny_inner + 1
         self.nb_sq = self.sq_per_row * self.sq_per_col
+
+        # Test function values for FEM assembly
+        self.test_fun_vals_left = get_N_left_test_vals()
+        self.test_fun_vals_right = get_N_right_test_vals()
+        self.test_fun_vals_left_dx = get_N_left_test_vals_dx(self.dx)
+        self.test_fun_vals_right_dx = get_N_right_test_vals_dx(self.dx)
+        self.test_fun_vals_left_dy = get_N_left_test_vals_dy(self.dy)
+        self.test_fun_vals_right_dy = get_N_right_test_vals_dy(self.dy)
+        self.quad_weights = get_quad_weights()
 
     def _get_active_terms(self) -> None:
         """Initialize list of active terms from problem fem_solver config."""
@@ -321,220 +330,64 @@ class FEMSolver2D:
     # Solution Vector Management
     # =========================================================================
 
-    def isBC(self, i: int, j: int) -> bool:
-        """Check if the point at (i, j) is a boundary condition point."""
-        if (i == 0 and self.bc_at_S) or (i == self.Ny_padded - 1 and self.bc_at_N):
+    def isBC(self, x: int, y: int) -> bool:
+        """Check if the point at (x, y) is a boundary condition point."""
+        if (x == 0 and self.bc_at_W) or (x == self.Nx_padded - 1 and self.bc_at_E):
             return True
-        if (j == 0 and self.bc_at_W) or (j == self.Nx_padded - 1 and self.bc_at_E):
+        if (y == 0 and self.bc_at_S) or (y == self.Ny_padded - 1 and self.bc_at_N):
             return True
         return False
 
     @cached_property
     def index_mask_inner_local(self) -> NDArray:
-        """2D Index mask for inner points with local indices."""
-        mask = np.full((self.Ny_padded, self.Nx_padded), -1, dtype=int)
-        inner_shape = (self.Ny_inner, self.Nx_inner)
-        mask[1:-1, 1:-1] = np.arange(np.prod(inner_shape)).reshape(inner_shape)
-        return mask
-
-    @cached_property
[…]
