### 2025-12-27 613800e
WIP: First draft PETSc assembly running
