+++ GaPFlow/solver_fem_2d.py
@@ -48,6 +48,7 @@ BASE_FIELDS = {
     'U', 'V', 'Ls',
     'dp_drho',
     'rho_prev', 'jx_prev', 'jy_prev',
+    'pressure_stab',
 }
 
 STRESS_XZ_FIELDS = {
@@ -77,17 +78,18 @@ class FEMSolver2D:
         self.num_solver = Solver(problem.fem_solver)
         self.nodal_fields: Dict[str, Any] = {}
         self.quad_fields: Dict[str, Any] = {}
+        self.petsc = None  # Optional PETSc system, enabled via enable_petsc()
 
     def _init_convenience_accessors(self) -> None:
         """Initialize convenience accessors for problem and grid properties."""
         p = self.problem
+        self.decomp = p.decomp
 
-        self.per_x = p.grid['bc_xE_P'][0]  # TODO periodicity information from decomp
-        self.per_y = p.grid['bc_yN_P'][0]
+        self.per_x = p.decomp.periodic_x
+        self.per_y = p.decomp.periodic_y
         self.energy = p.fem_solver['equations']['energy']
         self.dynamic = p.fem_solver['dynamic']
-
-        self.decomp = p.decomp
+        
         self.global_coords = p.decomp.icoordsg  # shape (2, Nx_padded, Ny_padded)
 
         nb_subdomain_pts = p.decomp.nb_subdomain_grid_pts
@@ -96,10 +98,11 @@ class FEMSolver2D:
         self.Nx_padded = self.Nx_inner + 2
         self.Ny_padded = self.Ny_inner + 2
 
-        self.bc_at_W = p.decomp.is_at_xW
-        self.bc_at_E = p.decomp.is_at_xE
-        self.bc_at_S = p.decomp.is_at_yS
-        self.bc_at_N = p.decomp.is_at_yN
+        # BC flags: True only if at domain edge AND not periodic
+        self.bc_at_W = p.decomp.is_at_xW and not self.per_x
+        self.bc_at_E = p.decomp.is_at_xE and not self.per_x
+        self.bc_at_S = p.decomp.is_at_yS and not self.per_y
+        self.bc_at_N = p.decomp.is_at_yN and not self.per_y
 
         self.dx = p.grid['dx']
         self.dy = p.grid['dy']
@@ -131,6 +134,26 @@ class FEMSolver2D:
         self.test_fun_vals_left_dy = get_N_left_test_vals_dy(self.dy)
         self.test_fun_vals_right_dy = get_N_right_test_vals_dy(self.dy)
         self.quad_weights = get_quad_weights()
+
+        # Precomputed element tensors: elem[i,j,q] = N_i[q] * N_j[q] * w[q] * A
+        # For zero-der terms: M[i,j] += sum_q(elem[i,j,q] * res[q])
+        w = self.quad_weights
+        Nl, Nr = self.test_fun_vals_left, self.test_fun_vals_right
+        self.elem_tensor_left = Nl[:, None, :] * Nl[None, :, :] * w * self.A   # (3,3,3)
+        self.elem_tensor_right = Nr[:, None, :] * Nr[None, :, :] * w * self.A  # (3,3,3)
+
+        # Precomputed test function integrals for derivative terms: integral[i,q] = N_i[q] * w[q] * A
+        self.test_wA_left = Nl * w * self.A       # (3, 3)
+        self.test_wA_right = Nr * w * self.A      # (3, 3)
+        self.test_wA_left_dx = self.test_fun_vals_left_dx * w * self.A
+        self.test_wA_right_dx = self.test_fun_vals_right_dx * w * self.A
+        self.test_wA_left_dy = self.test_fun_vals_left_dy * w * self.A
[…]
