### 2026-01-05 3759047
ENH: full vectorization, sparse matrices, iterative solvers
