+++ GaPFlow/fem/petsc_system.py
@@ -23,232 +23,173 @@
 #
 import numpy as np
 import numpy.typing as npt
+from dataclasses import dataclass
 
 from petsc4py import PETSc
 
-from typing import TYPE_CHECKING
-if TYPE_CHECKING:
-    from ..solver_fem_2d import FEMSolver2D
 
 NDArray = npt.NDArray[np.floating]
 
 
+@dataclass(frozen=True)
+class PETScLayout:
+    """
+    Explicit link between Solver/DomainDecomposition and PETSc:
+    - l2g_pts maps local point indices to global point indices
+    - mat_rows/cols define global matrix structure (COO format)
+    - All arrays computed once at solver initialization.
+
+    Attributes
+    ----------
+    nb_vars : int
+        Number of variables (3 or 4 with energy).
+    nb_inner_pts : int
+        Number of inner points owned by this rank.
+    nb_contributors : int
+        Number of unique contributor indices (inner + ghost).
+    nb_global_pts : int
+        Total global grid points.
+    l2g_pts : NDArray
+        Local-to-global point mapping, shape (nb_contributors,).
+    mat_rows : NDArray
+        Global row indices for each non-zero, shape (nnz,).
+    mat_cols : NDArray
+        Global column indices for each non-zero, shape (nnz,).
+    mat_local_idx : NDArray
+        Local indices [row, col] in M_local for each non-zero, shape (nnz, 2).
+        This attribute is used by sparse_layout to invert the mapping from
+        coo_idx → (local_row, local_col) to (local_row, local_col) → coo_idx.
+    rhs_rows : NDArray
+        Global row indices for RHS assembly, shape (local_size,).
+    """
+    nb_vars: int
+    nb_inner_pts: int
+    nb_contributors: int
+    nb_global_pts: int
+    l2g_pts: NDArray
+    mat_rows: NDArray
+    mat_cols: NDArray
+    mat_local_idx: NDArray
+    rhs_rows: NDArray
+
+
 class PETScSystem:
     """
-    Translates local FEM system to distributed PETSc and solves.
-
-    This class manages:
-    1. Local-to-global index translation
-    2. PETSc matrix/vector creation and assembly
+    Manages distributed PETSc linear system and solves.
+    1. PETSc matrix/vector creation with COO preallocation
+    2. O(nnz) vectorized assembly from COO values
     3. KSP solver configuration and execution
 
-    The translation is built by comparing index_mask_padded_local (local indices)
-    with index_mask_padded_global (global indices) at each grid position.
+    All index translation is handled by PETScLayout.
 
     Parameters
     ----------
-    solver : FEMSolver2D
-        The FEM solver instance containing index masks and problem info.
+    layout : PETScLayout
+        Precomputed index arrays for assembly and solution extraction.
+    solver_type : str, optional
+        Solver type: "direct" (MUMPS LU) or "iterative" (BiCGSTAB + ILU).
+        Default is "direct".
+
+    Notes
+    -----
+    Iterative solver performance (vs MUMPS direct):
+    - Grid ≤160×160: BiCGSTAB + ILU(1) → 4.7-5.3× faster
+    - Grid ≥192×192: BiCGSTAB + ILU(2) → 1.9-4.5× faster
+    ILU fill level is selected automatically based on matrix size.
     """
 
-    def __init__(self, solver: "FEMSolver2D"):
-        self.solver = solver
-        self.decomp = solver.decomp
+    # Threshold for switching from ILU(1) to ILU(2)
+    # ~110k corresponds to 192×192 grid with 3 variables
+    _ILU2_THRESHOLD = 110000
+
+    def __init__(self, layout: PETScLayout, solver_type: str = "direct"):
+        self._layout = layout
[…]
