+++ GaPFlow/fem2d/test_solve_sphere.py
@@ -11,6 +11,11 @@ Run serial:
 Run with MPI:
     mpirun -n 2 python test_solve_sphere.py
 """
+import resource
+# Limit memory to 12GB to prevent OOM crashes in WSL
+_MEM_LIMIT_GB = 12
+resource.setrlimit(resource.RLIMIT_AS, (_MEM_LIMIT_GB * 1024**3, _MEM_LIMIT_GB * 1024**3))
+
 from mpi4py import MPI
 import numpy as np
 import sys
@@ -32,8 +37,8 @@ options:
 grid:
     Lx: 0.1
     Ly: 0.1
-    Nx: 64
-    Ny: 64
+    Nx: 128
+    Ny: 128
     xE: ['D', 'N', 'N']
     xW: ['D', 'N', 'N']
     yS: ['D', 'N', 'N']
@@ -44,7 +49,7 @@ grid:
     yN_D: 1.1853
 
 geometry:
-    type: inclined
+    type: parabolic_2d
     hmax: 6.6e-5
     hmin: 1.0e-5
     U: 50.
@@ -66,6 +71,7 @@ properties:
 
 fem_solver:
     type: newton_alpha
+    linear_solver: direct
     dynamic: True
     R_norm_tol: 1e-11
     pressure_stab_alpha: 500
@@ -74,145 +80,21 @@ fem_solver:
 """
 
 
-def set_spherical_topography(problem, hmin, hmax):
-    """
-    Set a spherical topography with minimum height at center, maximum at corners.
-
-    h(x, y) = hmin + (hmax - hmin) * r^2 / r_max^2
-
-    where r^2 = (x - x_c)^2 + (y - y_c)^2 and r_max^2 = (Lx/2)^2 + (Ly/2)^2
-    """
-    Lx = problem.grid['Lx']
-    Ly = problem.grid['Ly']
-
-    # Get coordinates from topography
-    xx = problem.topo.x
-    yy = problem.topo.y
-
-    # Center of domain
-    x_c = Lx / 2
-    y_c = Ly / 2
-
-    # Maximum radius (corner distance from center)
-    r_max_sq = (Lx / 2)**2 + (Ly / 2)**2
-
-    # Compute distance from center
-    r_sq = (xx - x_c)**2 + (yy - y_c)**2
-
-    # Spherical height profile: minimum at center, maximum at corners
-    h = hmin + (hmax - hmin) * r_sq / r_max_sq
-
-    # Set the height field (gradients are computed automatically via the setter)
-    problem.topo.h = h
-
-    return h
-
-
 def gather_global_solution(problem):
-    """
-    Gather solution from all MPI ranks to rank 0.
-
-    Domain decomposition is along Y-axis only, so each rank has:
-    - Full X dimension (Nx_inner = Nx_global)
-    - Partial Y dimension (local Ny_inner)
-
-    Returns (rho, jx, jy) as full global arrays on rank 0, None on other ranks.
-    """
-    comm = MPI.COMM_WORLD
-    rank = comm.Get_rank()
-    size = comm.Get_size()
-
-    solver = problem.solver
+    """Gather solution fields from all ranks using DomainDecomposition."""
     decomp = problem.decomp
-
-    # Get local solution
-    nb = solver.nb_inner_pts
-    q = solver.get_q_nodal()
-
-    # Local arrays (Ny_local, Nx) - note: reshape is (Ny, Nx) for row-major
-    rho_local = q[0:nb].reshape((solver.Ny_inner, solver.Nx_inner))
-    jx_local = q[nb:2*nb].reshape((solver.Ny_inner, solver.Nx_inner))
[…]
