+++ GaPFlow/fem2d/test_solve_slider.py
@@ -8,21 +8,21 @@ Run serial:
 Run with MPI:
     mpirun -n 2 python test_solve.py
 """
+import os
 import resource
+import sys
+
 # Limit memory to 12GB to prevent OOM crashes in WSL
 _MEM_LIMIT_GB = 12
 resource.setrlimit(resource.RLIMIT_AS, (_MEM_LIMIT_GB * 1024**3, _MEM_LIMIT_GB * 1024**3))
 
-from mpi4py import MPI
-import numpy as np
-import sys
-import os
-
 hans_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.insert(0, hans_dir)
 
-from GaPFlow.problem import Problem
-from GaPFlow.models.pressure import eos_pressure
+import numpy as np  # noqa: E402
+from mpi4py import MPI  # noqa: E402
+from GaPFlow.problem import Problem  # noqa: E402
+from GaPFlow.models.pressure import eos_pressure  # noqa: E402
 
 
 CONFIG = """
@@ -70,7 +70,8 @@ fem_solver:
     R_norm_tol: 1e-11
     pressure_stab_alpha: 500
     equations:
-        term_list: ['R11x', 'R11y', 'R11Sx', 'R11Sy', 'R1Stabx', 'R1Staby', 'R1T', 'R21x', 'R21y', 'R24x', 'R24y', 'R2Tx', 'R2Ty']
+        term_list: ['R11x', 'R11y', 'R11Sx', 'R11Sy', 'R1Stabx', 'R1Staby', 'R1T',
+                    'R21x', 'R21y', 'R24x', 'R24y', 'R2Tx', 'R2Ty']
 """
 
 
@@ -97,8 +98,8 @@ def gather_global_solution(problem):
 
     # Local arrays (Ny_local, Nx) - note: reshape is (Ny, Nx) for row-major
     rho_local = q[0:nb].reshape((solver.Ny_inner, solver.Nx_inner))
-    jx_local = q[nb:2*nb].reshape((solver.Ny_inner, solver.Nx_inner))
-    jy_local = q[2*nb:3*nb].reshape((solver.Ny_inner, solver.Nx_inner))
+    jx_local = q[nb:2 * nb].reshape((solver.Ny_inner, solver.Nx_inner))
+    jy_local = q[2 * nb:3 * nb].reshape((solver.Ny_inner, solver.Nx_inner))
 
     # Global dimensions
     Nx_global, Ny_global = decomp.nb_domain_grid_pts
@@ -171,12 +172,9 @@ def main():
         solver = problem.solver
 
         # Problem parameters
-        P0 = problem.prop['P0']
-        mu = problem.prop['shear']
         U = problem.geo['U']
         hmin = problem.geo['hmin']
         hmax = problem.geo['hmax']
-        Lx = problem.grid['Lx']
         rho0 = problem.prop['rho0']
 
         # For inclined slider bearing (shear-driven), mass flux ~ rho * U * h / 2
@@ -214,8 +212,8 @@ def plot_solution(solver, rho, jx, jy, problem, jx_expected, Nx_global=None, Ny_
     dx, dy = solver.dx, solver.dy
 
     # Grid coordinates (cell centers)
-    x = np.linspace(dx/2, Nx * dx - dx/2, Nx)
-    y = np.linspace(dy/2, Ny * dy - dy/2, Ny)
+    x = np.linspace(dx / 2, Nx * dx - dx / 2, Nx)
+    y = np.linspace(dy / 2, Ny * dy - dy / 2, Ny)
     X, Y = np.meshgrid(x, y)
 
     # Compute pressure from density using the actual EOS from problem
[…]
