+++ GaPFlow/solver_fem_2d.py
@@ -28,12 +28,22 @@ from .fem.utils2d import (
     get_triangle_2_operator_dx,
     get_triangle_2_operator_dy,
 )
-from .fem.utils2d import *
+from .fem.utils2d import (
+    get_N_left_test_vals,
+    get_N_right_test_vals,
+    get_N_left_test_vals_dx,
+    get_N_right_test_vals_dx,
+    get_N_left_test_vals_dy,
+    get_N_right_test_vals_dy,
+    get_quad_weights,
+    get_active_terms,
+)
 
 from dataclasses import dataclass
 from functools import cached_property, lru_cache
 import numpy as np
 import time
+from muGrid import Field, Timer
 
 import numpy.typing as npt
 from typing import TYPE_CHECKING, Tuple, Dict, Any, Optional
@@ -113,6 +123,7 @@ class FEMSolver2D:
         self.nodal_fields: Dict[str, Any] = {}
         self.quad_fields: Dict[str, Any] = {}
         self.petsc = None  # Optional PETSc system, enabled via enable_petsc()
+        self.timer = Timer()
 
     def _init_convenience_accessors(self) -> None:
         """Initialize convenience accessors for problem and grid properties."""
@@ -123,7 +134,7 @@ class FEMSolver2D:
         self.per_y = p.decomp.periodic_y
         self.energy = p.fem_solver['equations']['energy']
         self.dynamic = p.fem_solver['dynamic']
-        
+
         self.global_coords = p.decomp.icoordsg  # shape (2, Nx_padded, Ny_padded)
 
         nb_subdomain_pts = p.decomp.nb_subdomain_grid_pts
@@ -211,8 +222,8 @@ class FEMSolver2D:
             self.nodal_fields[name] = fc.real_field(f'{name}_nodal', 1, 'pixel')
 
         # existing single-component fields
-        self.nodal_fields['p'] = fc.get_real_field('pressure')
-        self.nodal_fields['eta'] = fc.get_real_field('shear_viscosity')
+        self.nodal_fields['p'] = Field(fc.get_real_field('pressure'))
+        self.nodal_fields['eta'] = Field(fc.get_real_field('shear_viscosity'))
 
         # quad output fields
         for name in self._get_needed_quad_fields():
@@ -339,7 +350,7 @@ class FEMSolver2D:
         apply = self._apply_2d_vmap
 
         # Pressure gradient
-        self.quad_fields['dp_drho'].pg = apply(p.pressure.dp_drho, q('rho'))
+        self.quad_fields['dp_drho'].pg[:] = apply(p.pressure.dp_drho, q('rho'))
 
         # Pressure stabilization parameter: tau = alpha * h_elem^2 / P0
         # This controls the strength of the Brezzi-Pitkäranta stabilization
@@ -347,39 +358,42 @@ class FEMSolver2D:
         alpha = p.fem_solver.get('pressure_stab_alpha', 0.0)
         P0 = p.prop.get('P0', 1.0)
         h_elem_sq = self.dx * self.dy  # element area as characteristic length squared
-        self.quad_fields['pressure_stab'].pg = np.full_like(q('h'), alpha * h_elem_sq / P0)
+        self.quad_fields['pressure_stab'].pg[:] = np.full_like(q('h'), alpha * h_elem_sq / P0)
 
         # Wall stress xz
         args_xz = (q('rho'), q('jx'), q('jy'), q('h'), q('dh_dx'), q('U'), q('V'), q('Ls'))
         for name in ['tau_xz', 'dtau_xz_drho', 'dtau_xz_djx',
[…]
