### 2026-01-08 f40e102
MAINT: muGrid dependency update, refactoring in solver_fem_2d
