+++ GaPFlow/solver_fem_2d.py
@@ -653,11 +653,11 @@ class FEMSolver2D:
             for dep_var in term.dep_vars:
                 key = (term.name, term.res, dep_var)
                 if term.d_dx_resfun:
-                    layout[key] = self._build_term_indices_dx(
-                        term, dep_var, local_to_coo)
+                    layout[key] = self._build_term_indices_deriv(
+                        term, dep_var, local_to_coo, 'x')
                 elif term.d_dy_resfun:
-                    layout[key] = self._build_term_indices_dy(
-                        term, dep_var, local_to_coo)
+                    layout[key] = self._build_term_indices_deriv(
+                        term, dep_var, local_to_coo, 'y')
                 else:
                     layout[key] = self._build_term_indices_zero_der(
                         term, dep_var, local_to_coo)
@@ -796,13 +796,19 @@ class FEMSolver2D:
             sq_y=np.concatenate(sq_y_list),
         )
 
-    def _build_term_indices_dx(self, term: NonLinearTerm, dep_var: str,
-                               local_to_coo: NDArray) -> TermIndices:
+    def _build_term_indices_deriv(self, term: NonLinearTerm, dep_var: str,
+                                    local_to_coo: NDArray, direction: str) -> TermIndices:
         """
-        Build indices for x-derivative term (±1/dx stencil contributions).
+        Build indices for derivative term (±1/d* stencil contributions).
 
         Each triangle contributes 3×2 = 6 entries (3 test nodes × 2 stencil pts).
-        Left triangle: stencil (bl-, br+), Right triangle: stencil (tl-, tr+).
+
+        Parameters
+        ----------
+        direction : str
+            'x' or 'y' - determines finite difference stencil direction:
+            - 'x': Left uses (bl-, br+), Right uses (tl-, tr+)
+            - 'y': Left uses (bl-, tl+), Right uses (br-, tr+)
         """
         res_idx = self.residuals.index(term.res)
         var_idx = self.variables.index(dep_var)
@@ -812,15 +818,23 @@ class FEMSolver2D:
 
         # Left triangle test nodes [bl, tl, br] = corners [0, 2, 1]
         TO_left = TO[:, [0, 2, 1]]
-        # Left triangle x-stencil: bl (neg), br (pos)
-        FROM_neg_left = FROM[:, 0]  # bl
-        FROM_pos_left = FROM[:, 1]  # br
-
         # Right triangle test nodes [tr, br, tl] = corners [3, 1, 2]
         TO_right = TO[:, [3, 1, 2]]
-        # Right triangle x-stencil: tl (neg), tr (pos)
-        FROM_neg_right = FROM[:, 2]  # tl
-        FROM_pos_right = FROM[:, 3]  # tr
+
+        # Direction-dependent stencil indices and test function weights
+        # Corner indices: bl=0, br=1, tl=2, tr=3
+        if direction == 'x':
+            # x-derivative: horizontal neighbors
+            FROM_neg_left, FROM_pos_left = FROM[:, 0], FROM[:, 1]    # bl-, br+
+            FROM_neg_right, FROM_pos_right = FROM[:, 2], FROM[:, 3]  # tl-, tr+
+            test_wA_left = self.test_wA_left_dx if term.der_testfun else self.test_wA_left
+            test_wA_right = self.test_wA_right_dx if term.der_testfun else self.test_wA_right
+        else:  # 'y'
+            # y-derivative: vertical neighbors
+            FROM_neg_left, FROM_pos_left = FROM[:, 0], FROM[:, 2]    # bl-, tl+
+            FROM_neg_right, FROM_pos_right = FROM[:, 1], FROM[:, 3]  # br-, tr+
[…]
