### 2026-01-08 8e46655
MAINT: continued refactor of solver_fem_2d, muGrid ci workflow fix
