+++ GaPFlow/solver_fem_2d.py
@@ -24,93 +24,30 @@
 from .fem.num_solver import Solver
 from .fem.utils2d import (
     NonLinearTerm,
-    get_triangle_3_operator,
-    get_triangle_2_operator_dx,
-    get_triangle_2_operator_dy,
-)
-from .fem.utils2d import (
-    get_N_left_test_vals,
-    get_N_right_test_vals,
-    get_N_left_test_vals_dx,
-    get_N_right_test_vals_dx,
-    get_N_left_test_vals_dy,
-    get_N_right_test_vals_dy,
-    get_quad_weights,
+    TriangleQuadrature,
     get_active_terms,
 )
+from .fem.assembly_layout import (
+    FEMAssemblyLayout,
+    MatrixCOOPattern,
+    RHSPattern,
+    JacobianTermMap,
+    RHSTermMap,
+)
+from .fem.grid_index import GridIndexManager
+from .fem.quad_fields import QuadFieldManager
 
-from dataclasses import dataclass
-from functools import cached_property, lru_cache
+from functools import cached_property
 import numpy as np
 import time
-from muGrid import Field, Timer
+from muGrid import Timer
 
 import numpy.typing as npt
-from typing import TYPE_CHECKING, Tuple, Dict, Any, Optional
+from typing import TYPE_CHECKING, Tuple, Dict
 if TYPE_CHECKING:
     from .problem import Problem
-    from .fem.petsc_system import PETScLayout
 
 NDArray = npt.NDArray[np.floating]
-
-
-BASE_FIELDS = {
-    'rho', 'jx', 'jy',
-    'p', 'h', 'dh_dx', 'dh_dy', 'eta',
-    'U', 'V', 'Ls',
-    'dp_drho',
-    'rho_prev', 'jx_prev', 'jy_prev',
-    'pressure_stab',
-}
-
-STRESS_XZ_FIELDS = {
-    'tau_xz', 'dtau_xz_drho', 'dtau_xz_djx',
-    'tau_xz_bot', 'dtau_xz_bot_drho', 'dtau_xz_bot_djx',
-}
-
-STRESS_YZ_FIELDS = {
-    'tau_yz', 'dtau_yz_drho', 'dtau_yz_djy',
-    'tau_yz_bot', 'dtau_yz_bot_drho', 'dtau_yz_bot_djy',
-}
-
-ENERGY_FIELDS = {
-    'E', 'Tb_top', 'Tb_bot',
-    'T', 'dT_drho', 'dT_djx', 'dT_djy', 'dT_dE',
-    'S', 'dS_drho', 'dS_djx', 'dS_djy', 'dS_dE',
-    'E_prev',
-}
-
-
-@dataclass
-class TermIndices:
-    """
-    Precomputed indices for sparse assembly of a single (term, res, var) block.
-
-    Enables O(nnz) vectorized assembly: values written directly to COO array
-    using precomputed indices, avoiding dense matrix allocation.
-
-    Attributes
-    ----------
-    coo_idx : NDArray
-        COO value array positions to write to, shape (nb_contributions,).
-    elem_weights : NDArray
-        Precomputed FEM weights (N_i * N_j * w * A for zero-der,
-        N_i * w * A for derivative terms), shape (nb_contributions,).
-    quad_idx : NDArray
-        Quadrature point index (0-5, where 0-2=left tri, 3-5=right tri).
-    sq_x : NDArray
-        Square x-index for fetching quadrature values.
-    sq_y : NDArray
-        Square y-index for fetching quadrature values.
-    signs : Optional[NDArray]
-        Sign multipliers for derivative terms (+1/-1), None for zero-der.
-    """
-    coo_idx: NDArray
[…]
