+++ GaPFlow/fem/quad_fields.py
@@ -1,0 +1,287 @@
+#
+# Copyright 2025 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+"""Quadrature field management for FEM assembly.
+
+Handles field storage, interpolation operators, update methods, and access
+for nodal and quadrature point fields on triangular elements.
+"""
+from typing import Any, Dict, List, Set, TYPE_CHECKING
+
+import numpy as np
+import numpy.typing as npt
+from muGrid import Field
+
+from .utils2d import TriangleQuadrature
+
+if TYPE_CHECKING:
+    from ..problem import Problem
+
+NDArray = npt.NDArray[np.floating]
+
+
+# Field name sets for different physics
+BASE_FIELDS = {
+    'rho', 'jx', 'jy',
+    'p', 'h', 'dh_dx', 'dh_dy', 'eta',
+    'U', 'V', 'Ls',
+    'dp_drho',
+    'rho_prev', 'jx_prev', 'jy_prev',
+    'pressure_stab',
+}
+
+STRESS_XZ_FIELDS = {
+    'tau_xz', 'dtau_xz_drho', 'dtau_xz_djx',
+    'tau_xz_bot', 'dtau_xz_bot_drho', 'dtau_xz_bot_djx',
+}
+
+STRESS_YZ_FIELDS = {
+    'tau_yz', 'dtau_yz_drho', 'dtau_yz_djy',
+    'tau_yz_bot', 'dtau_yz_bot_drho', 'dtau_yz_bot_djy',
+}
+
+ENERGY_FIELDS = {
+    'E', 'Tb_top', 'Tb_bot',
+    'T', 'dT_drho', 'dT_djx', 'dT_djy', 'dT_dE',
+    'S', 'dS_drho', 'dS_djx', 'dS_djy', 'dS_dE',
+    'E_prev',
+}
+
+
+class QuadFieldManager:
+    """Manages quadrature fields, operators, and updates for FEM.
+
+    Handles:
+    - Field storage (nodal and quadrature point fields)
+    - Interpolation operators (nodal → quad, derivatives)
+    - Field access methods (get values, compute derivatives)
+    - Physics updates (nodal fields, interpolation, derived quantities)
+    - Time stepping (store previous values)
+
+    Parameters
+    ----------
+    problem : Problem
+        Problem instance for physics model access.
+    energy : bool
+        Whether energy equation is active.
+    variables : list of str
[…]
