+++ .github/workflows/ci.yaml
@@ -20,11 +20,12 @@ jobs:
       with:
         fetch-depth: 0
 
-    - name: Install MPI & NetCDF
+    - name: Install MPI, NetCDF & FFTW
       run: |
         sudo apt-get update
-        sudo apt-get install -y openmpi-bin libopenmpi-dev 
+        sudo apt-get install -y openmpi-bin libopenmpi-dev
         sudo apt-get install -y libhdf5-mpi-dev libpnetcdf-dev libnetcdf-pnetcdf-dev
+        sudo apt-get install -y libfftw3-dev libfftw3-mpi-dev
         sudo apt-get install -y ffmpeg
 
     - name: Set up Python

[…]
