### 2026-01-09 6889898
BUILD: make petsc optional, provide install script, skip petsc tests when not installed
