+++ tests/test_fem2d_jacobian.py
@@ -30,6 +30,7 @@ for various grid sizes and boundary condition configurations.
 import pytest
 import numpy as np
 from GaPFlow.problem import Problem
+from GaPFlow import HAS_PETSC
 
 from GaPFlow.solver_fem_2d import FEMSolver2D  # type: ignore
 
@@ -158,6 +159,7 @@ GRID_SIZES = [
 ]
 
 
+@pytest.mark.skipif(not HAS_PETSC, reason="PETSc not installed")
 class TestJacobianFiniteDifference:
     """Test Jacobian assembly against finite differences."""
 
@@ -337,6 +339,7 @@ class TestShapeFunctionConsistency:
         )
 
 
+@pytest.mark.skipif(not HAS_PETSC, reason="PETSc not installed")
 class TestJacobianSymmetry:
     """Test expected symmetry properties of Jacobian blocks."""
 
