### 2026-01-12 078cd56
DOC: 2D FEM Implementation Overview
