### 2026-01-12 679bf2a
ENH: file restructure, added scipy solver fallback for serial mode
