### 2026-01-12 bab5827
MAINT: removed dynamic option from fem solver
