### 2026-01-12 c19eb4f
MAINT: updated dependency chain for muGrid 0.105.1
