### 2026-01-12 1837cb8
BUILD: Use scikit-build-core as build system and vendor LAMMPS Python library CI: build wheels
