### 2026-01-12 f4db70b
CI: checkout submodules
