+++ .github/workflows/ci.yaml
@@ -19,6 +19,7 @@ jobs:
       uses: actions/checkout@v5
       with:
         fetch-depth: 0
+        submodules: 'true'
 
     - name: Install MPI & NetCDF
       run: |
@@ -35,7 +36,6 @@ jobs:
     - name: Install GaPFlow
       run: |
         python3 -m pip install --upgrade pip
-        python3 -m pip install -v --force-reinstall --no-cache --no-binary muGrid muGrid
         python3 -m pip install .[test]
 
     - name: Test with pytest (<3.13)
