+++ .github/workflows/wheels.yml
@@ -21,7 +21,7 @@ jobs:
           - macos-14-large
           - windows-2022
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
         with:
           submodules: 'true'
 
