### 2026-01-12 626011d
CI: test with parallel build
