+++ .github/workflows/ci.yaml
@@ -25,7 +25,6 @@ jobs:
       run: |
         sudo apt-get update
         sudo apt-get install -y openmpi-bin libopenmpi-dev 
-        sudo apt-get install -y libhdf5-mpi-dev libpnetcdf-dev libnetcdf-pnetcdf-dev
         sudo apt-get install -y ffmpeg
 
     - name: Set up Python
@@ -36,7 +35,7 @@ jobs:
     - name: Install GaPFlow
       run: |
         python3 -m pip install --upgrade pip
-        python3 -m pip install .[test]
+        python3 -m pip install .[test,parallel]
 
     - name: Test with pytest (<3.13)
       if: matrix.python-version != '3.13'
