### 2026-01-12 a459b33
CI: disable windows and coverage for now
