### 2026-01-12 b10717e
CI: netcdf library
