+++ .github/workflows/ci.yaml
@@ -21,10 +21,11 @@ jobs:
         fetch-depth: 0
         submodules: 'true'
 
-    - name: Install MPI & NetCDF
+    - name: Install MPI, NetCDF & FFMPEG
       run: |
         sudo apt-get update
         sudo apt-get install -y openmpi-bin libopenmpi-dev 
+        sudo apt-get install -y libnetcdf-dev 
         sudo apt-get install -y ffmpeg
 
     - name: Set up Python
