### 2026-01-12 2fe188a
CI: bump muGrid version, exclude Intel Macs
