+++ .github/workflows/wheels.yml
@@ -18,7 +18,7 @@ jobs:
         os: 
           - ubuntu-latest
           - macos-14
-          - macos-14-large
+          # - macos-14-large
           # - windows-2022
     steps:
       - uses: actions/checkout@v5
