+++ pyproject.toml
@@ -21,7 +21,7 @@ requires-python = ">=3.10"
 dynamic = ["version"]
 dependencies = [
 	"DiscoverVersion",
-	"muGrid==0.95.0",
+	"muGrid==0.96.0",
 	"tinygp>=0.3.0",
 	"jax",
 	"jaxopt==0.8.5",
