### 2026-01-12 76b2b38
MAINT: moved term handling to solver, 2d plotting fix
