+++ GaPFlow/fem_2d/terms.py
@@ -230,14 +230,228 @@ R2Ty = NonLinearTerm(
     d_dy_resfun=False,
     der_testfun=False)
 
+# -----------------------------------------------------------------------------
+# Energy equation terms (R3*)
+# -----------------------------------------------------------------------------
+
+R31x = NonLinearTerm(
+    name='R31x',
+    description='energy convection x',
+    res='energy',
+    dep_vars=['rho', 'jx', 'E'],
+    dep_vals=[],
+    fun=lambda ctx: lambda rho, jx, E: - (jx / rho) * E,
+    der_funs=[lambda ctx: lambda rho, jx, E: (jx / rho**2) * E,
+              lambda ctx: lambda rho, jx, E: - (1 / rho) * E,
+              lambda ctx: lambda rho, jx, E: - (jx / rho)],
+    d_dx_resfun=True,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+R31y = NonLinearTerm(
+    name='R31y',
+    description='energy convection y',
+    res='energy',
+    dep_vars=['rho', 'jy', 'E'],
+    dep_vals=[],
+    fun=lambda ctx: lambda rho, jy, E: - (jy / rho) * E,
+    der_funs=[lambda ctx: lambda rho, jy, E: (jy / rho**2) * E,
+              lambda ctx: lambda rho, jy, E: - (1 / rho) * E,
+              lambda ctx: lambda rho, jy, E: - (jy / rho)],
+    d_dx_resfun=False,
+    d_dy_resfun=True,
+    der_testfun=False)
+
+R31Sx = NonLinearTerm(
+    name='R31Sx',
+    description='energy convection height source x',
+    res='energy',
+    dep_vars=['rho', 'jx', 'E'],
+    dep_vals=['h', 'dh_dx'],
+    fun=lambda ctx: lambda rho, jx, E: - (jx / rho) * E * (1 / ctx['h']() * ctx['dh_dx']()),
+    der_funs=[lambda ctx: lambda rho, jx, E: (jx / rho**2) * E * (1 / ctx['h']() * ctx['dh_dx']()),
+              lambda ctx: lambda rho, jx, E: - (1 / rho) * E * (1 / ctx['h']() * ctx['dh_dx']()),
+              lambda ctx: lambda rho, jx, E: - (jx / rho) * (1 / ctx['h']() * ctx['dh_dx']())],
+    d_dx_resfun=False,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+R31Sy = NonLinearTerm(
+    name='R31Sy',
+    description='energy convection height source y',
+    res='energy',
+    dep_vars=['rho', 'jy', 'E'],
+    dep_vals=['h', 'dh_dy'],
+    fun=lambda ctx: lambda rho, jy, E: - (jy / rho) * E * (1 / ctx['h']() * ctx['dh_dy']()),
+    der_funs=[lambda ctx: lambda rho, jy, E: (jy / rho**2) * E * (1 / ctx['h']() * ctx['dh_dy']()),
+              lambda ctx: lambda rho, jy, E: - (1 / rho) * E * (1 / ctx['h']() * ctx['dh_dy']()),
+              lambda ctx: lambda rho, jy, E: - (jy / rho) * (1 / ctx['h']() * ctx['dh_dy']())],
+    d_dx_resfun=False,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+R32x = NonLinearTerm(
+    name='R32x',
+    description='pressure work x',
+    res='energy',
+    dep_vars=['rho', 'jx'],
+    dep_vals=[],
[…]
