### 2026-01-12 6a2aed2
DOC: started 2D FEM examples, extended 1D FEM examples
