+++ doc/tutorials/utils/__init__.py
@@ -1,0 +1,10 @@
+#
+# Copyright 2025 Christoph Huber
+#
+# ### MIT License
+#
+"""Utility functions for tutorials."""
+
+from .analytic import heat_equation_1d
+
+__all__ = ['heat_equation_1d']
