### 2026-01-12 9e69163
MAINT: added get_real_field wrapper and slice access for setting .p
