### 2026-01-12 13b85c1
CI: install serial netcdf (temporary solution, newer muGrid should have this)
