### 2026-01-12 ad94218
CI: don't need SurfaceTopography
