### 2026-01-12 90e0e3f
CI: roll back to old installation for tests (WIP)
