+++ .github/workflows/ci.yaml
@@ -25,7 +25,7 @@ jobs:
       run: |
         sudo apt-get update
         sudo apt-get install -y openmpi-bin libopenmpi-dev 
-        sudo apt-get install -y libnetcdf-dev 
+        sudo apt-get install -y libhdf5-mpi-dev libpnetcdf-dev libnetcdf-pnetcdf-dev
         sudo apt-get install -y ffmpeg
 
     - name: Set up Python
@@ -36,7 +36,8 @@ jobs:
     - name: Install GaPFlow
       run: |
         python3 -m pip install --upgrade pip
-        python3 -m pip install .[test,parallel]
+        python3 -m pip install -v --force-reinstall --no-cache --no-binary muGrid muGrid==0.96.0
+        python3 -m pip install -v .[test,parallel]
 
     - name: Test with pytest
       run:
@@ -53,13 +54,13 @@ jobs:
     #     coverage run -m pytest
     #     bash maintenance/update_coverage_badge.sh
 
-    - name: Upload coverage badge
-      if: matrix.python-version == '3.13' && github.ref == 'refs/heads/main'
-      uses: exuanbo/actions-deploy-gist@v1
-      with:
-        token: ${{ secrets.GIST_SECRET }}
-        gist_id: fac7fa61e1899b1e74b3bab598fe6513
-        file_path: badge.svg
+    # - name: Upload coverage badge
+    #   if: matrix.python-version == '3.13' && github.ref == 'refs/heads/main'
+    #   uses: exuanbo/actions-deploy-gist@v1
+    #   with:
+    #     token: ${{ secrets.GIST_SECRET }}
+    #     gist_id: fac7fa61e1899b1e74b3bab598fe6513
+    #     file_path: badge.svg
 
   deploy:
     runs-on: ubuntu-latest
