### 2026-01-13 2488b94
MAINT: use setuptools_scm for dynamic versioning
