### 2026-01-13 ee29401
MAINT: correct version from tag in wheels
