+++ pyproject.toml
@@ -36,7 +36,6 @@ dependencies = [
 	"ContactMechanics==1.7.0",
 	"IPython"]
 
-
 [project.optional-dependencies]
 parallel = ['mpi4py']
 
@@ -68,7 +67,6 @@ filterwarnings = "ignore::DeprecationWarning"
 
 [tool.coverage.run]
 source = ["GaPFlow"]
-
 
 [tool.scikit-build]
 metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
