+++ .github/workflows/wheels.yml
@@ -23,6 +23,7 @@ jobs:
     steps:
       - uses: actions/checkout@v5
         with:
+          fetch-depth: 0
           submodules: 'true'
 
       - name: Build wheels
