### 2026-01-13 bb32ca2
CI: re-organize CI/CD workflow
