+++ .github/workflows/test.yaml
@@ -8,6 +8,7 @@ on:
       - '**'
   pull_request:
     branches: [ main ]
+
 jobs:
   test:
     runs-on: ubuntu-latest
@@ -61,38 +62,3 @@ jobs:
     #     token: ${{ secrets.GIST_SECRET }}
     #     gist_id: fac7fa61e1899b1e74b3bab598fe6513
     #     file_path: badge.svg
-
-  deploy:
-    runs-on: ubuntu-latest
-    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
-    needs: test
-    environment:
-      name: ci
-    permissions:
-      # IMPORTANT: this permission is mandatory for trusted publishing
-      id-token: write
-
-    steps:
-      - name: Checkout source
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.13'
-
-      - name: Install dependencies
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y openmpi-bin libopenmpi-dev
-          sudo apt-get install -y libhdf5-mpi-dev libpnetcdf-dev libnetcdf-pnetcdf-dev
-          python3 -m pip install --upgrade pip
-          python3 -m pip install wheel twine build
-
-      - name: Package dist
-        run:
-          # Only serial, no LAMMPS
-          python3 -m build
-
-      - name: Publish package on PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
