+++ .github/workflows/wheels.yml
@@ -1,37 +1,0 @@
-name: Build wheels
-
-on:
-  push:
-    branches:
-      - '**'
-    tags:
-      - '**'
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  build_wheels:
-    name: Build wheels on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: 
-          - ubuntu-latest
-          - macos-14
-          # - macos-14-large
-          # - windows-2022
-    steps:
-      - uses: actions/checkout@v5
-        with:
-          fetch-depth: 0
-          submodules: 'true'
-
-      - name: Build wheels
-        uses: pypa/cibuildwheel@v3.3.1
-        env:
-          MACOSX_DEPLOYMENT_TARGET: 14.0
-
-      - uses: actions/upload-artifact@v4
-        with:
-          path: ./wheelhouse/*.whl
-          name: ${{ matrix.os }}
