+++ pyproject.toml
@@ -72,6 +72,16 @@ source = ["GaPFlow"]
 metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
 build-dir = "build/{wheel_tag}"
 build.verbose = true
+sdist.exclude = [
+  "lammps/**"
+]
+sdist.include = [
+  "lammps/src/**",
+  "lammps/cmake/**",
+  "lammps/doc/**",
+  "lammps/bench/**",
+  "lammps/potentials/**",
+]
 
 [tool.setuptools_scm]
 version_scheme = "guess-next-dev"
