### 2026-01-14 f6e7bd4
FIX: energy boundary condition bug
