### 2026-01-14 1e02452
MAINT: bump ContactMechanics version
