### 2026-01-14 e4a62d3
ENH: added diagonal scaling to improve general conditioning
