+++ GaPFlow/fem_2d/assembly_layout.py
@@ -34,18 +34,19 @@ Architecture:
     PETScAssemblyInfo     - Extracted info for PETSc (references, not copies)
 """
 from dataclasses import dataclass
-from typing import Dict, Optional, Tuple
+from typing import Dict, List, Optional, Tuple
 
 import numpy as np
 import numpy.typing as npt
 
 NDArray = npt.NDArray[np.floating]
 IntArray = npt.NDArray[np.signedinteger]
+Int8Array = npt.NDArray[np.int8]
 
 
 @dataclass
 class MatrixCOOPattern:
-    """Sparse matrix COO structure.
+    """Sparse matrix COO structure with block metadata for scaling.
 
     Attributes
     ----------
@@ -55,17 +56,23 @@ class MatrixCOOPattern:
         Local indices for each COO entry. Used for dense reconstruction.
     global_rows, global_cols : IntArray
         Global indices for PETSc preallocation.
+    res_block_idx : Int8Array
+        Residual block index for each entry (0=mass, 1=mom_x, 2=mom_y, 3=energy).
+    var_block_idx : Int8Array
+        Variable block index for each entry (0=rho, 1=jx, 2=jy, 3=E).
     """
     nnz: int
     local_rows: IntArray
     local_cols: IntArray
     global_rows: IntArray
     global_cols: IntArray
+    res_block_idx: Int8Array
+    var_block_idx: Int8Array
 
 
 @dataclass
 class RHSPattern:
-    """RHS vector assembly structure.
+    """RHS vector structure with block metadata for scaling.
 
     Attributes
     ----------
@@ -73,9 +80,12 @@ class RHSPattern:
         Local RHS size (nb_vars * nb_inner_pts).
     global_rows : IntArray
         Global row indices for PETSc assembly.
+    res_block_idx : Int8Array
+        Residual block index for each entry (0=mass, 1=mom_x, 2=mom_y, 3=energy).
     """
     size: int
     global_rows: IntArray
+    res_block_idx: Int8Array
 
 
 @dataclass
@@ -131,6 +141,69 @@ RHSTermKey = Tuple[str, str]            # (term_name, residual)
 
 
 @dataclass
+class ScalingInfo:
+    """Precomputed scaling factors for linear system conditioning.
+
+    Transforms the linear system J·dq = -R into a well-conditioned
+    scaled system J*·dq* = -R* where all matrix entries are O(1).
+
+    The transformation is:
+        J*[i,j] = J[i,j] · D_q[j] / D_R[i]
+        R*[i] = R[i] / D_R[i]
+        dq[j] = dq*[j] · D_q[j]
+
+    Attributes
+    ----------
+    coo_scale : NDArray
+        Scale factors for COO values, shape (nnz,).
+    rhs_scale : NDArray
+        Scale factors for residual vector, shape (rhs_size,).
+    sol_scale : NDArray
+        Scale factors for solution vector, shape (sol_size,).
+    char_scales : dict
+        Original characteristic scales for reference.
+    """
+    coo_scale: NDArray
+    rhs_scale: NDArray
+    sol_scale: NDArray
+    char_scales: Dict[str, float]
+
+    def scale_system(self, M_coo: NDArray, R: NDArray) -> Tuple[NDArray, NDArray]:
+        """Scale Jacobian COO values and residual vector.
+
+        Parameters
+        ----------
+        M_coo : NDArray, shape (nnz,)
+            Unscaled Jacobian in COO value format.
+        R : NDArray, shape (n,)
+            Unscaled residual vector.
+
+        Returns
+        -------
+        M_scaled : NDArray, shape (nnz,)
+            Scaled Jacobian COO values.
+        R_scaled : NDArray, shape (n,)
[…]
