### 2026-01-14 c00ff52
CI: enable windows wheels, rm before-all installation of netCDF (comes with muGrid wheels)
