+++ pyproject.toml
@@ -92,15 +92,11 @@ skip = ["*cp314t*", "*musllinux*"]
 test-command = "pytest -v {project}/tests"
 test-requires = "pytest"
 
-# binary installations of muGrid require at least serial netCDF
 [tool.cibuildwheel.linux]
 archs = ["x86_64"]
-before-all = "yum install -y netcdf-devel"
 
 [tool.cibuildwheel.macos]
 archs = ["arm64"]
-before-all = "brew install netcdf"
 
-# excluded (for now)
 [tool.cibuildwheel.windows]
 archs = ["AMD64"]
