+++ .github/workflows/test.yaml
@@ -22,12 +22,10 @@ jobs:
         fetch-depth: 0
         submodules: 'true'
 
-    - name: Install MPI, NetCDF & FFMPEG
+    - name: Install MPI
       run: |
         sudo apt-get update
-        sudo apt-get install -y openmpi-bin libopenmpi-dev 
-        sudo apt-get install -y libhdf5-mpi-dev libpnetcdf-dev libnetcdf-pnetcdf-dev
-        sudo apt-get install -y ffmpeg
+        sudo apt-get install -y openmpi-bin libopenmpi-dev
 
     - name: Set up Python
       uses: actions/setup-python@v5
@@ -37,28 +35,23 @@ jobs:
     - name: Install GaPFlow
       run: |
         python3 -m pip install --upgrade pip
-        python3 -m pip install -v --force-reinstall --no-cache --no-binary muGrid muGrid==0.96.0
         python3 -m pip install -v .[test,parallel]
 
-    - name: Test with pytest
+    - name: Test with pytest (!=3.13)
+      if: matrix.python-version != '3.13'
       run:
         pytest
-
-    # - name: Test with pytest (<3.13)
-    #   if: matrix.python-version != '3.13'
-    #   run:
-    #     pytest
     
-    # - name: Test with pytest & coverage (3.13)
-    #   if: matrix.python-version == '3.13'
-    #   run: |
-    #     coverage run -m pytest
-    #     bash maintenance/update_coverage_badge.sh
+    - name: Test with pytest & coverage (3.13)
+      if: matrix.python-version == '3.13'
+      run: |
+        coverage run -m pytest
+        bash maintenance/update_coverage_badge.sh
 
-    # - name: Upload coverage badge
-    #   if: matrix.python-version == '3.13' && github.ref == 'refs/heads/main'
-    #   uses: exuanbo/actions-deploy-gist@v1
-    #   with:
-    #     token: ${{ secrets.GIST_SECRET }}
-    #     gist_id: fac7fa61e1899b1e74b3bab598fe6513
-    #     file_path: badge.svg
+    - name: Upload coverage badge
+      if: matrix.python-version == '3.13' && github.ref == 'refs/heads/main'
+      uses: exuanbo/actions-deploy-gist@v1
+      with:
+        token: ${{ secrets.GIST_SECRET }}
+        gist_id: fac7fa61e1899b1e74b3bab598fe6513
+        file_path: badge.svg
