+++ .github/workflows/wheels.yaml
@@ -18,8 +18,7 @@ jobs:
         os: 
           - ubuntu-latest
           - macos-14
-          - windows-2022
-          # - macos-14-large
+          # - windows-2022
     steps:
       - uses: actions/checkout@v5
         with:
