+++ pyproject.toml
@@ -40,7 +40,7 @@ dependencies = [
 parallel = ['mpi4py']
 
 test = ['pytest',
-		'coverage']
+		'pytest-cov']
 
 doc = ['sphinx',
 	   'sphinx_book_theme',
@@ -64,9 +64,6 @@ Source = "https://github.com/hannes-holey/gapflow"
 addopts = "-v"
 testpaths = ["tests"]
 filterwarnings = "ignore::DeprecationWarning"
-
-[tool.coverage.run]
-source = ["GaPFlow"]
 
 [tool.scikit-build]
 metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
