### 2026-01-14 23d4dab
MAINT: license change due to bundled LAMMPS (MIT --> GPL-2.0)
