### 2026-01-14 24ed9b9
MAINT: changed gpf_info script output slightly
