### 2026-01-15 9dffd3c
MAINT: fix typo in install instructions
