### 2026-01-16 c5dc1c3
TST/DOC: added solver comparison
