+++ doc/tutorials/utils/plotting.py
@@ -1,0 +1,170 @@
+#
+# Copyright 2025 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+"""Plotting utilities for FEM tutorials."""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib import animation
+from IPython.display import HTML
+
+
+def animate_comparison(x_vec, t_vec, T_numeric, T_analytic, L):
+    """Animated comparison of numeric vs analytic temperature solutions.
+
+    Parameters
+    ----------
+    x_vec : array_like
+        Spatial coordinates, shape (nx,)
+    t_vec : array_like
+        Time values, shape (nt,)
+    T_numeric : ndarray
+        Numeric temperature field, shape (nt, nx)
+    T_analytic : ndarray
+        Analytic temperature field, shape (nt, nx)
+    L : float
+        Domain length
+
+    Returns
+    -------
+    HTML
+        Jupyter HTML animation object
+    """
+    # Print error summary
+    print(f"Initial max error: {np.max(np.abs(T_analytic[0] - T_numeric[0])):.2e} (should be ~0)")
+    print(f"Overall max error: {np.max(np.abs(T_analytic - T_numeric)):.2e} (numerical discretization)")
+
+    # Create animation
+    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(10, 3))
+
+    line_fem, = ax1.plot([], [], 'b-', linewidth=2, label='FEM')
+    line_ana, = ax1.plot([], [], 'r--', linewidth=2, label='Analytic')
+    ax1.set_xlim(0, L)
+    ax1.set_ylim(-0.1, 1.1)
+    ax1.set_xlabel('x [m]')
+    ax1.set_ylabel('T [K]')
+    ax1.set_title('Temperature Profiles')
+    ax1.legend()
+    ax1.grid(True, alpha=0.3)
+
+    line_err, = ax2.plot([], [], 'g-', linewidth=2)
+    ax2.set_xlim(0, L)
+    ax2.set_ylim(0, 0.02)
+    ax2.set_xlabel('x [m]')
+    ax2.set_ylabel('|T_FEM - T_analytic|')
+    ax2.set_title('Absolute Error')
+    ax2.grid(True, alpha=0.3)
+
[…]
