### 2026-01-16 ebbc33b
ENH: parabolic slider with elastic deformation use case, fixed height gradient communication
