+++ doc/tutorials/utils/__init__.py
@@ -5,6 +5,6 @@
 #
 """Utility functions for tutorials."""
 
-from .plotting import animate_comparison
+from .plotting import animate_comparison, plot_solver_comparison_rho_jx
 
-__all__ = ['animate_comparison']
+__all__ = ['animate_comparison', 'plot_solver_comparison_rho_jx']
