+++ GaPFlow/io.py
@@ -480,8 +480,8 @@ def sanitize_fem_solver(d):
 
     # Stabilization parameters
     out['tau_mode'] = str(d.get('tau_mode', 'physics'))  # 'physics' or 'constant'
-    out['pressure_stab_alpha'] = float(d.get('pressure_stab_alpha', 0.))
-    out['momentum_stab_alpha'] = float(d.get('momentum_stab_alpha', 0.))
+    out['pressure_stab_alpha'] = float(d.get('pressure_stab_alpha', 500.))
+    out['momentum_stab_alpha'] = float(d.get('momentum_stab_alpha', 10000.))
 
     out['equations'] = {}
     out['equations']['energy'] = bool(d.get('equations', {}).get('energy', False))
