+++ tests/configs/parabolic_slider.yaml
@@ -44,5 +44,3 @@ fem_solver:
     type: newton_alpha
     R_norm_tol: 1e-04
     max_iter: 10
-    pressure_stab_alpha: {pressure_stab_alpha}
-    momentum_stab_alpha: {momentum_stab_alpha}

[…]
