### 2026-01-16 b45231f
TST: removed skip from fem_2d FD test, CI can use scipy fallback
