### 2026-01-16 f9f2abe
ENH/DOC: added reference pressure/displacement option, finished parabolic slider + elastic deform use case
