### 2026-01-19 0860b17
MAINT: add similarity check for candidate training points
