### 2026-01-20 d31c66c
MAINT/DOC: few clean-ups, added 2D parabolic slider + energy + deform use case
