+++ doc/tutorials/utils/__init__.py
@@ -5,6 +5,20 @@
 #
 """Utility functions for tutorials."""
 
-from .plotting import animate_comparison, plot_solver_comparison_rho_jx
+from .plotting import (
+    animate_comparison,
+    plot_solver_comparison_rho_jx,
+    animate_advection,
+    animate_3d_surface,
+    animate_3d_advection,
+    plot_3d_snapshot,
+)
 
-__all__ = ['animate_comparison', 'plot_solver_comparison_rho_jx']
+__all__ = [
+    'animate_comparison',
+    'plot_solver_comparison_rho_jx',
+    'animate_advection',
+    'animate_3d_surface',
+    'animate_3d_advection',
+    'plot_3d_snapshot',
+]
