+++ doc/tutorials/utils/plotting.py
@@ -26,10 +26,11 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib import animation
+from mpl_toolkits.mplot3d import Axes3D
 from IPython.display import HTML
 
 
-def animate_comparison(x_vec, t_vec, T_numeric, T_analytic, L):
+def animate_comparison(x_vec, t_vec, T_numeric, T_analytic, L, save_path=None):
     """Animated comparison of numeric vs analytic temperature solutions.
 
     Parameters
@@ -44,6 +45,9 @@ def animate_comparison(x_vec, t_vec, T_numeric, T_analytic, L):
         Analytic temperature field, shape (nt, nx)
     L : float
         Domain length
+    save_path : str, optional
+        Path to save the animation (e.g., 'animation.mp4' or 'animation.gif').
+        Requires ffmpeg for mp4 or pillow for gif.
 
     Returns
     -------
@@ -97,6 +101,11 @@ def animate_comparison(x_vec, t_vec, T_numeric, T_analytic, L):
 
     ani = animation.FuncAnimation(fig, animate, init_func=init, frames=len(t_vec),
                                   interval=50, blit=True, repeat=True)
+
+    if save_path is not None:
+        ani.save(save_path, writer='ffmpeg' if save_path.endswith('.mp4') else 'pillow')
+        print(f"Animation saved to: {save_path}")
+
     plt.close(fig)
     return HTML(ani.to_jshtml())
 
@@ -151,7 +160,8 @@ def plot_solver_comparison_rho_jx(results, title):
     plt.show()
 
 
-def animate_advection(x_vec, t_vec, T_numeric, L, U_avg, T_min=None, T_max=None):
+def animate_advection(x_vec, t_vec, T_numeric, L, U_avg, T_min=None, T_max=None,
+                      save_path=None):
     """Animated visualization of temperature advection.
 
     Parameters
@@ -168,6 +178,9 @@ def animate_advection(x_vec, t_vec, T_numeric, L, U_avg, T_min=None, T_max=None)
         Average flow velocity for advection
     T_min, T_max : float, optional
         Temperature range for y-axis. If None, computed from data.
+    save_path : str, optional
+        Path to save the animation (e.g., 'animation.mp4' or 'animation.gif').
+        Requires ffmpeg for mp4 or pillow for gif.
 
     Returns
     -------
@@ -216,5 +229,268 @@ def animate_advection(x_vec, t_vec, T_numeric, L, U_avg, T_min=None, T_max=None)
 
     ani = animation.FuncAnimation(fig, animate, init_func=init, frames=len(t_vec),
                                   interval=50, blit=True, repeat=True)
+
+    if save_path is not None:
+        ani.save(save_path, writer='ffmpeg' if save_path.endswith('.mp4') else 'pillow')
+        print(f"Animation saved to: {save_path}")
+
     plt.close(fig)
     return HTML(ani.to_jshtml())
+
+
+def animate_3d_surface(x_vec, y_vec, t_vec, T_field, Lx=1.0, Ly=1.0,
+                       T_min=None, T_max=None, title='Temperature Field',
+                       view_angle=(30, -60), cmap='coolwarm', figsize=(8, 5),
+                       save_path=None):
+    """3D surface animation of a scalar field (e.g., temperature) over time.
+
+    Parameters
+    ----------
+    x_vec : array_like
+        Spatial x-coordinates, shape (nx,)
+    y_vec : array_like
+        Spatial y-coordinates, shape (ny,)
+    t_vec : array_like
+        Time values, shape (nt,)
[…]
