+++ GaPFlow/fem_2d/petsc_system.py
@@ -98,7 +98,8 @@ class PETScSystem:
             pc.setFactorLevels(fill_level)
         else:
             # MUMPS direct solver (default)
-            print("Using MUMPS direct solver for PETSc KSP.")
+            if self.comm.getRank() == 0:
+                print("Using MUMPS direct solver for PETSc KSP.")
             self.ksp.setType('preonly')
             pc = self.ksp.getPC()
             pc.setType('lu')

[…]
