### 2026-01-20 be884b7
MAINT/TST: improved FEM 2D - FD test including energy
