+++ GaPFlow/fem_2d/terms.py
@@ -71,6 +71,9 @@ class NonLinearTerm():
         i = self.dep_vars.index(dep_var)
         return self.der_funs[i](*args)
 
+# -----------------------------------------------------------------------------
+# Mass equation terms (R1*)
+# -----------------------------------------------------------------------------
 
 R11x = NonLinearTerm(
     name='R11x',
@@ -122,24 +125,24 @@ R11Sy = NonLinearTerm(
 
 R1Stabx = NonLinearTerm(
     name='R1Stabx',
-    description='pressure stabilization x (PSPG)',
+    description='pressure stabilization x',
     res='mass',
     dep_vars=['rho'],
     dep_vals=[],
-    fun=lambda ctx: lambda rho: -ctx['tau_pspg']() * ctx['p'](),
-    der_funs=[lambda ctx: lambda rho: -ctx['tau_pspg']() * ctx['dp_drho']()],
+    fun=lambda ctx: lambda rho: -ctx['tau_mass']() * ctx['p'](),
+    der_funs=[lambda ctx: lambda rho: -ctx['tau_mass']() * ctx['dp_drho']()],
     d_dx_resfun=True,
     d_dy_resfun=False,
     der_testfun=True)
 
 R1Staby = NonLinearTerm(
     name='R1Staby',
-    description='pressure stabilization y (PSPG)',
+    description='pressure stabilization y',
     res='mass',
     dep_vars=['rho'],
     dep_vals=[],
-    fun=lambda ctx: lambda rho: -ctx['tau_pspg']() * ctx['p'](),
-    der_funs=[lambda ctx: lambda rho: -ctx['tau_pspg']() * ctx['dp_drho']()],
+    fun=lambda ctx: lambda rho: -ctx['tau_mass']() * ctx['p'](),
+    der_funs=[lambda ctx: lambda rho: -ctx['tau_mass']() * ctx['dp_drho']()],
     d_dx_resfun=False,
     d_dy_resfun=True,
     der_testfun=True)
@@ -155,6 +158,10 @@ R1T = NonLinearTerm(
     d_dx_resfun=False,
     d_dy_resfun=False,
     der_testfun=False)
+
+# -----------------------------------------------------------------------------
+# Momentum equation terms (R2*)
+# -----------------------------------------------------------------------------
 
 R21x = NonLinearTerm(
     name='R21x',
@@ -230,28 +237,26 @@ R2Ty = NonLinearTerm(
     d_dy_resfun=False,
     der_testfun=False)
 
-# Momentum stabilization terms (Laplacian-type diffusion)
-# Matches PSPG pattern: -τ * ∫ (∂ψ/∂x) * (∂jx/∂x) dΩ
 R2Stabx = NonLinearTerm(
     name='R2Stabx',
-    description='momentum stabilization x (SUPG)',
+    description='momentum stabilization x',
     res='momentum_x',
     dep_vars=['jx'],
     dep_vals=[],
-    fun=lambda ctx: lambda jx: -ctx['tau_supg']() * jx,
-    der_funs=[lambda ctx: lambda jx: -ctx['tau_supg']() * np.ones_like(jx)],
+    fun=lambda ctx: lambda jx: -ctx['tau_mom']() * jx,
+    der_funs=[lambda ctx: lambda jx: -ctx['tau_mom']() * np.ones_like(jx)],
     d_dx_resfun=True,
     d_dy_resfun=False,
     der_testfun=True)
 
 R2Staby = NonLinearTerm(
     name='R2Staby',
-    description='momentum stabilization y (SUPG)',
+    description='momentum stabilization y',
[…]
