### 2026-01-20 1cbc2c4
MAINT: fix missing lammps import
