### 2026-01-21 ca12d00
MAINT: remove leading '/' on windows
