### 2026-01-21 be82a39
MAINT: undo some changes from last merge commit
