### 2026-01-22 f58b78a
ENH: solution guards, full momentum terms, more flexible bc treatment
