+++ GaPFlow/fem_2d/terms.py
@@ -262,6 +262,170 @@ R2Staby = NonLinearTerm(
     der_testfun=True)
 
 # -----------------------------------------------------------------------------
+# Convective momentum terms (R22*)
+# These represent momentum advection: transport of momentum by the flow velocity
+# -----------------------------------------------------------------------------
+
+R22xx = NonLinearTerm(
+    name='R22xx',
+    description='convective momentum flux jx*jx in x',
+    res='momentum_x',
+    dep_vars=['rho', 'jx'],
+    dep_vals=[],
+    fun=lambda ctx: lambda rho, jx: -(jx * jx) / rho,
+    der_funs=[
+        lambda ctx: lambda rho, jx: (jx * jx) / (rho ** 2),
+        lambda ctx: lambda rho, jx: -2 * jx / rho
+    ],
+    d_dx_resfun=True,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+R22xxS = NonLinearTerm(
+    name='R22xxS',
+    description='convective momentum flux jx*jx height source',
+    res='momentum_x',
+    dep_vars=['rho', 'jx'],
+    dep_vals=['h', 'dh_dx'],
+    fun=lambda ctx: lambda rho, jx: -1 / ctx['h']() * ctx['dh_dx']() * (jx * jx) / rho,
+    der_funs=[
+        lambda ctx: lambda rho, jx: 1 / ctx['h']() * ctx['dh_dx']() * (jx * jx) / (rho ** 2),
+        lambda ctx: lambda rho, jx: -1 / ctx['h']() * ctx['dh_dx']() * 2 * jx / rho
+    ],
+    d_dx_resfun=False,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+R22yx = NonLinearTerm(
+    name='R22yx',
+    description='convective momentum flux jx*jy in y',
+    res='momentum_x',
+    dep_vars=['rho', 'jx', 'jy'],
+    dep_vals=[],
+    fun=lambda ctx: lambda rho, jx, jy: -(jx * jy) / rho,
+    der_funs=[
+        lambda ctx: lambda rho, jx, jy: (jx * jy) / (rho ** 2),
+        lambda ctx: lambda rho, jx, jy: -jy / rho,
+        lambda ctx: lambda rho, jx, jy: -jx / rho
+    ],
+    d_dx_resfun=False,
+    d_dy_resfun=True,
+    der_testfun=False)
+
+R22yxS = NonLinearTerm(
+    name='R22yxS',
+    description='convective momentum flux jx*jy height source',
+    res='momentum_x',
+    dep_vars=['rho', 'jx', 'jy'],
+    dep_vals=['h', 'dh_dy'],
+    fun=lambda ctx: lambda rho, jx, jy: -1 / ctx['h']() * ctx['dh_dy']() * (jx * jy) / rho,
+    der_funs=[
+        lambda ctx: lambda rho, jx, jy: 1 / ctx['h']() * ctx['dh_dy']() * (jx * jy) / (rho ** 2),
+        lambda ctx: lambda rho, jx, jy: -1 / ctx['h']() * ctx['dh_dy']() * jy / rho,
+        lambda ctx: lambda rho, jx, jy: -1 / ctx['h']() * ctx['dh_dy']() * jx / rho
+    ],
+    d_dx_resfun=False,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+R22xy = NonLinearTerm(
+    name='R22xy',
+    description='convective momentum flux jx*jy in x',
+    res='momentum_y',
+    dep_vars=['rho', 'jx', 'jy'],
+    dep_vals=[],
[…]
