+++ GaPFlow/topography.py
@@ -42,29 +42,24 @@ from ContactMechanics.FFTElasticHalfSpace import (
 NDArray = npt.NDArray[np.floating]
 
 
-def create_midpoint_grid(grid, decomp: DomainDecomposition | None = None) -> Tuple[NDArray, NDArray]:
-    """Create cell-center coordinate arrays (xx, yy) for the grid."""
-    if decomp is not None:
-        # MPI-aware: use decomposition to get local coordinates
-        return decomp.local_coordinates_midpoint(grid['dx'], grid['dy'])
+def create_midpoint_grid(grid: dict, decomp: DomainDecomposition) -> Tuple[NDArray, NDArray]:
+    """Create cell-center coordinate arrays (xx, yy) for the grid.
 
-    # Legacy global grid (backward compatibility)
-    Lx = grid['Lx']
-    Ly = grid['Ly']
-    Nx = grid['Nx']
-    Ny = grid['Ny']
+    Parameters
+    ----------
+    grid : dict
+        Grid configuration with 'dx' and 'dy' keys.
+    decomp : DomainDecomposition
+        Domain decomposition instance for MPI-aware coordinate generation.
 
-    dx = Lx / Nx
-    ix = np.arange(-1, Nx + 1)
-    x = ix / Nx * Lx + dx / 2.
-
-    dy = Ly / Ny
-    iy = np.arange(-1, Ny + 1)
-    y = iy / Ny * Ly + dy / 2.
-
-    xx, yy = np.meshgrid(x, y, indexing='ij')
-
-    return xx, yy
+    Returns
+    -------
+    xx : NDArray
+        X-coordinate grid (local subdomain including ghost cells).
+    yy : NDArray
+        Y-coordinate grid (local subdomain including ghost cells).
+    """
+    return decomp.local_coordinates_midpoint(grid['dx'], grid['dy'])
 
 
 def journal_bearing(xx, grid, geo):
@@ -234,9 +229,6 @@ class Topography:
     due to the fluid pressure field.
     """
 
-    # Timer attribute for profiling (set externally after solver init)
-    timer = None
-
     def __init__(self,
                  fc: Any,
                  grid: dict,
@@ -270,6 +262,8 @@ class Topography:
 
         self.dx = grid['dx']
         self.dy = grid['dy']
+        self._Lx = grid['Lx']
+        self._Ly = grid['Ly']
 
         # Store for MPI communication
         self._decomp = decomp
@@ -336,24 +330,50 @@ class Topography:
             if self.ElasticDeformation.periodicity in ['half', 'none']:
                 p_ref = self.get_reference_pressure()
                 p = self.__pressure.pg - p_ref
-                deformation = self._timed_deformation_calc(p)
+                deformation = self._calc_deformation(p)
                 d_ref = self.get_reference_displacement(deformation)
                 deformation = deformation - d_ref
             else:
                 p = self.__pressure.pg
-                deformation = self._timed_deformation_calc(p)
+                deformation = self._calc_deformation(p)
             self.deformation = deformation
             self.h = self.h_undeformed + deformation
         else:
             pass
 
-    def _timed_deformation_calc(self, p):
-        """Wrapper for deformation calculation with optional timing."""
-        if self.timer is not None:
-            with self.timer("elastic_deformation_calc"):
-                return self.ElasticDeformation.get_deformation_underrelax(p, self.timer)
+    def _calc_deformation(self, p):
+        """Calculate elastic deformation from pressure field."""
+        return self.ElasticDeformation.get_deformation_underrelax(p)
+
+    def set_height_manual(self, height_field: NDArray) -> None:
+        """Set the height field manually from external source (e.g., a callback).
+
+        This is the recommended API for dynamically modifying topography during
+        simulation, such as when introducing obstacles via callback functions.
+
+        Parameters
[…]
