### 2026-01-23 hannes-holey/GaPFlow#55 @DanWaxman
[JOSS-Review] Installation-Related Issues
Hi, I'm one of the reviewers for [the corresponding JOSS submission](https://github.com/openjournals/joss-reviews/issues/9645).

On fresh installation, I had two small issues:
- [ ] OpenMPI requirement should be more clear.
- [ ] Current PyPi version does not install gpf_info

The current documentation seems to imply OpenMPI is required only for optional parallelization (or, at least, I possibly misread as such). But I can't load the library on a fresh MacOS install without OpenMPI:
<code>% pip install GaPFlow
% python
Python 3.13.3 (main, May 30 2025, 05:45:55) [Clang 20.1.4 ] on darwin
>>> import gapflow
Traceback (most recent call last):
  File "<python-input-0>", line 1, in <module>
    import gapflow
ModuleNotFoundError: No module named 'gapflow'
>>> from GaPFlow.problem import Problem
Traceback (most recent call last):
  File "<python-input-1>", line 1, in <module>
    from GaPFlow.problem import Problem
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/GaPFlow/__init__.py", line 32, in <module>
    from .problem import Problem  # noqa: F401, E402
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/GaPFlow/problem.py", line 32, in <module>
    from muGrid import GlobalFieldCollection, FileIONetCDF, OpenMode
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/muGrid/__init__.py", line 38, in <module>
    from mpi4py import MPI
  File "<frozen importlib._bootstrap>", line 1360, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1322, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 1262, in _find_spec
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/mpi4py/_mpiabi.py", line 290, in find_spec
    mpiabi_suffix = _get_mpiabi_suffix(fullname)
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/mpi4py/_mpiabi.py", line 276, in _get_mpiabi_suffix
    mpiabi = _get_mpiabi()
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/mpi4py/_mpiabi.py", line 258, in _get_mpiabi
    mpiabi = _get_mpiabi_from_libmpi(libmpi)
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/mpi4py/_mpiabi.py", line 217, in _get_mpiabi_from_libmpi
    lib = _dlopen_libmpi(libmpi)
  File ".../Documents/GaPFlow_Review/.venv/lib/python3.13/site-packages/mpi4py/_mpiabi.py", line 210, in _dlopen_libmpi
    raise RuntimeError("\n".join(errors))
RuntimeError: cannot load MPI library
[...]
dlopen(/opt/local/lib, 0x000B): tried: '/opt/local/lib' (no such file), '/System/Volumes/Preboot/Cryptexes/OS/opt/local/lib' (no such file), '/opt/local/lib' (no such file)
</code>

For the latter, installing via PyPi doesn't result in any executable `gpf_info`. Installing via `git` resolves this (i.e., `pip install git+https://github.com/hannes-holey/GaPFlow.git`).
