### 2026-01-24 e1de5a7
ENH/DOC/TST: fem_2d analytic tests, scaling test metrics, config+scripting doc, .xx .yy unified access from DomainDecomposition
