+++ GaPFlow/problem.py
@@ -26,7 +26,6 @@ import os
 import io
 import signal
 import numpy as np
-from copy import deepcopy
 from collections import deque
 from datetime import datetime
 from muGrid import FileIONetCDF, OpenMode
@@ -45,6 +44,7 @@ from . import __version__
 from .db import Database
 from .topography import Topography
 from .io import read_yaml_input, write_yaml, create_output_directory, history_to_csv
+from .analysis import compute_metrics, print_metrics, create_overview_plot
 from .models import WallStress, BulkStress, Pressure, Energy, Viscosity
 from .md import Mock, LennardJones, GoldAlkane
 from .viz.plotting import _plot_height_1d_from_field, _plot_height_2d_from_field
@@ -124,6 +124,9 @@ class Problem:
         # Callback functions called after each time step
         self._callbacks = []
 
+        # BC callback functions: {var_name: {boundary: callback}}
+        self._bc_callbacks = {}
+
         # Initialize solver
         if self.numerics['solver'] == 'explicit':
             from .solver_explicit import ExplicitSolver
@@ -171,14 +174,15 @@ class Problem:
             self.energy = Energy(self.fc, energy_spec, grid)
             self.energy.initialize()
 
-        # I/O
-        if not self.options['silent']:
-
+        # I/O - create output directory if saving data or plots
+        if self.options['save_output'] or self.options['output_plots']:
             self.outdir = create_output_directory(options['output'], options['use_tstamp'])
+
+        # Data files only when save_output is enabled
+        if self.options['save_output']:
 
             if database is not None:
                 # Set training path inside output path
-                # if database.overwrite_training_path:
                 database.set_training_path(os.path.join(self.outdir, 'train'),
                                            check_temporary=True)
 
@@ -217,11 +221,9 @@ class Problem:
 
             if gpx is not None:
                 field_names.append('wall_stress_xz_var')
-
             if gpy is not None:
                 field_names.append('wall_stress_yz_var')
-
-            if gpz:
+            if gpz is not None:
                 field_names.append('pressure_var')
 
             self.file.register_field_collection(self.fc, field_names=field_names)
@@ -433,8 +435,11 @@ class Problem:
         while not self.converged and self.step < self.numerics['max_it'] and not self._stop:
             self.update()
 
-            if self.step % self.options['write_freq'] == 0 and not self.options['silent']:
-                self.write()
+            if self.step % self.options['write_freq'] == 0:
+                if self.options['print_progress']:
+                    self.solver.print_status(True)
+                if self.options['save_output']:
+                    self.write(scalars=False)
 
             _handle_signals(self._receive_signal)
 
@@ -454,7 +459,7 @@ class Problem:
         self.wall_stress_xz.init()
         self.wall_stress_yz.init()
 
-        if not self.options['silent']:
+        if self.options['save_output']:
             self.pressure.write()
             self.wall_stress_xz.write()
             self.wall_stress_yz.write()
@@ -479,13 +484,23 @@ class Problem:
         """
         Finalize run: write history, print timing and GP timing info.
         """
+        # Print metrics if requested
+        if self.options.get('print_metrics', False):
+            from mpi4py import MPI
+            metrics = compute_metrics(self)
+            print_metrics(metrics, MPI.COMM_WORLD)
+
+        # Output overview plot if requested
+        if self.options.get('output_plots', False):
+            plot_path = os.path.join(self.outdir, 'overview.png')
+            create_overview_plot(self, plot_path)
 
[…]
