+++ GaPFlow/analysis/postprocessing.py
@@ -1,0 +1,420 @@
+#
+# Copyright 2025 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+"""
+Postprocessing utilities for GaPFlow simulations.
+
+Provides metrics computation and overview plotting with MPI support
+for scaling studies and general analysis.
+
+Features
+--------
+- Metrics: Automatically printed when `output_metrics: True` in config
+- Plots: Overview figure saved when `output_plots: True` in config
+
+Metrics Output Format::
+
+    ==================================================
+    METRICS OUTPUT
+    ==================================================
+    coeff_friction: 0.00234
+    density_max: 1.2345
+    ...
+    ==================================================
+"""
+from datetime import datetime
+
+import numpy as np
+from mpi4py import MPI
+
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from ..problem import Problem
+
+
+def compute_metrics(problem: "Problem") -> dict:
+    """
+    Compute all characteristic metrics from a completed simulation.
+
+    Parameters
+    ----------
+    problem : Problem
+        The completed GaPFlow problem instance.
+
+    Returns
+    -------
+    dict
+        Dictionary of metric names and values.
+    """
+    comm = MPI.COMM_WORLD
+
+    metrics = {}
+    metrics.update(_density_metrics(problem, comm))
+    metrics.update(_pressure_metrics(problem, comm))
+    metrics.update(_flow_metrics(problem, comm))
+    metrics.update(_friction_metrics(problem, comm))
+    metrics.update(_deformation_metrics(problem, comm))
+    metrics.update(_energy_metrics(problem, comm))
+    metrics.update(_timing_metrics(problem))
+    metrics.update(_solver_metrics(problem))
+
+    return metrics
+
+
+def _density_metrics(problem: "Problem", comm: MPI.Comm) -> dict:
+    """Compute density field statistics."""
+    rho = problem.q[0, 1:-1, 1:-1]  # Inner cells only (no ghosts)
+    grid_total = problem.grid['Nx'] * problem.grid['Ny']
+
+    return {
[…]
