### 2026-01-26 6d2259d
TST: Bernoulli and temperature diffusion added to fem_2d analytic tests
