+++ tests/test_fem_2d_analytic.py
@@ -28,6 +28,7 @@ import pytest
 import numpy as np
 
 from GaPFlow import Problem
+from GaPFlow.models.pressure import eos_pressure
 
 CONFIG_DIR = os.path.join(os.path.dirname(__file__), 'configs')
 
@@ -124,3 +125,138 @@ def test_couette_2d(couette_problem):
     l2_err = compute_l2_error(vx, u_ana)
 
     assert l2_err < 0.01, f"L2 error = {l2_err:.2e} exceeds tolerance 0.01"
+
+
+# =============================================================================
+# Test: Bernoulli Venturi Flow
+# =============================================================================
+
+# Geometry parameters
+H_INLET, H_THROAT = 1.0e-3, 0.5e-3
+X_RAMP_START, X_RAMP_END = 0.03, 0.07
+THROAT_HALF_WIDTH = 0.005
+RHO0, V_INLET = 1000.0, 5.0
+
+def _venturi_topography(xx):
+    """Symmetric venturi with cosine transitions."""
+    x_mid = (X_RAMP_START + X_RAMP_END) / 2
+    ramp_len = x_mid - THROAT_HALF_WIDTH - X_RAMP_START
+    
+    # Distance from throat edge, clipped to [0, ramp_len]
+    dist = np.clip(np.abs(xx - x_mid) - THROAT_HALF_WIDTH, 0, ramp_len)
+    xi = dist / ramp_len
+    
+    return np.where(
+        (xx >= X_RAMP_START) & (xx <= X_RAMP_END),
+        (H_INLET + H_THROAT)/2 - (H_INLET - H_THROAT)/2 * np.cos(np.pi * xi),
+        H_INLET
+    )
+
+
+@pytest.fixture
+def bernoulli_problem():
+    """Create and run Bernoulli venturi flow problem."""
+    config_path = os.path.join(CONFIG_DIR, 'bernoulli_venturi.yaml')
+    problem = Problem.from_yaml(config_path)
+    problem.topo.set_mapped_height(_venturi_topography(problem.topo.xx))
+    problem.run()
+    return problem
+
+
+def test_bernoulli_venturi(bernoulli_problem):
+    """Bernoulli flow through symmetric venturi.
+
+    Tests convective terms (R22xx, R22yx, ...) against Bernoulli equation.
+    Validates: velocity scaling with area, pressure drop, mass conservation.
+    """
+    problem = bernoulli_problem
+
+    # Extract centerline data
+    rho = problem.q[0][1:-1, 1:-1]
+    jx = problem.q[1][1:-1, 1:-1]
+    h = problem.topo.h[1:-1, 1:-1]
+    j_center = rho.shape[1] // 2
+
+    dx = problem.grid['dx']
+    x = np.arange(rho.shape[0]) * dx + dx / 2
+    vx = jx[:, j_center] / rho[:, j_center]
+    p = np.asarray(eos_pressure(rho[:, j_center], problem.prop))
+
+    # Define regions
+    inlet = x < X_RAMP_START
+    throat = (x >= 0.045) & (x <= 0.055)
+    outlet = x > X_RAMP_END
+
+    # Measured values
+    v_in, v_throat, v_out = vx[inlet].mean(), vx[throat].mean(), vx[outlet].mean()
+    p_in, p_throat = p[inlet].mean(), p[throat].mean()
+
+    # Bernoulli predictions
+    v_throat_theory = V_INLET * H_INLET / H_THROAT
+    dp_theory = 0.5 * RHO0 * (v_throat_theory**2 - V_INLET**2)
+    dp_sim = p_in - p_throat
+
+    # Mass flux conservation
+    mass_flux = rho[:, j_center] * vx * h[:, j_center]
+    mass_var = (mass_flux.max() - mass_flux.min()) / mass_flux.mean()
+
[…]
