+++ GaPFlow/fem_2d/terms.py
@@ -71,10 +71,12 @@ class NonLinearTerm():
         i = self.dep_vars.index(dep_var)
         return self.der_funs[i](*args)
 
+
 # -----------------------------------------------------------------------------
 # Mass equation terms (R1*)
 # -----------------------------------------------------------------------------
 
+# R11: Flux divergence
 R11x = NonLinearTerm(
     name='R11x',
     description='flux divergence x',
@@ -123,6 +125,20 @@ R11Sy = NonLinearTerm(
     d_dy_resfun=False,
     der_testfun=False)
 
+# R1T: Time derivative
+R1T = NonLinearTerm(
+    name='R1T',
+    description='time derivative',
+    res='mass',
+    dep_vars=['rho'],
+    dep_vals=[],
+    fun=lambda ctx: lambda rho: - (rho - ctx['rho_prev']()) / ctx['dt'],
+    der_funs=[lambda ctx: lambda rho: - np.full_like(rho, 1.0) / ctx['dt']],
+    d_dx_resfun=False,
+    d_dy_resfun=False,
+    der_testfun=False)
+
+# R1Stab: Pressure stabilization
 R1Stabx = NonLinearTerm(
     name='R1Stabx',
     description='pressure stabilization x',
@@ -147,22 +163,11 @@ R1Staby = NonLinearTerm(
     d_dy_resfun=True,
     der_testfun=True)
 
-R1T = NonLinearTerm(
-    name='R1T',
-    description='time derivative',
-    res='mass',
-    dep_vars=['rho'],
-    dep_vals=[],
-    fun=lambda ctx: lambda rho: - (rho - ctx['rho_prev']()) / ctx['dt'],
-    der_funs=[lambda ctx: lambda rho: - np.full_like(rho, 1.0) / ctx['dt']],
-    d_dx_resfun=False,
-    d_dy_resfun=False,
-    der_testfun=False)
-
 # -----------------------------------------------------------------------------
 # Momentum equation terms (R2*)
 # -----------------------------------------------------------------------------
 
+# R21: Pressure gradient
 R21x = NonLinearTerm(
     name='R21x',
     description='pressure gradient x',
@@ -187,111 +192,7 @@ R21y = NonLinearTerm(
     d_dy_resfun=True,
     der_testfun=False)
 
-R24x= NonLinearTerm(
-    name='R24x',
-    description='wall stress',
-    res='momentum_x',
-    dep_vars=['rho', 'jx'],
-    dep_vals=['h', 'tau_xz', 'dtau_xz_drho', 'dtau_xz_djx'],
-    fun=lambda ctx: lambda *args: 1 / ctx['h']() * ctx['tau_xz'](),
-    der_funs=[lambda ctx: lambda *args: 1 / ctx['h']() * ctx['dtau_xz_drho'](),
-              lambda ctx: lambda *args: 1 / ctx['h']() * ctx['dtau_xz_djx']()],
-    d_dx_resfun=False,
-    d_dy_resfun=False,
-    der_testfun=False)
-
-R24y= NonLinearTerm(
-    name='R24y',
-    description='wall stress',
-    res='momentum_y',
-    dep_vars=['rho', 'jy'],
-    dep_vals=['h', 'tau_yz', 'dtau_yz_drho', 'dtau_yz_djy'],
-    fun=lambda ctx: lambda *args: 1 / ctx['h']() * ctx['tau_yz'](),
-    der_funs=[lambda ctx: lambda *args: 1 / ctx['h']() * ctx['dtau_yz_drho'](),
-              lambda ctx: lambda *args: 1 / ctx['h']() * ctx['dtau_yz_djy']()],
[…]
