### 2026-01-26 599acb5
MAINT: pass zero to mean function
