### 2026-01-26 226abc8
MAINT: use mean_value arg
