### 2026-01-26 47f7fa6
MAINT: try build with tinygp fork
