+++ pyproject.toml
@@ -30,7 +30,8 @@ requires-python = ">=3.10"
 dynamic = ["version"]
 dependencies = [
 	"muGrid>=0.105.0",
-	"tinygp>=0.3.0",
+	# "tinygp>=0.3.0",
+	"tinygp = { git = 'https://github.com/hannes-holey/tinygp', branch = 'main' }",
 	"jax",
 	"numpy",
 	"scipy",
