### 2026-01-26 83dccda
MAINT: fix syntax
