### 2026-01-27 f5952f3
MAINT: pass mean to gp
