+++ .github/workflows/wheels.yaml
@@ -17,8 +17,8 @@ jobs:
       matrix:
         os: 
           - ubuntu-latest
-          - macos-14
-          - windows-2022
+          - macos-latest
+          - windows-latest
     steps:
       - uses: actions/checkout@v5
         with:
