### 2026-01-27 e34a6ce
CI: fix macos deployment target
