+++ .github/workflows/wheels.yaml
@@ -28,7 +28,7 @@ jobs:
       - name: Build wheels
         uses: pypa/cibuildwheel@v3.3.1
         env:
-          MACOSX_DEPLOYMENT_TARGET: 14.0
+          MACOSX_DEPLOYMENT_TARGET: 15.0
 
       - uses: actions/upload-artifact@v4
         with:
