### 2026-01-28 5efb5c2
WIP: increase memory efficiency and stabilization robustness
