+++ GaPFlow/solver_fem_2d.py
@@ -30,12 +30,15 @@ from .fem_2d.assembly_layout import (
     RHSPattern,
     JacobianTermMap,
     RHSTermMap,
+    COOLookup,
 )
 from .fem_2d.grid_index import GridIndexManager
 from .fem_2d.quad_fields import QuadFieldManager
 from .fem_2d.solution_guards import apply_guards
+from .fem_2d.stab_diagnostics import StabDiagnostics
 
 from functools import cached_property
+import os
 import numpy as np
 import time
 from muGrid import Timer
@@ -242,11 +245,14 @@ class FEMSolver2D:
             nb_vars=nb_vars,
             nb_inner_pts=self.nb_inner_pts,
             nb_global_pts=Nx * Ny,
+            nb_sq=self.nb_sq,
+            sq_per_row=self.sq_per_row,
+            sq_per_col=self.sq_per_col,
         )
 
     def _build_matrix_coo_pattern(self, l2g_pts: NDArray, nb_vars: int,
-                                  nb_res: int) -> Tuple[MatrixCOOPattern, dict]:
-        """Build sparse matrix COO pattern and local-to-COO map."""
+                                  nb_res: int) -> Tuple[MatrixCOOPattern, COOLookup]:
+        """Build sparse matrix COO pattern and memory-efficient COO lookup."""
         inner_pts, contrib_pts = self.grid_idx.get_stencil_connectivity()
         nnz_per_block = len(inner_pts)
         total_nnz = nb_res * nb_vars * nnz_per_block
@@ -279,11 +285,8 @@ class FEMSolver2D:
 
                 idx += n
 
-        # Build local-to-COO lookup dictionary
-        local_to_coo = {}
-        for coo_idx in range(total_nnz):
-            key = (int(local_rows[coo_idx]), int(local_cols[coo_idx]))
-            local_to_coo[key] = coo_idx
+        # Build memory-efficient COO lookup (replaces Python dict)
+        coo_lookup = COOLookup.from_coo_pattern(local_rows, local_cols)
 
         matrix_coo = MatrixCOOPattern(
             nnz=total_nnz,
@@ -295,7 +298,7 @@ class FEMSolver2D:
             var_block_idx=var_block_idx,
         )
 
-        return matrix_coo, local_to_coo
+        return matrix_coo, coo_lookup
 
     def _build_rhs_pattern(self, l2g_pts: NDArray, nb_res: int) -> RHSPattern:
         """Build RHS vector assembly pattern."""
@@ -342,58 +345,49 @@ class FEMSolver2D:
 
         output_idx_list = []
         weights_list = []
-        quad_idx_list = []
-        sq_x_list = []
-        sq_y_list = []
+        flat_field_idx_list = []
 
         for TO_tri, test_wA, quad_offset in triangle_configs:
             for i in range(3):
                 valid = TO_tri[:, i] >= 0
                 sq_valid = np.where(valid)[0]
                 if len(sq_valid) > 0:
+                    sq_x_valid = sq_x[sq_valid]
+                    sq_y_valid = sq_y[sq_valid]
                     for q in range(3):
                         output_idx_list.append(
                             res_idx * self.nb_inner_pts + TO_tri[sq_valid, i])
                         weights_list.append(
                             np.full(len(sq_valid), test_wA[i, q], dtype=np.float64))
-                        quad_idx_list.append(
-                            np.full(len(sq_valid), q + quad_offset, dtype=np.int32))
-                        sq_x_list.append(sq_x[sq_valid])
-                        sq_y_list.append(sq_y[sq_valid])
+                        # Flat field index: quad_idx * nb_sq + sq_x * sq_per_col + sq_y
[…]
