### 2026-01-28 2be6a14
BUG: rotation magnitude
