### 2026-01-29 c790bb9
WIP: weak memory enhancement done, working on Neumann stabilization through explicit BC
