+++ GaPFlow/fem_2d/scaling.py
@@ -1,0 +1,234 @@
+#
+# Copyright 2025 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+
+"""
+Linear system scaling for FEM solver conditioning.
+
+This module provides scaling utilities to improve the conditioning of the
+linear systems arising from Newton iteration in the FEM solver.
+
+The transformation is:
+    J*[i,j] = J[i,j] · D_q[j] / D_R[i]
+    R*[i] = R[i] / D_R[i]
+    dq[j] = dq*[j] · D_q[j]
+
+where D_q and D_R are characteristic scales for variables and residuals.
+"""
+from dataclasses import dataclass
+from typing import Dict, List, Tuple, TYPE_CHECKING
+
+import numpy as np
+import numpy.typing as npt
+
+if TYPE_CHECKING:
+    from .assembly_layout import MatrixCOOPattern, RHSPattern
+    from GaPFlow import Problem
+
+NDArray = npt.NDArray[np.floating]
+
+
+@dataclass
+class ScalingInfo:
+    """Precomputed scaling factors for linear system conditioning.
+
+    Transforms the linear system J·dq = -R into a well-conditioned
+    scaled system J*·dq* = -R* where all matrix entries are O(1).
+
+    The transformation is:
+        J*[i,j] = J[i,j] · D_q[j] / D_R[i]
+        R*[i] = R[i] / D_R[i]
+        dq[j] = dq*[j] · D_q[j]
+
+    Attributes
+    ----------
+    coo_scale : NDArray
+        Scale factors for COO values, shape (nnz,).
+    rhs_scale : NDArray
+        Scale factors for residual vector, shape (rhs_size,).
+    sol_scale : NDArray
+        Scale factors for solution vector, shape (sol_size,).
+    char_scales : dict
+        Original characteristic scales for reference.
+    """
+    coo_scale: NDArray
+    rhs_scale: NDArray
+    sol_scale: NDArray
+    char_scales: Dict[str, float]
+
+    def scale_system(self, M_coo: NDArray, R: NDArray) -> Tuple[NDArray, NDArray]:
+        """Scale Jacobian COO values and residual vector.
+
+        Parameters
+        ----------
+        M_coo : NDArray, shape (nnz,)
+            Unscaled Jacobian in COO value format.
+        R : NDArray, shape (n,)
+            Unscaled residual vector.
+
+        Returns
[…]
