### 2026-01-30 af00740
ENH: choose methods for input and output normalization
