### 2026-01-30 eef39fe
ENH: more intuitive fem solver term specification and general clean-up
