### 2026-01-31 5eccc4e
WIP: increased TermMaps RAM efficiency
