+++ GaPFlow/solver_fem_2d.py
@@ -30,6 +30,8 @@ from .fem_2d.assembly_layout import (
     MatrixCOOPattern,
     RHSPattern,
     JacobianTermMap,
+    JacobianTermTemplate,
+    JacobianTermRef,
     COOLookup,
 )
 from .fem_2d.scaling import compute_characteristic_scales
@@ -194,10 +196,14 @@ class FEMSolver2D:
         return self._build_assembly_layout()
 
     def _build_assembly_layout(self) -> FEMAssemblyLayout:
-        """Build complete assembly layout with COO pattern and term mappings."""
+        """Build complete assembly layout with COO pattern and term mappings.
+
+        Uses template-based approach: builds ~5-20 shared templates instead of
+        ~80 individual JacobianTermMaps. Each term gets a lightweight ref that
+        points to its template plus a block offset.
+        """
         nb_vars = len(self.variables)
         nb_res = len(self.residuals)
-        Nx, Ny = self.decomp.nb_domain_grid_pts
 
         # 1. Build local-to-global point mapping
         mask_local = self.grid_idx.index_mask_padded_local('')
@@ -207,50 +213,47 @@ class FEMSolver2D:
         l2g_pts[mask_local[valid]] = mask_global[valid]
 
         # 2. Build matrix COO pattern
-        matrix_coo, local_to_coo = self._build_matrix_coo_pattern(
+        matrix_coo, coo_lookup = self._build_matrix_coo_pattern(
             l2g_pts, nb_vars, nb_res)
 
         # 3. Build RHS pattern
         rhs = self._build_rhs_pattern(l2g_pts, nb_res)
 
-        # 4. Build Jacobian term maps with deduplication
-        # Terms with identical structural keys share the same JacobianTermMap object
+        # 4. Build shared Jacobian templates (grouped by FROM pattern)
+        inner_pts, _ = self.grid_idx.get_stencil_connectivity()
+        nnz_per_block = len(inner_pts)
+        jacobian_templates, var_to_from_hash = self._build_jacobian_templates(
+            coo_lookup, nnz_per_block)
+
+        # 5. Create lightweight refs for each (term, res, dep_var) combination
         jacobian_terms = {}
-        structural_key_to_map = {}  # temporary, for deduplication during build
         for term in self.terms:
             for dep_var in term.dep_vars:
                 term_key = (term.name, term.res, dep_var)
-                structural_key = self._get_structural_key(term, dep_var)
+                deriv_type, der_testfun = self._get_template_key(term)
+                from_hash = var_to_from_hash[dep_var]
+                template_key = (from_hash, deriv_type, der_testfun)
 
-                # Only build if this structural pattern hasn't been seen
-                if structural_key not in structural_key_to_map:
-                    if term.d_dx_resfun:
-                        term_map = self._build_jacobian_term_deriv(
-                            term, dep_var, local_to_coo, 'x')
-                    elif term.d_dy_resfun:
-                        term_map = self._build_jacobian_term_deriv(
-                            term, dep_var, local_to_coo, 'y')
-                    else:
-                        term_map = self._build_jacobian_term_zero_der(
-                            term, dep_var, local_to_coo)
-                    structural_key_to_map[structural_key] = term_map
+                res_idx = self.residuals.index(term.res)
+                var_idx = self.variables.index(dep_var)
+                block_offset = (res_idx * nb_vars + var_idx) * nnz_per_block
 
-                # Store reference (same object if structural_key already existed)
-                jacobian_terms[term_key] = structural_key_to_map[structural_key]
+                jacobian_terms[term_key] = JacobianTermRef(
[…]
