### 2026-02-09 ce1bda4
MAINT: always write training data to file
