### 2026-02-09 6156830
ENH: different options for tolerance criterion
