### 2026-02-09 9ada33a
MAINT: fix assert
