+++ GaPFlow/models/gp.py
@@ -89,7 +89,7 @@ class GaussianProcessSurrogate:
             self._cache = None
             self._database = database
             self._last_fit_train_size = 0
-            self._pause = 0
+            self._pause = self.pause_steps
 
             # Initialize timers
             ref = datetime.now()
