### 2026-02-10 609866a
MAINT: modify logging, write upto three logfiles into output directory
