+++ GaPFlow/md/md.py
@@ -42,11 +42,14 @@ from ..models.pressure import eos_pressure
 from ..models.viscous import stress_bottom, stress_top
 from .moltemplate import write_template, build_template
 from .utils import read_output_files, read_output_files_X
-from ..utils import bordered_text, make_dumpable
+from ..utils import make_dumpable
+from ..logging import get_logger
 
 yaml = YAML()
 yaml.explicit_start = True
 yaml.indent(mapping=4, sequence=4, offset=2)
+
+logger = get_logger("gapflow.md")
 
 
 class MolecularDynamics:
@@ -79,6 +82,9 @@ class MolecularDynamics:
     _dtool_basepath: str = '/tmp/'
     _readme_template: str = ""
     _input_names: list[str] = ['ρ', 'jx', 'jy', 'h', '∂h/∂x', '∂h/∂y'] + [f'extra_{i}' for i in range(10)]
+    _output_names: list[str] = ['p',
+                                'τ(xx|bot)', 'τ(yy|bot)', 'τ(zz|bot)', 'τ(yz|bot)', 'τ(xz|bot)', 'τ(xy|bot)',
+                                'τ(xx|top)', 'τ(yy|top)', 'τ(zz|top)', 'τ(yz|top)', 'τ(xz|top)', 'τ(xy|top)']
     _ascii_art: str = r"""
   _        _    __  __ __  __ ____  ____
  | |      / \  |  \/  |  \/  |  _ \/ ___|
@@ -118,22 +124,64 @@ class MolecularDynamics:
         """Read simulation output and returns observations and their standard error."""
         raise NotImplementedError
 
-    def _pretty_print(self, proto_datapath, X) -> None:
-        """Print header before the start of a LAMMPS simulation.
+    def _write_log_input(self, proto_datapath, X) -> None:
+        """Write logging info before an MD run.
 
         Parameters
         ----------
         proto_datapath : str
             The data path inside the dtool proto dataset
         X : Array
-            The input array
+            The target input array
         """
-        text = ['Run next MD simulation in:', f'{proto_datapath}']
-        text.append(self._ascii_art)
-        text.append('---')
+
+        logger.info(120 * '=')
+        logger.info('Running MD simulation')
+        logger.info('---')
+        logger.info(self._ascii_art)
+        logger.info('---')
+
         for i, (Xi, name) in enumerate(zip(X, self._input_names)):
-            text.append(f'Input {i + 1}: {Xi:+.3e}    ({name})')
-        print(bordered_text('\n'.join(text)))
+            logger.info(f'Target input {i + 1}: {Xi:+.3e}    ({name})')
+
+        logger.info('---')
+        logger.info('View log:')
+        logger.info(f'{proto_datapath}/data/log.lammps')
+        logger.info('---')
+
+    def _write_log_output(self, X, X_target, Y, Ye, walltime) -> None:
+        """Write logging info after an MD run.
+
+        Parameters
+        ----------
+        X : Array
+            The realized input
+        X_target : Array
+            The target input
+        Y : Array
+            Measured output
+        Ye : Array
+            Measured output error
+        walltime: datetime.timedelta
+            Wall time of an MD run
+        """
+
+        logger.info('---')
+        logger.info(f'MD run completed (walltime: {str(walltime).split('.')[0]})')
+        logger.info('---')
+
+        for i, (Xi, Xi_t, name) in enumerate(zip(X, X_target, self._input_names)):
+            deviation = (Xi - Xi_t) / max(abs(Xi_t), 1e-12) * 100
+            logger.info(f'Realized input {i + 1}: {Xi:+.3e} ({deviation:+.1f}%)   ({name})')
+
+        logger.info('---')
+
[…]

[…]
