### 2026-02-10 0527dc6
MAINT: use prediction instead of training data to compute variance tol
