### 2026-02-10 8e88c88
MAINT: show which tol crit is used in log
