### 2026-02-10 561abd4
MAINT: show no relative error for zero targets
