### 2026-02-10 #9645 @vyastreb
Dear @matthewfeickert and @hannes-holey,

That is a serious software which nicely complement the existing landscape with rich multiscale framework.
I would recommend the publication of a paper in JOSS.

Please find below some additional comments in complement to my checklist.

**Pros**

+ Well integrated in the current software infrastructure
+ Installation worked smoothly
+ Interfaces many different tools
+ Enables high accuracy and rich multi-scale computations with smart integration and adaptive learning
+ All worked smoothly for all tests I run
+ Nice documentation with extensive tutorial
+ Has a cavitation model implemented and an elastohydrodynamic coupling
+ Helpful plotting and animation tools.
+ Plenty of tests within `pytest` env are available.

**Cons**

+ It seems to me that `examples/` contain more examples than described in the sphinx documentation
+ Maybe separate folders could have a short `README.md` and a short description for files on the top after the license could be provide to guide the user/contributor.
+ Target audience is not explicitly provided, but it is quite clear from the context.
+ It would be helpful to add a short paragraph on the state of the art in the field, so that we could checkmark [ ] "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

**Remarks & Questions**

+ In the tutorial, the authors mention that "the code comes with a few useful command lines for visualizations", it would be helpful to give a reference there.
+ From a physical point of view, it is strange to show absolute negative pressure in a couple of examples; should it be the pressure with the subtracted average pressure?
+ On the same topic: it was not clear from the beginning whether a lubrication model was available: in tutorials `Lubrication: wall-driven flows` and `Journal Slider, Periodic`, there's no cavitation model, therefore resulting in negative pressures.
+ Cavitation is not mentioned in the accompanying paper, however it was used in the reference paper [Holey, H. et al., Tribology Letters 70 (2022)](https://doi.org/10.1007/s11249-022-01576-5). Finally, I've ended up by finding `def bayada_chupin(...)` for the cavitation modelling and a tutorial `Elastic Deformation/Parabolic Slider, Non-Periodic`.
Nevertheless, I would recommend to mention this in the paper too.
+ **Question:** It is not clear whether we can handle problems with contact using ContactMechanics library, please specify.
+ **Performance:** For the minimal example (`journal.yaml`) and for other small examples that I run, the problem setup takes longer than the simulation itself, i.e when one executes `python -m GaPFlow -i journal.yaml`, approximately 2/3 of running time is spent on the simulation preparation and 1/3 on the simulation itself, maybe it could be improved.
+ Typo: in the last sentence of the paper, before the Acknowledgement "updated,." the comma should be removed.
