### 2026-02-11 972dad4
MAINT: store also target inputs in db (measured ones might slightly differ from target)
