+++ GaPFlow/md/md.py
@@ -173,10 +173,10 @@ class MolecularDynamics:
         for i, (Xi, Xi_t, name) in enumerate(zip(X, X_target, self._input_names)):
 
             if abs(Xi_t) < 1e-9:
-                logger.info(f'Realized input {i + 1}: {Xi:+.3e} (-----%)   ({name})')
+                logger.info(f'Measured input {i + 1}: {Xi:+.3e} (----%)   ({name})')
             else:
                 deviation = (Xi - Xi_t) / abs(Xi_t) * 100
-                logger.info(f'Realized input {i + 1}: {Xi:+.3e} ({deviation:+.1f}%)   ({name})')
+                logger.info(f'Measured input {i + 1}: {Xi:+.3e} ({deviation:+.1f}%)   ({name})')
 
         logger.info('---')
 
