### 2026-02-11 e29f8b7
ENH: apply random perturbation to new training points if too close to existing ones
