### 2026-02-18 d09d742
Fix Some Docstring Typos
While going through the docs, I found a few typos and misformatted lists in docstrings. I had Claude take a pass at things, which found a few more.
