### 2026-02-19 2f3e4ce
JOSS: Fix typo
