+++ paper/paper.md
@@ -95,7 +95,7 @@ For simulations in this *Elastohydrodynamic Regime*, `GaPFlow` uses the [Contact
 Under the assumption that the elastic deformation of the walls responds on a timescale much shorter than that of the fluid-dynamic system, the elastic response can be treated as quasi-static and represented by its steady-state solution. 
 Further assuming linear-elastic and isotropic walls, the tool utilizes a *Green's function* formulation that reduces the computational effort to a convolution operation, which can be efficiently solved in Fourier space [@stanley1997_FFT].
 The elastic deformation is determined based on the fluid pressure field and automatically adapts to the boundary conditions specified in the fluid-flow problem.
-Fluid flow and wall deformation are linked through a *weak coupling* scheme in which both solutions are iteratively updated,.
+Fluid flow and wall deformation are linked through a *weak coupling* scheme in which both solutions are iteratively updated.
 
 # Acknowledgments
 
