### 2026-02-19 968472b
WIP: separate GP animation test
