### 2026-02-19 a5a4c11
MAINT: undo Python version change and clean unused imports
