+++ pyproject.toml
@@ -26,7 +26,7 @@ license-files = [
 	"GaPFlow/_vendor/lammps/LICENSE",
 ]
 
-requires-python = ">=3.12"
+requires-python = ">=3.10"
 dynamic = ["version"]
 dependencies = [
 	"muGrid>=0.105.0",
