### 2026-02-19 a1a7d86
MAINT: update license headers
