### 2026-02-19 #9645 @DanWaxman
I've completed a review of the material and updated my checklist (apologies for being a bit slow). As full disclosure (which I discussed a bit with the editor), my expertise is in GP/Bayesian ML and I'm not very knowledgeable about the application-specific parts of the software. Generally, though, the library seems nicely implemented and documented well, if a bit terse. The GP-based components (which mostly comprise simple MLL-based optimization and some max-entropy-based active learning) appear correct.

A few high-level comments, that I think would improve the software but don't stand in the way of acceptance:
- Documentation can be quite terse in some places; I suppose part of this is my lack of subject-matter expertise, and I understand that much of this is covered in tutorials, but more abundant references (e.g., to papers, equations in papers, or tutorials) in docstrings would likely be helpful.
- There is a lots of configuration being done in YAML files. There can be merits to this approach, but it can also become unruly if you want to run lots of similarly-parameterized experiments. This is somewhat a matter of taste, but I often find Python dataclasses a much better solution since they can be inhereted and modified easily.
