### 2026-02-20 e90ef5f
DOC: notes on cavitation and negative pressure
