### 2026-02-20 54faa26
PAPER: note on solid-solid contact
