### 2026-02-20 f4af1b6
CI: separate jobs for each python version
