+++ .github/workflows/wheels.yaml
@@ -11,14 +11,15 @@ on:
 
 jobs:
   wheels:
-    name: Build wheels on ${{ matrix.os }}
+    name: Build wheels on ${{ matrix.os }} ${{ matrix.python }}
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: 
-          - ubuntu-latest
-          - macos-latest
-          - windows-latest
+        buildplat:
+        - [ubuntu-latest, manylinux, x86_64]
+        - [macos-latest, macosx, arm64]
+        - [windows-latest, win, AMD64]
+        python: ['cp310', 'cp311', 'cp312', 'cp313', 'cp314']
     steps:
       - uses: actions/checkout@v5
         with:
@@ -33,7 +34,7 @@ jobs:
       - uses: actions/upload-artifact@v4
         with:
           path: ./wheelhouse/*.whl
-          name: ${{ matrix.os }}
+          name: ${{ matrix.os }}-${{ matrix.python-version }}
 
   sdist:
     name: Build sdist
