### 2026-02-20 f104bb0
CI: separate jobs for each python version
