### 2026-02-20 5e6549c
DOC: visualization line commands page
