### 2026-02-23 a742b7b
MAINT: exclude ghost cells in active learning
