### 2026-02-23 c8f7b4c
MAINT: Don't use tuple of slices
