+++ GaPFlow/models/gp.py
@@ -504,17 +504,16 @@ class GaussianProcessSurrogate:
         """
 
         # Consicer only interior cells (no ghosts)
-        inner = slice(1, -1, None), slice(1, -1, None)
         nx, ny = var.shape
         n_inner = (nx - 2) * (ny - 2)
 
         # normalized
-        Xtest = self.Xtest.reshape(nx, ny, -1)[*inner].reshape(n_inner, -1)
+        Xtest = self.Xtest.reshape(nx, ny, -1)[1:-1, 1:-1].reshape(n_inner, -1)
         # not normalized
-        _Xtest = self._Xtest.reshape(nx, ny, -1)[*inner].reshape(n_inner, -1)
+        _Xtest = self._Xtest.reshape(nx, ny, -1)[1:-1, 1:-1].reshape(n_inner, -1)
 
         # from large to small
-        sorted_indices = jnp.argsort(var[*inner], axis=None)[::-1]
+        sorted_indices = jnp.argsort(var[1:-1, 1:-1], axis=None)[::-1]
 
         # start with largest variance (currently only implemented strategy)
         selected = sorted_indices[0]
