### 2026-02-23 5ceb7ab
MAINT: change default behavior of active learning: similarity check optional, random perturbation optional, allow max. number of 'skipped' targets
