### 2026-02-24 4513117
JOSS: add more explanation solid-solid contact
