### 2026-02-24 9b5ea2b
JOSS: add state of the field
