+++ paper/paper.md
@@ -60,13 +60,16 @@ The source term $\mathbf{s}$ accounts for fluxes across the bottom and top walls
 The current implementation uses a finite volume discretization on a regular grid and the MacCormack explicit time-integration scheme [@maccormack2003_effect] to solve the transient lubrication problem. 
 The [µGrid](https://muspectre.github.io/muGrid/) library is used to assemble the discretized density and flux fields into a unified container and to export the simulation results in the [NetCDF](https://www.unidata.ucar.edu/software/netcdf) file format.
 
-Next to the numerical inegration of the continuum equations, `GaPFlow` serves as a *glue code* that integrates the various components for multiscale or multiphysics simulations. 
+Next to the numerical integration of the continuum equations, `GaPFlow` serves as a *glue code* that integrates the various components for multiscale or multiphysics simulations. 
 Therefore, it relies on a small set of external dependencies which are summarized below.
 
 ## GP regression
 
 The fluxes required to close the macroscopic equations can either be obtained from deterministic constitutive laws or modeled using GP regression [@rasmussen2006_gaussian].
-The GP models are trained on data generated by MD, or, for testing purposes, on sparsified datasets sampled from predefined constitutive laws. 
+The GP models are trained on data generated by MD, or, for testing purposes, on sparsified datasets sampled from predefined constitutive laws.
+These include non-Newtonian fluid models capturing shear-thinning and piezoviscous effects, as well as equations of state for the fluid pressure.
+To account for cavitation, `GaPFlow` implements a homogenized gas–fluid mixture model following @bayada2013_compressible, enabling the description of compressible two-phase behavior within the lubrication framework.
+
 `GaPFlow` employs the [tinygp](https://tinygp.readthedocs.io/en/stable/index.html) library for constructing and training GP models, taking advantage of its flexibility.
 For example, it allows the implementation of custom kernels for the joint prediction of wall shear stresses at the top and bottom walls using a multi-output GP that shares a common noise process.
 Since [tinygp](https://tinygp.readthedocs.io/en/stable/index.html) is built on [JAX](https://github.com/jax-ml/jax), `GaPFlow` also benefits from automatic differentiation of the GP models, e.g. to compute the speed of sound from the pressure model.
