+++ GaPFlow/md/base.py
@@ -1,0 +1,237 @@
+#
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+from .runner import run_parallel, run_serial
+from ..utils import bordered_text, make_dumpable
+
+import os
+import abc
+import dtoolcore
+from datetime import datetime, date
+from dateutil.relativedelta import relativedelta
+from ruamel.yaml import YAML
+from getpass import getuser
+from urllib.parse import urlparse
+
+
+yaml = YAML()
+yaml.explicit_start = True
+yaml.indent(mapping=4, sequence=4, offset=2)
+
+
+class MolecularDynamics:
+    """Driver for molecular dynamics simulations.
+
+    Abstract base class for MD setup, running, and reading outputs.
+    Derived classes need to implement methods to write LAMMPS input files
+    into a dtool dataset, and to read the output of this simulation.
+
+    Attributes
+    ----------
+    name : str
+        Name of the MD object
+    params : dict
+        Parameters to control the MD setup, will be written to the dtool metadata.
+    main_file : str
+        File name of the main LAMMPS input file.
+    num_worker : int
+        Number of cores to run the parallel MD simulation.
+    is_mock : bool
+        Whether the subclass is only a mock object, which does not run an actual MD simulation.
+    """
+    __metaclass__ = abc.ABCMeta
+
+    name = str
+    params: dict
+    main_file: str
+    num_worker: int
+    is_mock: bool
+    _dtool_basepath: str = '/tmp/'
+    _readme_template: str = ""
+    _input_names: list[str] = ['ρ', 'jx', 'jy', 'h', '∂h/∂x', '∂h/∂y'] + [f'extra_{i}' for i in range(10)]
+    _ascii_art: str = r"""
+  _        _    __  __ __  __ ____  ____
+ | |      / \  |  \/  |  \/  |  _ \/ ___|
+ | |     / _ \ | |\/| | |\/| | |_) \___ \
+ | |___ / ___ \| |  | | |  | |  __/ ___) |
+ |_____/_/   \_\_|  |_|_|  |_|_|   |____/
+
+"""
+
+    @property
+    def dtool_basepath(self):
+        """File location, where dtool datasets are written into (default is '/tmp/')."""
+        return self._dtool_basepath
+
+    @dtool_basepath.setter
+    def dtool_basepath(self, name):
+        self._dtool_basepath = name
+
+    @abc.abstractmethod
[…]
