+++ GaPFlow/md/__init__.py
@@ -21,4 +21,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 #
-from .runner import MolecularDynamics, Mock, LennardJones, GoldAlkane, PARALLEL  # noqa: F401
+from .runner import PARALLEL  # noqa: F401
+from .mock import Mock  # noqa: F401
+from .lj import LennardJones  # noqa: F401
+from .mol import GoldAlkane  # noqa: F401

[…]
