### 2026-02-24 bbc3366
MAINT: update license headers
