+++ GaPFlow/problem.py
@@ -1,5 +1,6 @@
 #
-# Copyright 2025-2026 Hannes Holey
+# Copyright 2026 Dan Waxman
+#           2025-2026 Hannes Holey
 #           2025-2026 Christoph Huber
 #
 # ### MIT License
