### 2026-02-24 2ec7f31
DOC: fix unfound attribute, and 'wrong' script example
