+++ GaPFlow/md/__init__.py
@@ -22,6 +22,7 @@
 # SOFTWARE.
 #
 from .runner import PARALLEL  # noqa: F401
+from .base import MolecularDynamics  # noqa: F401
 from .mock import Mock  # noqa: F401
 from .lj import LennardJones  # noqa: F401
 from .mol import GoldAlkane  # noqa: F401
