### 2026-02-25 0487926
ENH: enable GP animations from problem object
