### 2026-02-25 7ca5963
ENH: consistent PSPG and GLS stabilization scheme, without inertia
