+++ GaPFlow/solver_fem_2d.py
@@ -25,21 +25,17 @@
 from . import HAS_PETSC
 from .fem_2d.elements import TriangleQuadrature
 from .fem_2d.terms import NonLinearTerm, get_active_terms
-from .fem_2d.assembly_layout import (
-    FEMAssemblyLayout,
-    MatrixCOOPattern,
-    RHSPattern,
-    JacobianTermMap,
-    JacobianTermTemplate,
-    JacobianTermRef,
-    COOLookup,
-)
-from .fem_2d.scaling import compute_characteristic_scales
+from .fem_2d.assembly import Assembly
+from .fem_2d.scaling import build_scaling
 from .fem_2d.grid_index import GridIndexManager
 from .fem_2d.quad_fields import QuadFieldManager
 from .fem_2d.solution_guards import apply_guards
+from .fem_2d.residual_analysis import (
+    create_residual_analysis_plot,
+    plot_residual_history,
+    print_residual_analysis,
+)
 
-from functools import cached_property
 import numpy as np
 import time
 from muGrid import Timer
@@ -59,18 +55,13 @@ class FEMSolver2D:
     def __init__(self, fem_spec: dict, problem: "Problem") -> None:
         self.fem_spec = fem_spec
         self.problem = problem
-        self.quad_mgr = None  # QuadFieldManager, initialized via init()
-        self.linear_solver = None  # Linear solver (PETSc or SciPy)
         self.timer = Timer()
         self.R_norm_history = []  # Nested list: [[step1_iters], [step2_iters], ...]
 
     def _init_convenience_accessors(self) -> None:
         """Initialize convenience accessors for problem and grid properties."""
         p = self.problem
-        self.decomp = p.decomp
         self.energy = p.fem_solver['equations']['energy']
-
-        self.global_coords = p.decomp.icoordsg
 
         nb_subdomain_pts = p.decomp.nb_subdomain_grid_pts
         self.Nx_inner = nb_subdomain_pts[0]
@@ -78,7 +69,6 @@ class FEMSolver2D:
 
         self.dx = p.grid['dx']
         self.dy = p.grid['dy']
-        self.A = self.dx * self.dy
 
         self.variables = ['rho', 'jx', 'jy']
         self.residuals = ['mass', 'momentum_x', 'momentum_y']
@@ -96,33 +86,13 @@ class FEMSolver2D:
             energy_spec=energy_spec,
         )
 
-        # Convenience aliases from grid index manager
         self.nb_inner_pts = self.grid_idx.nb_inner_pts
-        self.sq_per_row = self.grid_idx.sq_per_row
-        self.sq_per_col = self.grid_idx.sq_per_col
-        self.nb_sq = self.grid_idx.nb_sq
-        self.sq_x_arr = self.grid_idx.sq_x_arr
-        self.sq_y_arr = self.grid_idx.sq_y_arr
 
         self.res_size = len(self.residuals) * self.nb_inner_pts
         self.var_size = len(self.variables) * self.grid_idx.nb_contributors
 
-        # Quadrature data for triangular elements
-        quad = TriangleQuadrature(self.dx, self.dy)
-        w = quad.weights
-        Nl, Nr = quad.N_left, quad.N_right
-
-        # Precomputed element tensors: elem[i,j,q] = N_i[q] * N_j[q] * w[q] * A
-        self.elem_tensor_left = Nl[:, None, :] * Nl[None, :, :] * w * self.A
-        self.elem_tensor_right = Nr[:, None, :] * Nr[None, :, :] * w * self.A
-
-        # Precomputed test function integrals: integral[i,q] = N_i[q] * w[q] * A
-        self.test_wA_left = Nl * w * self.A
[…]
