### 2026-02-25 7fc5608
MAINT: small fixes to plotting
