### 2026-02-25 79fd155
DOC: added new examples for 2D simulations and GP with MD data
