### 2026-02-25 0583d36
MAINT: duplicate class definition
