### 2026-02-25 0b7dc01
MAINT: update license headers
