+++ GaPFlow/viz/plotting.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2025 Hannes Holey
+# Copyright 2025-2026 Hannes Holey
 #           2025 Christoph Huber
 #
 # ### MIT License
