### 2026-02-25 ff53798
DOC: update function and module level docstrings
