### 2026-02-25 809af43
MAINT: enable notebook execution in doc
