+++ _sources/api/summary/GaPFlow.models.pressure.rst
@@ -1,0 +1,19 @@
+﻿GaPFlow.models.pressure
+=======================
+
+.. automodule:: GaPFlow.models.pressure
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      bayada_chupin
+      bwr
+      cubic
+      dowson_higginson
+      eos_pressure
+      murnaghan_tait
+      power_law
+      van_der_waals
+   
