+++ _sources/api/summary/GaPFlow.models.viscosity.rst
@@ -1,0 +1,21 @@
+﻿GaPFlow.models.viscosity
+========================
+
+.. automodule:: GaPFlow.models.viscosity
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      barus_piezo
+      carreau_shear
+      dukler_mixture
+      eyring_shear
+      mc_adams_mixture
+      piezoviscosity
+      roelands_piezo
+      shear_rate_avg
+      shear_thinning_factor
+      srate_wall_newton
+   
