+++ _sources/api/summary/GaPFlow.models.viscous.rst
@@ -1,0 +1,14 @@
+﻿GaPFlow.models.viscous
+======================
+
+.. automodule:: GaPFlow.models.viscous
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      stress_avg
+      stress_bottom
+      stress_top
+   
