+++ _sources/api/summary/GaPFlow.models.sound.rst
@@ -1,0 +1,19 @@
+﻿GaPFlow.models.sound
+====================
+
+.. automodule:: GaPFlow.models.sound
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      bayada_chupin
+      bwr
+      cubic
+      dowson_higginson
+      eos_sound_velocity
+      murnaghan_tait
+      power_law
+      van_der_waals
+   
