### 2026-02-26 #9645 @hannes-holey
We thank the reviewers for taking their time to review our submission and the constructive feedback. Please find our response to the issues raised by the reviewers below. We collected all changes to the code and the documentation related to the reviewer’s comments in hannes-holey/GaPFlow#60. We merged this PR already, to trigger deployment of the online documentation.

Response to @vyastreb's comments:

*It seems to me that examples/ contain more examples than described in the sphinx documentation*

Indeed, there is a larger collection of configuration files in the examples/ directory than in the documentation. We identified that there are mainly two components missing in the tutorials which were covered in the examples: 2D simulations and multiscale simulations with *real* MD data. We now added both to the tutorials in the Sphinx documentation. Since running MD simulations is time-consuming, we uploaded training datasets for the two implemented MD systems (Lennard-Jones & gold/alkane) to [Zenodo](https://zenodo.org/records/18761223). The notebooks in the tutorials fetch these datasets and use them in simulations without active learning, but interested users can of course try to reproduce and/or extend the MD datasets. See hannes-holey/GaPFlow@79fd155.

*Maybe separate folders could have a short README.md and a short description for files on the top after the license could be provide to guide the user/contributor.*

We did not add separate README files to different directories but added more module level docstrings which will now also appear in the automatically generated API reference (see hannes-holey/GaPFlow@ff53798). We will continuously work on improving the documentation.

*Target audience is not explicitly provided, but it is quite clear from the context.*

We agree that it should be clear from the context. No additions have been made.

*It would be helpful to add a short paragraph on the state of the art in the field, so that we could checkmark [ ] "State of the field: Do the authors describe how this software compares to other commonly-used packages?"*

We added a paragraph to the paper (see hannes-holey/GaPFlow@9b5ea2b).

*In the tutorial, the authors mention that "the code comes with a few useful command lines for visualizations", it would be helpful to give a reference there.*

We agree with the reviewer and added an overview page for these command line tools to the documentation and included links to it (see hannes-holey/GaPFlow@5e6549c).

*From a physical point of view, it is strange to show absolute negative pressure in a couple of examples; should it be the pressure with the subtracted average pressure?*

Large negative pressures for liquids as in some of the examples are indeed unphysical. The reason for this is that the used compressibility models (such as Dowson-Higginson) are not suited for the gas phase or mixtures of liquid and gases. Furthermore, cavitation models such as the JFO formalism as in classical Reynolds calculations do not apply here. Thus, we implemented cavitation through Bayada & Chupin’s equation of state for fluid vapor mixtures. A similar behavior is expected from our MD data-driven models.

Nevertheless, the numerical solution—even with an unphysical equation of state—is technically correct and is useful for validating our simulation with analytical solutions that do not include cavitation either, such as in tutorial 5 in the limit of nearly incompressible fluids. We added clarifying remarks to the corresponding examples (see hannes-holey/GaPFlow@e90ef5f163c283f60dbb6c8fb0ca5a3c1f4bc1aa)

*On the same topic: it was not clear from the beginning whether a lubrication model was available: in tutorials Lubrication: wall-driven flows and Journal Slider, Periodic, there's no cavitation model, therefore resulting in negative pressures.*

and

*Cavitation is not mentioned in the accompanying paper, however it was used in the reference paper[ Holey, H. et al., Tribology Letters 70 (2022)](https://doi.org/10.1007/s11249-022-01576-5). Finally, I've ended up by finding def bayada_chupin(...) for the cavitation modelling and a tutorial Elastic Deformation/Parabolic Slider, Non-Periodic. Nevertheless, I would recommend to mention this in the paper too.*

We agree that the cavitation model was somewhat hidden. We added a section to the paper that mentions the implemented model (see hannes-holey/GaPFlow@7677ad4).

*Question: It is not clear whether we can handle problems with contact using ContactMechanics library, please specify.*

Currently, there is no support for contact and the ContactMechanics package is only used for the calculation of the elastic deformation based on the fluid pressure, i.e. we assume a continuous fluid film throughout the full simulation domain. The main objective of our multiscale framework is to use molecular insights to improve predictions at the nanometer scale instead of relying on ad-hoc assumptions about asperity contact. We added a clarification to the paper (see hannes-holey/GaPFlow@4513117).

*Performance: For the minimal example (journal.yaml) and for other small examples that I run, the problem setup takes longer than the simulation itself, i.e when one executes python -m GaPFlow -i journal.yaml, approximately 2/3 of running time is spent on the simulation preparation and 1/3 on the simulation itself, maybe it could be improved.*

The minimal example is limited to 200 time steps to check the basic functionality, which is why the runtime of the actual simulation is much shorter than setting up the system. Nevertheless, we profiled the script and found that most of the time is spent importing the problem class `from GaPFlow import Problem` (~2s) while the system setup `Problem.from_yaml(‘input.yaml’)` is not particularly slow (~1 ms). Thus, we don’t think there is much we can do here.

*Typo: in the last sentence of the paper, before the Acknowledgement "updated,." the comma should be removed.*

Thank you. Fixed in hannes-holey/GaPFlow@2f3e4ce.

---

Response to @DanWaxman's comments:

*Documentation can be quite terse in some places; I suppose part of this is my lack of subject-matter expertise, and I understand that much of this is covered in tutorials, but more abundant references (e.g., to papers, equations in papers, or tutorials) in docstrings would likely be helpful.*

We agree that the API documentation is terse in some places. We updated some of the docstrings and included more modules in the API reference in hannes-holey/GaPFlow@ff53798d4cb6a119589ae1dafb96449ead5ca136. We will continuously work on improving the documentation.

*There is a lots of configuration being done in YAML files. There can be merits to this approach, but it can also become unruly if you want to run lots of similarly-parameterized experiments. This is somewhat a matter of taste, but I often find Python dataclasses a much better solution since they can be inhereted and modified easily.*

We agree that this is a matter of taste. The decision to use YAML files was done a long time ago and we prefer not to change this at the moment. Nevertheless, for large parametric studies, there is also the option to create instances of a problem from a dictionary, which might be more convenient.
