### 2026-02-26 hannes-holey/hans#48 @hannes-holey
I am closing this as we now have wheels with packaged LAMMPS for the most common systems.
