### 2026-02-27 aec2530
ENH: wall velocity on both sides
