+++ GaPFlow/models/profiles.py
@@ -1,5 +1,6 @@
 #
 # Copyright 2025 Hannes Holey
+#           2026 Christoph Huber
 #
 # ### MIT License
 #
@@ -24,8 +25,12 @@
 
 # flake8: noqa: W503
 
-def get_velocity_profiles(z, q, Ls=0.0, U=1.0, V=0.0, slip="both"):
-    """Velocity profiles for a given flow rate and wall velcoty
+import numpy as np
+
+
+def get_velocity_profiles(z, q, U_bot=0.0, V_bot=0.0, U_top=0.0, V_top=0.0,
+                          Ls_bot=0.0, Ls_top=0.0):
+    """Velocity profiles for a given flow rate and wall velocity
 
     Parameters
     ----------
@@ -34,15 +39,18 @@ def get_velocity_profiles(z, q, Ls=0.0, U=1.0, V=0.0, slip="both"):
     q : array-like
         Height-averaged solution, (rho, jx, jy) for a single point (shape=(3,))
         or a field (shape=(3, nx, ny))
-    Ls : float, optional
-        Slip length (the default is 0.0, which means no-slip)
-    U : float, optional
-        Lower wall velocity in x direction (the default is 1.0)
-    V : float, optional
-        Lower wall velocity in y direction (the default is 1.0)
-    slip : str, optional
-        Type of slip boundary conditions ("both", "top", "bottom", or "none")
-        (the default is "both", which means Ls applies to both the upper and lower wall)
+    U_bot : float, optional
+        Bottom wall velocity in x direction (the default is 0.0)
+    V_bot : float, optional
+        Bottom wall velocity in y direction (the default is 0.0)
+    U_top : float, optional
+        Top wall velocity in x direction (the default is 0.0)
+    V_top : float, optional
+        Top wall velocity in y direction (the default is 0.0)
+    Ls_bot : float, optional
+        Slip length at bottom wall (the default is 0.0, which means no-slip)
+    Ls_top : float, optional
+        Slip length at top wall (the default is 0.0, which means no-slip)
 
     Returns
     -------
@@ -52,88 +60,52 @@ def get_velocity_profiles(z, q, Ls=0.0, U=1.0, V=0.0, slip="both"):
 
     h = z[-1]
 
-    if slip == "both":
-        u = (
-            12 * Ls**2 * h * q[1]
-            + 4 * Ls * U * h ** 2 * q[0]
-            - 12 * Ls * U * h * q[0] * z
-            + 6 * Ls * U * q[0] * z**2
-            + 6 * Ls * h ** 2 * q[1]
-            + 12 * Ls * h * q[1] * z
-            - 12 * Ls * q[1] * z**2
-            + U * h ** 3 * q[0]
-            - 4 * U * h ** 2 * q[0] * z
-            + 3 * U * h * q[0] * z**2
-            + 6 * h ** 2 * q[1] * z
-            - 6 * h * q[1] * z**2
-        ) / (h * q[0] * (12 * Ls**2 + 8 * Ls * h + h ** 2))
-        v = (
-            12 * Ls**2 * h * q[2]
[…]
