### 2026-03-02 0060126
ENH: Domain initialization with dry contact solution
