+++ GaPFlow/solver_fem_2d.py
@@ -199,8 +199,8 @@ class FEMSolver2D:
         if len(template.template_coo_idx) == 0:
             return
 
-        res_deriv = self.get_res_deriv_vals(term, dep_var)
-        res_vals = res_deriv.ravel()[template.flat_field_idx]
+        res_deriv = self.get_res_deriv_vals(term, dep_var)  # (6, sq_per_row, sq_per_col)
+        res_vals = res_deriv.ravel()[template.flat_field_idx] # (nb_sq * 54) = (3 * 3 * np_quad)
 
         # Compute actual COO indices from template + block offset
         actual_coo_idx = template.template_coo_idx + ref.block_offset
