+++ GaPFlow/models/dry_contact.py
@@ -1,0 +1,267 @@
+#
+# Copyright 2026 Christoph Huber
+#
+# ### MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+
+from datetime import datetime
+from mpi4py import MPI
+import numpy as np
+import copy
+import os
+
+from ContactMechanics.Systems import NonSmoothContactSystem
+from ContactMechanics import PeriodicFFTElasticHalfSpace
+from SurfaceTopography import Topography
+
+from ..topography import Topography as GaPFlowTopography
+
+
+class DryContact:
+
+    def __init__(self, input_dict, dir):
+
+        self.domain_inlet = input_dict['force_balance']['init_dry_contact']['domain_inlet']
+        self.domain_outlet = input_dict['force_balance']['init_dry_contact']['domain_outlet']
+        self.domain_sides = input_dict['force_balance']['init_dry_contact']['domain_sides']
+
+        assert self.domain_inlet > 0, "domain_inlet must be positive"
+        assert self.domain_outlet > 0, "domain_outlet must be positive"
+        assert self.domain_sides > 0, "domain_sides must be positive"
+
+        self.geo = input_dict['geometry']
+        self.grid = input_dict['grid']
+        self.force = self.get_force(input_dict)  # check if force or pressure is specified
+
+        self.elastic = input_dict['properties']['elastic']
+        assert self.elastic['enabled'], "Elastic deformation must be enabled for dry contact initialization"
+
+        self.dir = dir
+        assert os.path.isdir(dir), f"Directory {dir} does not exist. Please check the path of the input YAML file."
+
+        self.input_dict = input_dict
+
+    def main(self):
+
+        xx, yy = self.get_initial_grid()
+        h = self.get_topography(xx, yy)
+        p, u = self.solve_contact(h)
+        contact_bounds = self.get_bounding_box(p, xx, yy)
+        domain_bounds = self.get_domain_bounds(contact_bounds)
+        h_new = self.update_input_dict(h, domain_bounds)
+
+        debug_plot(p, u, contact_bounds, domain_bounds, self.grid, h, h_new)
+
+        return self.input_dict_new
+
+    def update_input_dict(self, h, domain_bounds):
+
+        self.input_dict_new = copy.deepcopy(self.input_dict)
[…]

[…]
