### 2026-03-04 288d6ff
MAINT: improved topography structure, several clean-ups
