### 2026-03-04 #9645 @matthewfeickert
@vyastreb, @DanWaxman, @hannes-holey to keep the review tracking issue from becoming increasingly complex to follow (given that not all notifications are relevant to all participants) please have all relevant discussions on issues or questions for the review in GitHub Issues on https://github.com/hannes-holey/GaPFlow/issues that reference this Issue (`https://github.com/openjournals/joss-reviews/issues/9645`). It would be great to see https://github.com/openjournals/joss-reviews/issues/9645#issuecomment-3881073811 and https://github.com/openjournals/joss-reviews/issues/9645#issuecomment-3930018560 ported over as Issues with the relevant components of https://github.com/openjournals/joss-reviews/issues/9645#issuecomment-3966115517. Iteration can happen as needed and then once things have concluded please report the status here.

@vyastreb, @DanWaxman, aside from this, if you are able to respond (on the to-be-created GitHub issues) in the next week that would be useful to given @hannes-holey's nice revision work.
