### 2026-03-18 7c4e70c
MAINT: interim save of all local tests and analyses
