+++ dry_contact/generate_dimpled_topography.py
@@ -1,0 +1,124 @@
+"""
+Generate a flat surface with four square dimples (sharp height jumps) for
+texturing + cavitation tests.  Each dimple is a rectangular well with a
+near-step edge profile (smoothed over ~1 cell via tanh) to avoid spurious
+Gibbs-type noise while keeping the jump visually sharp.
+
+Saves the height field as a .npy file in the topography/ subfolder and plots it.
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+import os
+
+# --- Grid parameters (must match the config) ---
+Nx = 128
+Ny = 128
+Lx = 0.002   # [m]
+Ly = 0.002   # [m]
+
+# --- Flat baseline height ---
+h_flat = 5.0e-6   # [m]
+
+# --- Dimple parameters ---
+dimple_depth  = 3.0e-6   # [m]   extra height inside each dimple
+dimple_half_w = 2.5e-4   # [m]   half-width (and half-height) of the square well
+# Transition width: ~1.5 cell widths for a near-step but well-resolved edge
+dx = Lx / Nx
+transition = 1.5 * dx    # [m]
+
+def square_well(XX, YY, x0, y0, hw, depth, t):
+    """Return depth * smooth_square_indicator centred at (x0,y0) with half-width hw."""
+    # Product of two tanh "ramp-down" functions along x and y
+    wx = 0.5 * (np.tanh((XX - (x0 - hw)) / t) - np.tanh((XX - (x0 + hw)) / t))
+    wy = 0.5 * (np.tanh((YY - (y0 - hw)) / t) - np.tanh((YY - (y0 + hw)) / t))
+    return depth * wx * wy
+
+# Four dimple centres at (±Lx/4, ±Ly/4)
+cx = Lx / 4.0
+cy = Ly / 4.0
+centres = [
+    ( cx,  cy),
+    (-cx,  cy),
+    ( cx, -cy),
+    (-cx, -cy),
+]
+
+# --- Build coordinate arrays (cell centres) ---
+dy = Ly / Ny
+x = np.linspace(-Lx/2 + dx/2, Lx/2 - dx/2, Nx)
+y = np.linspace(-Ly/2 + dy/2, Ly/2 - dy/2, Ny)
+XX, YY = np.meshgrid(x, y, indexing='ij')   # shape (Nx, Ny)
+
+# --- Compute height field ---
+h = h_flat * np.ones((Nx, Ny))
+for (x0, y0) in centres:
+    h += square_well(XX, YY, x0, y0, dimple_half_w, dimple_depth, transition)
+
+# Clip so height is never negative
+h = np.maximum(h, 1.0e-8)
+
+# --- Save ---
+out_dir = os.path.join(os.path.dirname(__file__), 'topography')
+os.makedirs(out_dir, exist_ok=True)
+out_file = os.path.join(out_dir, 'dimpled_flat.npy')
+np.save(out_file, h)
+print(f"Saved height field {h.shape} to: {out_file}")
+print(f"  h_min = {h.min():.3e} m,  h_max = {h.max():.3e} m")
+
+# --- Plot ---
+# Downsample for the 3-D surface (full 128x128 is fine, but stride keeps it crisp)
+stride = 1
+xs = x[::stride] * 1e3        # mm
+ys = y[::stride] * 1e3        # mm
+hs = h[::stride, ::stride] * 1e6   # µm
+XS, YS = np.meshgrid(xs, ys, indexing='ij')
+
+fig = plt.figure(figsize=(16, 5))
+
[…]
