### 2026-03-19 #9645 @DanWaxman
Hi! I am satisfied with the authors responses, and appreciate the authors' efforts at more approachable documentation.
