### 2026-03-19 #9645 @vyastreb
Apologies, @hannes-holey, for my late reaction. I am fully satisfied with all your comments and answers and with all improvements for the documentation that you implemented. Thank you! I am happy to recommend your paper and software for JOSS!
@matthewfeickert, I am sorry for such a long delay.
