### 2026-03-25 d60945f
WIP: matching FD, working on diffusive terms
