+++ GaPFlow/fem_2d/grid_index.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2025 Christoph Huber
+# Copyright 2026 Christoph Huber
 #
 # ### MIT License
 #
@@ -21,10 +21,23 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 #
-"""Grid index management for FEM assembly.
+"""Grid index management for Taylor-Hood P2P1 FEM assembly.
 
-Handles index masks and boundary condition handling for triangular FEM
-on structured grids with domain decomposition support.
+Two grids are managed by a single GridIndexManager:
+  - Density (P1): coarse grid, Nx_p x Ny_p inner nodes, ghost depth 1.
+  - Mass flux (P2): fine grid, Nx_v x Ny_v inner nodes, ghost depth 2.
+
+Index mask convention:
+  - index_mask_inner_local: inner nodes get sequential indices 0..(Nx*Ny)-1
+    (row-major: x varies fastest); ghost nodes are -1.
+  - index_mask_padded_local: equals inner mask on inner nodes; ghost nodes
+    that represent DOF are assigned new sequential indices
+      - inter-subdomain ghost nodes (DOFs that live on another process)
+      - Neumann ghost nodes (reflecting DOF on adjacent inner node)
+      - Dirichlet ghost nodes (no DOF, stay -1)
+      - serial periodic ghost nodes (reflecting DOF on opposite side of domain)
+  - index_mask_padded_global: unique global indices from DomainDecomposition.
+  - l2g_list: maps local contributor index -> global index.
 """
 from functools import cached_property, lru_cache
 from typing import List, TYPE_CHECKING
@@ -40,199 +53,359 @@ IntArray = npt.NDArray[np.signedinteger]
 
 
 class GridIndexManager:
-    """Manages grid index masks and element connectivity for FEM assembly.
+    """Index masks and square connectivity for Taylor-Hood P2P1 assembly.
 
-    Handles:
-    - Index masks for inner (residual) and padded (contributor) grids
-    - Square element coordinates and corner connectivity
-    - Local-to-global point mapping
-    - Boundary condition handling (Dirichlet removal, Neumann forwarding)
-    - Does NOT know about triangular elements and quadrature
+    Manages two grids simultaneously:
+      - Density (P1) on the coarse grid via decomp._decomp (ghost depth 1).
+      - Mass flux (P2) on the fine grid via decomp._decomp_v (ghost depth 2).
+
+    Parameters
+    ----------
+    decomp : DomainDecomposition
+        Must have been initialised with solver='fem'.
+    variables : list of str
+        Ordered variable names, e.g. ['rho', 'jx', 'jy']. Used to look up
+        per-variable BC types.
     """
 
     def __init__(self,
                  decomp: "DomainDecomposition",
-                 variables: List[str],
-                 energy_spec: dict = None):
+                 variables: List[str]):
+
         self._decomp = decomp
         self._variables = variables
-        grid = decomp.grid
 
-        # Grid dimensions from decomp
-        Nx_inner, Ny_inner = decomp.nb_subdomain_grid_pts
-        self.Nx_inner = Nx_inner
-        self.Ny_inner = Ny_inner
-        self.Nx_padded = Nx_inner + 2
-        self.Ny_padded = Ny_inner + 2
+        # ------------------------------------------------------------------
+        # Density (P1) grid dimensions
+        # ------------------------------------------------------------------
+        Nx_p, Ny_p = decomp.nb_subdomain_grid_pts
+        self.Nx_p_inner  = Nx_p
+        self.Ny_p_inner  = Ny_p
[…]
