+++ GaPFlow/fem_2d/quad_fields.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2025 Christoph Huber
+# Copyright 2026 Christoph Huber
 #
 # ### MIT License
 #
@@ -21,36 +21,33 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 #
-"""Quadrature field management for FEM assembly.
 
-Handles field storage, interpolation operators, update methods, and access
-for nodal and quadrature point fields on triangular elements.
-"""
-from typing import Any, Dict, List, Set, TYPE_CHECKING
+from typing import Dict, List, Set, TYPE_CHECKING
 
 import numpy as np
 import numpy.typing as npt
 from muGrid import Field
+from scipy.ndimage import zoom
 
-from .elements import TriangleQuadrature
+from .elements import TaylorHoodP2P1
 
 if TYPE_CHECKING:
     from ..problem import Problem
+    from ..parallel import DomainDecomposition
 
 NDArray = npt.NDArray[np.floating]
 
 
-# Field name sets for different physics
+# ---------------------------------------------------------------------------
+# Field name sets
+# ---------------------------------------------------------------------------
+
 BASE_FIELDS = {
     'rho', 'jx', 'jy',
     'p', 'h', 'dh_dx', 'dh_dy', 'eta',
     'U_bot', 'V_bot', 'U_top', 'V_top', 'Ls',
     'dp_drho',
     'rho_prev', 'jx_prev', 'jy_prev',
-    'tau_mass',
-    'tau_mom',
-    'tau_pspg',
-    'tau_gls',
     'force_x', 'force_y',
 }
 
@@ -69,390 +66,241 @@ ENERGY_FIELDS = {
     'T', 'dT_drho', 'dT_djx', 'dT_djy', 'dT_dE',
     'S', 'dS_drho', 'dS_djx', 'dS_djy', 'dS_dE',
     'E_prev',
-    'tau_energy',
 }
+
+# Fields whose nodal values live on the fine (mass-flux) grid
+_FINE_GRID_FIELDS = {'jx', 'jy'}
 
 
 class QuadFieldManager:
-    """Manages quadrature fields, operators, and updates for FEM.
-
-    Handles:
-    - Field storage (nodal and quadrature point fields)
-    - Interpolation operators (nodal → quad, derivatives)
-    - Field access methods (get values, compute derivatives)
-    - Physics updates (nodal fields, interpolation, derived quantities)
-    - Time stepping (store previous values)
+    """Manages nodal and quadrature fields for Taylor-Hood P2P1 assembly.
 
     Parameters
     ----------
     problem : Problem
-        Problem instance for physics model access.
+        Problem instance (coarse-grid state, physics models).
     energy : bool
-        Whether energy equation is active.
+        Whether the energy equation is active.
     variables : list of str
-        Variable names for time stepping (e.g., ['rho', 'jx', 'jy']).
+        Newton variables in order, e.g. ['jx', 'jy', 'rho'] or with 'E'.
+    elements : TaylorHoodP2P1
+        Element instance carrying P1 and P2 operators.
+    decomp : DomainDecomposition
+        Decomposition holding both coarse (_p) and fine (_v) grid info.
     """
 
-    def __init__(self, problem: "Problem", energy: bool, variables: List[str],
-                 quad: TriangleQuadrature):
+    def __init__(self, problem: "Problem",
+                 energy: bool,
+                 variables: List[str],
+                 elements: TaylorHoodP2P1,
+                 decomp: "DomainDecomposition") -> None:
         self.problem = problem
         self.energy = energy
         self.variables = variables
[…]
