### 2026-03-25 cf506c4
WIP: first working version for periodic/semi-periodic
