+++ GaPFlow/fem_2d/tests/test_parabolic_slider.py
@@ -23,7 +23,7 @@ CONFIG_DIR = Path(__file__).parent / 'configs'
 # =============================================================================
 
 Nx = 200
-Ny = 4
+Ny = 2
 max_it = 10
 dt = 0.1
 
