### 2026-03-25 f6fd978
WIP: works on non-periodic cases with increased quadrature
