+++ GaPFlow/fem_2d/tests/test_inclined_slider.py
@@ -22,7 +22,7 @@ CONFIG_DIR = Path(__file__).parent / 'configs'
 # =============================================================================
 
 Nx = 100
-Ny = 1
+Ny = 3
 max_it = 30
 dt = 0.1
 
