+++ tests/configs/poiseuille_2d_body_force.yaml
@@ -23,9 +23,9 @@ geometry:
 
 numerics:
     solver: fem
-    dt: 0.01
+    dt: 0.001
     tol: 1e-10
-    max_it: 1000
+    max_it: 5
 
 properties:
     EOS: PL
@@ -34,25 +34,22 @@ properties:
     shear: 0.1
     bulk: 0.
     alpha: 1e-8
-    force_x: 1.0              # Body force in x (drives flow)
+    force_x: 1e-05            # Body force in x (drives flow)
     force_y: 0.0
 
 fem_solver:
     type: newton_alpha
     dynamic: True
-    alpha: 1.0
+    alpha: 0.01
     R_norm_tol: 1e-8
-    linear_solver: direct
-    pspg_boundary_decay: 2.0
     max_iter: 20
     physics:
         gap_shear: False
         plane_shear: True
         body_force: True
-        pspg: true
-        gls: true
 
 options:
     output: /tmp/poiseuille_2d_test
-    write_freq: 100
-    save_output: False
+    write_freq: 1
+    save_output: True
+    output_plots: True

[…]
