### 2026-03-27 53ff09e
MAINT: NetCDF files are always closed and re-opened in append mode to ensure readability
