### 2026-04-01 cf51bdb
MAINT: initial length scales
