### 2026-04-08 ac22a00
WIP: status before pressure gradient term change
