+++ GaPFlow/solver_fem_2d.py
@@ -42,6 +42,7 @@ from .fem_2d.assembly import Assembly
 from .fem_2d.terms import NonLinearTerm, get_active_terms
 from .fem_2d.scaling import build_scaling
 from .fem_2d.scipy_system import ScipySystem
+from .fem_2d.solution_guards import apply_guards
 
 if TYPE_CHECKING:
     from .problem import Problem
@@ -169,7 +170,7 @@ class FEMSolver2d:
         all_field_names = self.quad_mgr._needed_fields()
         for term in self.terms:
             ctx = {name: make_getter(name) for name in all_field_names}
-            ctx['dt'] = p.numerics['dt']
+            ctx['dt'] = lambda: p.numerics['dt']
             if self.energy:
                 ctx['k'] = lambda: p.energy.k
             term.build(ctx)
@@ -196,6 +197,38 @@ class FEMSolver2d:
 
         self.scaling = build_scaling(
             self.problem, self.energy, self.variables, self.assembly)
+
+        debug_from = self.problem.fem_solver.get('newton_debug', None)
+        if debug_from is not None:
+            from .fem_2d.newton_debug import NewtonDebugger
+            self.debugger = NewtonDebugger(
+                output_dir=self.problem.options['output'],
+                variables=self.variables,
+                residuals=self.residuals,
+                res_slices=self.assembly._res_slices,
+                sol_slices=self.assembly._sol_slices,
+                Nx_p=self.grid_idx.Nx_p_inner,
+                Ny_p=self.grid_idx.Ny_p_inner,
+                Nx_v=self.grid_idx.Nx_v_inner,
+                Ny_v=self.grid_idx.Ny_v_inner,
+                terms=self.terms,
+                problem=self.problem,
+            )
+            self._debug_from = debug_from
+            self._debug_steps_done = 0
+        else:
+            self.debugger = None
+            self._debug_from = None
+            self._debug_steps_done = 0
+
+    @property
+    def _debug_active(self) -> bool:
+        """True when newton_debug is enabled and the step threshold has been reached."""
+        return (
+            self.debugger is not None
+            and self.problem.step >= self._debug_from
+            and self._debug_steps_done < 5
+        )
 
     # =========================================================================
     # Quadrature field update
@@ -307,7 +340,7 @@ class FEMSolver2d:
         np.add.at(M, (block_rows, block_cols), coo)
         return M
 
-    def get_R(self, qf: dict = None) -> NDArray:
+    def get_R_(self, qf: dict = None) -> NDArray:
         """Assemble residual vector."""
         if qf is None:
             qf = self._build_all_quad_fields()
@@ -323,11 +356,30 @@ class FEMSolver2d:
 
     def solver_step_fun(self, q_guess: NDArray) -> Tuple[NDArray, NDArray]:
         self.set_q_nodal(q_guess)
+        self._exchange_ghosts()
         self.update_quad()
         qf = self._build_all_quad_fields()
         M = self.get_M(qf)
-        R = self.get_R(qf)
+        R = self.get_R_(qf).copy()
+        if self._debug_active:
+            self._last_R_per_term = self.assembly.assemble_rhs_per_term(qf, self.terms)
         return M, R
+
+    def get_R(self, q_guess: NDArray) -> float:
+        self.set_q_nodal(q_guess)
+        self._exchange_ghosts()
+        self.update_quad()
+        qf = self._build_all_quad_fields()
+        R = self.get_R_(qf).copy()
+        return R
+
+    def get_R_norm_global(self, R: NDArray) -> float:
+        p = self.problem
+        comm = p.decomp._mpi_comm
+        R_norm_local_sq = float(np.linalg.norm(R) ** 2)
[…]
