### 2026-04-13 9399761
MAINT: write csv files during simulation run instead of only once
