### 2026-04-14 8479e97
MAINT: modify history plot
