### 2026-04-15 2d78866
MAINT: use variable random seed for initial velocity distribution
