### 2026-04-16 e42b1dc
MAINT: pause active learning for a few steps at simulation start
