+++ GaPFlow/models/sound.py
@@ -68,6 +68,10 @@ def eos_sound_velocity(density, prop):
     elif prop['EOS'] == 'Bayada':
         func = bayada_chupin
         args = ['rho_l', 'rho_v', 'c_l', 'c_v']
+    elif prop['EOS'] == 'user':
+        import jax
+        dp_drho = jax.vmap(jax.grad(prop['EOS_user']))
+        return np.sqrt(np.abs(np.array(dp_drho(density.ravel()).reshape(density.shape))))
 
     # TODO: split EOS and stress arguments already in input
     kwargs = {k: v for k, v in prop.items() if k in args}
