### 2026-04-22 2f1de74
WIP: allow trainable jitter parameter
