### 2026-04-22 3e3bb72
WIP: modify hyperparameter log
