### 2026-04-27 759bc1a
WIP: penalty method implemented
