+++ GaPFlow/fem_2d/conv_slider_pocket_2D/generate_topography.py
@@ -1,0 +1,119 @@
+"""
+Generate topography for the 2D convergent slider with pocket benchmark from:
+  Bertocchi, L. et al. "Fluid film lubrication in the presence of cavitation:
+  a mass-conserving two-dimensional formulation for compressible, piezoviscous
+  and non-Newtonian fluids." Tribology International 67 (2013), 61–71.
+
+Domain layout (Lx = 20 mm, Ly = 10 mm):
+  Linear convergent wedge: h(x) = hmax - (hmax - hmin) * x / Lx
+  Rectangular pocket superimposed at x in [x_p_start, x_p_end],
+                                     y in [y_p_start, y_p_end]:
+    h(x, y) += h_pocket
+
+Geometry parameters:
+  hmin      = 1.0e-6 m   (outlet gap)
+  hmax      = 1.1e-6 m   (inlet gap)
+  h_pock    = 0.4e-6 m   (pocket depth)
+  x_p_start = 4.0e-3 m
+  x_p_end   = 11.0e-3 m  (pocket length = 7 mm)
+  y_p_start = 2.0e-3 m
+  y_p_end   = 8.0e-3 m   (pocket width = 6 mm, centred in y)
+
+Also prints the Bayada EOS cavitation pressure Pcav and the density
+corresponding to p_amb = 1e5 Pa (used as Dirichlet BC at inlet/outlet).
+"""
+
+import os
+import sys
+import numpy as np
+import matplotlib.pyplot as plt
+
+# ── Domain / grid parameters (must match the YAML config) ──────────────────
+Lx = 0.020      # [m]  total domain length
+Ly = 0.010      # [m]  domain width
+Nx = 100
+Ny = 50
+
+dx = Lx / Nx
+dy = Ly / Ny
+
+x = np.linspace(dx / 2, Lx - dx / 2, Nx)   # cell centres in x
+y = np.linspace(dy / 2, Ly - dy / 2, Ny)   # cell centres in y
+
+# ── Geometry parameters ────────────────────────────────────────────────────
+hmin   = 1.0e-6   # [m]  gap at outlet
+hmax   = 1.1e-6   # [m]  gap at inlet
+h_pock = 0.4e-6   # [m]  pocket depth
+
+x_p_start = 4.0e-3    # [m]  pocket start in x
+x_p_end   = 11.0e-3   # [m]  pocket end in x  (length = 7 mm)
+y_p_start = 2.0e-3    # [m]  pocket start in y
+y_p_end   = 8.0e-3    # [m]  pocket end in y  (width = 6 mm)
+
+# ── Build 2-D height field ─────────────────────────────────────────────────
+# Linear convergent wedge (x-dependent only)
+xx, yy = np.meshgrid(x, y, indexing='ij')   # shape (Nx, Ny)
+h2d = hmax - (hmax - hmin) * xx / Lx
+
+# Add pocket in x-y rectangle
+pocket = (
+    (xx >= x_p_start) & (xx < x_p_end) &
+    (yy >= y_p_start) & (yy < y_p_end)
+)
+h2d[pocket] += h_pock
+
+# ── Bayada EOS: compute Pcav and rho_BC for p_amb = 1e5 Pa ────────────────
+rho_l = 850.0
+rho_v = 0.019
+c_l   = 5989.28
+c_v   = 352.0
+p_amb = 1e5   # [Pa]
+
+N = (rho_v * c_v**2 * rho_l * c_l**2 * (rho_v - rho_l)
[…]
