+++ GaPFlow/models/pressure.py
@@ -330,3 +330,129 @@ def bayada_chupin(dens, rho_l, rho_v, c_l, c_v, smooth_width=0.0):
     )
 
     return p
+
+
+# ---------------------------------------------------------------------------
+# Inverse EoS — closed-form inverses for the pressure-based solver.
+#
+# Only 'DH' and 'Bayada' are implemented for now; other EoS variants raise
+# NotImplementedError in the eos_rho / eos_drho_dp dispatchers.
+# ---------------------------------------------------------------------------
+
+
+def rho_of_p_dh(pressure, rho0=877.7007, P0=101325., C1=3.5e8, C2=1.23):
+    """Inverse of Dowson-Higginson EoS.
+
+    Closed-form:
+        rho(p) = rho_0 * [C_2 (p - P_0) + C_1] / [(p - P_0) + C_1]
+
+    Inverts the forward form
+        p(rho) = P_0 + C_1 (rho/rho_0 - 1) / (C_2 - rho/rho_0).
+    """
+    u = pressure - P0
+    return rho0 * (C2 * u + C1) / (u + C1)
+
+
+def rho_of_p_pl(pressure, rho0=1.1853, P0=101325., alpha=0.):
+    """Inverse of power-law EoS.
+
+    Closed-form:
+        rho(p) = rho_0 * (p / P_0)^(1 - alpha/2)
+
+    Inverts the forward form
+        p(rho) = P_0 * (rho/rho_0)^(1 / (1 - alpha/2)).
+
+    For alpha=0 this is a linear EoS; widely used as the canonical test
+    case for FD Jacobian verification.
+    """
+    return rho0 * (pressure / P0) ** (1. - 0.5 * alpha)
+
+
+def rho_of_p_bayada(pressure, rho_l, rho_v, c_l, c_v):
+    """Inverse of Bayada-Chupin EoS (three-region piecewise).
+
+    Liquid  (p >= Pcav):          rho = rho_l + (p - Pcav) / c_l^2
+    Mixture (P_vt <= p <= Pcav):  rho = q*A / (1 - q*B) with
+                                  q = exp((p - Pcav)/N) / (rho_v*c_v^2)
+    Vapor   (p < P_vt):           rho = p / c_v^2
+
+    P_vt = rho_v*c_v^2 is the mixture-vapor boundary pressure; Pcav is the
+    cavitation (mixture-liquid) boundary pressure.
+
+    jnp.clip on the mixture branch is mandatory: exp((p-Pcav)/N) overflows
+    for liquid-region p, and jax.grad through jnp.where is poisoned by NaN
+    intermediates on the inactive branch unless the input is clamped.
+    """
+    N    = (rho_v * c_v**2 * rho_l * c_l**2 * (rho_v - rho_l)
+            / (rho_v**2 * c_v**2 - rho_l**2 * c_l**2))
+    Pcav = rho_v * c_v**2 - N * jnp.log(rho_v**2 * c_v**2 / (rho_l**2 * c_l**2))
+    P_vt = rho_v * c_v**2
+    A    = rho_l * (rho_v**2 * c_v**2 - rho_l**2 * c_l**2) / (rho_v - rho_l)
+    B    = rho_l * (rho_l * c_l**2 - rho_v * c_v**2)       / (rho_v - rho_l)
+
+    p_mix_safe = jnp.clip(pressure, P_vt, Pcav)
+
+    rho_liq = rho_l + (pressure - Pcav) / c_l**2
+    rho_vap = pressure / c_v**2
+    q       = jnp.exp((p_mix_safe - Pcav) / N) / (rho_v * c_v**2)
+    rho_mix = q * A / (1.0 - q * B)
+
+    return jnp.where(pressure >= Pcav, rho_liq,
[…]
