+++ GaPFlow/fem_2d/cavitation_FB.md
@@ -1,0 +1,319 @@
+# Fischer-Burmeister Cavitation — Implementation Plan
+
+## Physical Model
+
+Elrod-Adams: replace flux `j` with `(1−θ)·j` in the mass equation.
+θ is the void fraction (P1 DOF, same grid as p):
+- full film: θ=0, p>p_cav
+- cavitation: θ>0, p=p_cav
+
+`∂p/∂t = dp/drho · ∇·((1−θ)j) − (1/h) · ∇h · (1−θ)j`
+
+No ∂θ/∂t in the mass equation. θ is determined implicitly at each time step
+by the coupled (mass, FB) system. Momentum terms are unchanged.
+
+---
+
+## New DOF: theta (P1)
+
+- `DOF_GRID['theta'] = 'p'` — add to module-level dict in `assembly.py`
+- DOF index for theta is **computed dynamically** inside `Assembly.__init__`
+  from the ordered variables list (slot 2 = p, slot 3 = E if active, then theta).
+  Cannot be a static constant because the slot depends on which combination
+  of P1 DOFs is active.
+- `field_to_global`: replace `bEnergy: bool` with `p_factor: int`; change
+  P1-branch check from `res_type == 2 or res_type == 3` to `res_type >= 2`.
+- `p_factor` computed in `Assembly.__init__` as the count of P1 variables:
+  `p_factor = len([v for v in variables if DOF_GRID[v] == 'p'])`
+- `quad_fields`: add `theta` at quadrature points (no gradient needed — IBP
+  removes derivative from theta in R11x_fb/R11y_fb)
+
+---
+
+## New Residual: fb (P1)
+
+- `DOF_GRID['fb'] = 'p'` — add to module-level dict in `assembly.py`
+- Residual slot index computed dynamically (same logic as theta DOF index)
+- Standard Galerkin, no spatial derivatives
+
+---
+
+## New Mass Terms (Elrod-Adams variants)
+
+The original R11x, R11y, R11Sx, R11Sy terms are kept unchanged.
+New `_fb`-suffixed variants are added and selected instead when `cavitation: true`.
+The original `_corr` terms are also dropped when cavitation is active (dp/drho
+is weakly p-dependent for fluid EoS — correction is negligible).
+`R_cav_pen` is superseded by the FB approach and excluded when `cavitation: true`.
+
+### R11x_fb / R11y_fb — IBP (weak form)
+
+**Motivation**: moving the derivative to the test function via IBP gives a
+complete Jacobian for (mass, theta) with no correction terms, and avoids
+needing `d_dx_theta` in quad_fields.
+
+```python
+R11x_fb = NonLinearTerm(
+    name='R11x_fb', res='mass', dep_vars=['jx', 'theta'], dep_vals=['dp_drho'],
+    fun=lambda ctx: lambda jx, theta:  ctx['dp_drho']() * (1 - theta) * jx,
+    der_funs=[
+        lambda ctx: lambda jx, theta:  ctx['dp_drho']() * (1 - theta),  # ∂/∂jx
+        lambda ctx: lambda jx, theta: -ctx['dp_drho']() * jx,           # ∂/∂theta
+    ],
+    d_dx_resfun=False, d_dy_resfun=False, der_testfun='x')
+```
+
+R11y_fb: identical with `jy` and `der_testfun='y'`.
+
+Jacobian blocks produced (both complete, no correction terms):
+- `(mass, jx,    'none', 'x')`: `∫ ∂Nᵢ/∂x · dp_drho·(1−θ) · Nⱼᴾ² dΩ`
+- `(mass, theta, 'none', 'x')`: `∫ ∂Nᵢ/∂x · (−dp_drho·jx) · Nⱼᴾ¹ dΩ`
+
+### R11Sx_fb / R11Sy_fb — no IBP needed
+
[…]

[…]
