### 2026-04-29 cf68923
MAINT: suppress active learning during rapid changes in the solution, currently fixed at residuals > 1e-3.
