### 2026-04-30 #9645 @matthewfeickert
@vyastreb, @DanWaxman thank you for your reviews!

@hannes-holey As GaPFlow looks like it has passed all the reviewer comments I'll review the paper today for editorial comments and then we can move forward with publication. Apologies for the very delayed response time.

I've left some very minor comments on the paper at https://github.com/hannes-holey/GaPFlow/issues/63.
