### 2026-04-30 #9645 @matthewfeickert
> No DOI given, and none found for title: Gaussian Processes for Machine Learning

This is a book (https://gaussianprocess.org/gpml/) and so a missing DOI is okay.
