### 2026-04-30 hannes-holey/hans#63 @matthewfeickert
[JOSS Review] Paper Editorial Suggestions
These are small editorial suggestions for the paper as part of https://github.com/openjournals/joss-reviews/issues/9645

As of d7ae931ad46b61528f5c03679e74092c38096f02, [`paper/paper.md`](https://github.com/hannes-holey/GaPFlow/blob/d7ae931ad46b61528f5c03679e74092c38096f02/paper/paper.md) there are some pieces of software that are mentioned (and linked to, which is nice) but not cited. These include:

* µGrid and netcdf

https://github.com/hannes-holey/GaPFlow/blob/d7ae931ad46b61528f5c03679e74092c38096f02/paper/paper.md?plain=1#L61

* tinygp

https://github.com/hannes-holey/GaPFlow/blob/d7ae931ad46b61528f5c03679e74092c38096f02/paper/paper.md?plain=1#L73

* JAX

https://github.com/hannes-holey/GaPFlow/blob/d7ae931ad46b61528f5c03679e74092c38096f02/paper/paper.md?plain=1#L75

* ContactMechanics

https://github.com/hannes-holey/GaPFlow/blob/d7ae931ad46b61528f5c03679e74092c38096f02/paper/paper.md?plain=1#L100

Please cite all software referred to in the paper. In the event that the software does not have a clearly communicated preferred citation, please use the version information and the source code repository

Example:

```bibtex
@software{software_name_1_2_3,
author = {Authors},
license = {License info},
title = {{software name: v1.2.3}},
url = {https://github.com/org/repo/releases/tag/v1.2.3},
version = {1.2.3}
}
```

Additionally, while not required for publication, consider running `codespell` over the repository

```
pixi exec codespell .
```

<summary>Relevant results:</summary>

```
./doc/tutorials/01_macro_equations.ipynb:201: occuring ==> occurring
./doc/tutorials/05_lubrication_1d.ipynb:155: als ==> also
./doc/tutorials/05_lubrication_1d.ipynb:168: scetion ==> section
./doc/tutorials/07_gp_mock.ipynb:216: autmatically ==> automatically
./doc/tutorials/07_gp_mock.ipynb:217: wiht ==> with
./GaPFlow/problem.py:231: Intialize ==> Initialize
./GaPFlow/topography.py:255: inital ==> initial
./GaPFlow/viz/plotting.py:221: Presure ==> Pressure, Presume
./GaPFlow/md/base.py:208: traiing ==> trailing, training
./GaPFlow/md/mock.py:36: mimick ==> mimic
./GaPFlow/md/moltemplate.py:181: degreees ==> degrees
./GaPFlow/md/moltemplate.py:577: equil ==> equal
./GaPFlow/md/utils.py:185: Statisitical ==> Statistical
./paper/paper.md:91: re-use ==> reuse
```

to pick up some typos (some will be false positives, but there are some minor typos that can get fixed).
