### 2026-04-30 49d0666
MAINT: check cooldown flag on whole residual buffer
