### 2026-04-30 1b07e4d
WIP: save FB cavitation state
