+++ GaPFlow/fem_2d/assembly.py
@@ -69,7 +69,7 @@ class AssemblyTemplate:
 # Residual, Variable
 BLOCK = (('p', 'p'), ('p', 'v'), ('v', 'p'), ('v', 'v'))
 DOF_GRID = {
-    'jx': 'v', 'jy': 'v', 'p': 'p', 'E': 'p',
+    'jx': 'v', 'jy': 'v', 'p': 'p', 'E': 'p', 'theta': 'p', 'fb': 'p',
     'momentum_x': 'v', 'momentum_y': 'v', 'mass': 'p', 'energy': 'p',
 }
 DOF_IDX = {
@@ -85,11 +85,9 @@ class Assembly:
     ----------
     grid_idx : GridIndexManager
     variables : list of str
-        All variable names in block order, e.g. ['jx', 'jy', 'rho'].
+        All variable names in block order, e.g. ['jx', 'jy', 'p'].
     residuals : list of str
-        All residual names in block order, e.g. ['Rjx', 'Rjy', 'Rrho'].
-    energy : bool
-        Whether energy is active.
+        All residual names in block order, e.g. ['momentum_x', 'momentum_y', 'mass'].
     """
 
     def __init__(self,
@@ -97,14 +95,26 @@ class Assembly:
                  element: TaylorHoodP2P1,
                  variables: List[str],
                  residuals: List[str],
-                 energy: bool = False,
                  ) -> None:
 
         self.grid_idx = grid_idx
         self.variables = variables
         self.residuals = residuals
-        self.energy = energy
         self.element = element
+        self.p_factor = sum(1 for v in variables if DOF_GRID.get(v) == 'p')
+
+        # Build index lookup that extends DOF_IDX with dynamic P1 slots.
+        # Static entries (jx=0, jy=1, p=2, E=3) are correct for their fixed
+        # positions. theta and fb take the next available P1 slots after E.
+        p1_vars = [v for v in variables if DOF_GRID.get(v) == 'p']
+        self._dof_idx = dict(DOF_IDX)
+        for i, v in enumerate(p1_vars):
+            if v not in self._dof_idx:
+                self._dof_idx[v] = 2 + i  # p=2 is slot 0, E=3 slot 1, etc.
+        p1_res = [r for r in residuals if DOF_GRID.get(r) == 'p']
+        for i, r in enumerate(p1_res):
+            if r not in self._dof_idx:
+                self._dof_idx[r] = 2 + i
 
         self.res_to_grid = {r: DOF_GRID[r] for r in residuals}
         self.var_to_grid = {v: DOF_GRID[v] for v in variables}
@@ -260,9 +270,9 @@ class Assembly:
             inner_pts_global_idx = self.apply_l2g(rg, inner_pts)
             contrib_pts_global_idx = self.apply_l2g(vg, contrib_pts)
 
-            res_idx , var_idx = DOF_IDX[res], DOF_IDX[var]
-            global_rows = field_to_global(inner_pts_global_idx, res_idx, self.grid_idx, self.energy)
-            global_cols = field_to_global(contrib_pts_global_idx, var_idx, self.grid_idx, self.energy)
+            res_idx, var_idx = self._dof_idx[res], self._dof_idx[var]
+            global_rows = field_to_global(inner_pts_global_idx, res_idx, self.grid_idx, self.p_factor)
+            global_cols = field_to_global(contrib_pts_global_idx, var_idx, self.grid_idx, self.p_factor)
             self.nnz_global_rows = np.concatenate([self.nnz_global_rows, global_rows])
             self.nnz_global_cols = np.concatenate([self.nnz_global_cols, global_cols])
 
@@ -273,8 +283,7 @@ class Assembly:
         decomp = self.grid_idx._decomp
[…]
