+++ paper/paper.bib
@@ -389,3 +389,43 @@
   doi = {10.1103/PhysRevLett.87.096105},
   urldate = {2022-08-25}
 }
+
+@software{mugrid_0_105_0,
+author = {Junge, Till and Pastewka, Lars},
+license = {LGPL-3.0 License},
+title = {{muGrid: v0.105.0}},
+url = {https://github.com/muSpectre/muGrid/releases/tag/v0.105.0},
+version = {0.105.0}
+}
+
+@software{netcdf_4_9_3,
+author = {Unidata},
+title = {{Network Common Data Form (netCDF)}},
+url = {http://doi.org/10.5065/D6H70CW6},
+version = {4.9.3}
+}
+
+@software{tinygp_0_3_0,
+author = {Foreman-Mackey, Dan},
+license = {MIT License},
+title = {{tinygp: v0.3.0}},
+url = {https://github.com/dfm/tinygp/releases/tag/v0.3.0},
+version = {0.3.0}
+}
+
+@software{jax2018github,
+author = {Bradbury, James and Frostig, Roy and Hawkins, Peter and Johnson, Matthew James and Katariya, Yash and Leary, Chris and Maclaurin, Dougal and Necula, George and Paszke, Adam and Vander{P}las, Jake and Wanderman-{M}ilne, Skye and Zhang, Qiao},
+title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+url = {http://github.com/jax-ml/jax},
+license = {Apache-2.0 License},
+version = {0.9.0},
+year = {2018},
+}
+
+@software{contactmechanics_1_8_0,
+author = {Pastewka, Lars and R{\"o}ttger, Michael C. and Junge, Till and Sanner, Antoine},
+license = {MIT License},
+title = {{ContactMechanics: v1.8.0}},
+url = {https://github.com/ContactEngineering/ContactMechanics/releases/tag/v1.8.0},
+version = {1.8.0}
+}
