+++ paper/paper.md
@@ -58,7 +58,7 @@ $$
 where $\bar{\mathbf{q}}\equiv\bar{\mathbf{q}}(x,y,t)=h^{-1}\int_{h_0}^{h_1}\mathbf{q}(x,y,z,t)dz$ collects the densities of conserved variables (e.g. $\mathbf{q}=(\rho, j_x, j_y)^\top$ for mass and in-plane momentum) and $\bar{\mathbf{f}}_i\equiv\bar{\mathbf{f}}_i(x,y,t)=h^{-1}\int_{h_0}^{h_1}\mathbf{f}_i(x,y,z,t)dz$ are the corresponding fluxes in direction $i\in\{x,y\}$ with $h=h_1 - h_0$. 
 The source term $\mathbf{s}$ accounts for fluxes across the bottom and top walls ($\mathbf{f}_z$) as well as for changes in the conserved variable densities induced by flow within a spatially varying gap.
 The current implementation uses a finite volume discretization on a regular grid and the MacCormack explicit time-integration scheme [@maccormack2003_effect] to solve the transient lubrication problem. 
-The [µGrid](https://muspectre.github.io/muGrid/) library is used to assemble the discretized density and flux fields into a unified container and to export the simulation results in the [NetCDF](https://www.unidata.ucar.edu/software/netcdf) file format.
+The [µGrid](https://muspectre.github.io/muGrid/)[@mugrid_0_105_0] library is used to assemble the discretized density and flux fields into a unified container and to export the simulation results in the [NetCDF](https://www.unidata.ucar.edu/software/netcdf)[@netcdf_4_9_3] file format.
 
 Next to the numerical integration of the continuum equations, `GaPFlow` serves as a *glue code* that integrates the various components for multiscale or multiphysics simulations. 
 Therefore, it relies on a small set of external dependencies which are summarized below.
@@ -70,9 +70,9 @@ The GP models are trained on data generated by MD, or, for testing purposes, on 
 These include non-Newtonian fluid models capturing shear-thinning and piezoviscous effects, as well as equations of state for the fluid pressure.
 To account for cavitation, `GaPFlow` implements a homogenized gas–fluid mixture model following @bayada2013_compressible, enabling the description of compressible two-phase behavior within the lubrication framework.
 
-`GaPFlow` employs the [tinygp](https://tinygp.readthedocs.io/en/stable/index.html) library for constructing and training GP models, taking advantage of its flexibility.
+`GaPFlow` employs the [tinygp](https://tinygp.readthedocs.io/en/stable/index.html) [@tinygp_0_3_0] library for constructing and training GP models, taking advantage of its flexibility.
 For example, it allows the implementation of custom kernels for the joint prediction of wall shear stresses at the top and bottom walls using a multi-output GP that shares a common noise process.
-Since [tinygp](https://tinygp.readthedocs.io/en/stable/index.html) is built on [JAX](https://github.com/jax-ml/jax), `GaPFlow` also benefits from automatic differentiation of the GP models, e.g. to compute the speed of sound from the pressure model.
+Since [tinygp](https://tinygp.readthedocs.io/en/stable/index.html) is built on [JAX](https://github.com/jax-ml/jax) [@jax2018github], `GaPFlow` also benefits from automatic differentiation of the GP models, e.g. to compute the speed of sound from the pressure model.
 
 ## Automatic setup of MD runs
 
@@ -97,7 +97,7 @@ This makes it straightforward to discover previously computed configurations, or
 ## Elastic deformations
 
 In non-conforming lubricated contacts such as in ball or roller bearings, local fluid pressures can become large such that elastic deformation of the walls can no longer be neglected.
[…]
