### 2026-04-30 7502a96
JOSS: switch to latest versions, add release dates, replace NetCDF reference
