+++ paper/paper.bib
@@ -390,42 +390,52 @@
   urldate = {2022-08-25}
 }
 
-@software{mugrid_0_105_0,
+@software{mugrid_0_105_2,
 author = {Junge, Till and Pastewka, Lars},
 license = {LGPL-3.0 License},
-title = {{muGrid: v0.105.0}},
-url = {https://github.com/muSpectre/muGrid/releases/tag/v0.105.0},
-version = {0.105.0}
+title = {{muGrid: v0.105.2}},
+url = {https://github.com/muSpectre/muGrid/releases/tag/v0.105.2},
+version = {0.105.2},
+year = {2026}
 }
 
-@software{netcdf_4_9_3,
-author = {Unidata},
-title = {{Network Common Data Form (netCDF)}},
-url = {http://doi.org/10.5065/D6H70CW6},
-version = {4.9.3}
+@article{rew1990_netcdf,
+title = {{{NetCDF}}: An Interface for Scientific Data Access},
+shorttitle = {{{NetCDF}}},
+author = {Rew, R. and Davis, G.},
+year = 1990,
+month = jul,
+journal = {IEEE Computer Graphics and Applications},
+volume = {10},
+number = {4},
+pages = {76--82},
+issn = {1558-1756},
+doi = {10.1109/38.56302},
 }
 
-@software{tinygp_0_3_0,
+@software{tinygp_0_3_1,
 author = {Foreman-Mackey, Dan},
 license = {MIT License},
-title = {{tinygp: v0.3.0}},
-url = {https://github.com/dfm/tinygp/releases/tag/v0.3.0},
-version = {0.3.0}
+title = {{tinygp: v0.3.1}},
+url = {https://github.com/dfm/tinygp/releases/tag/v0.3.1},
+version = {0.3.1},
+year = {2026}
 }
 
-@software{jax2018github,
+@software{jax_0_10_0,
 author = {Bradbury, James and Frostig, Roy and Hawkins, Peter and Johnson, Matthew James and Katariya, Yash and Leary, Chris and Maclaurin, Dougal and Necula, George and Paszke, Adam and Vander{P}las, Jake and Wanderman-{M}ilne, Skye and Zhang, Qiao},
 title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
 url = {http://github.com/jax-ml/jax},
 license = {Apache-2.0 License},
-version = {0.9.0},
-year = {2018},
+version = {0.10.0},
+year = {2026},
 }
 
-@software{contactmechanics_1_8_0,
+@software{contactmechanics_1_8_1,
 author = {Pastewka, Lars and R{\"o}ttger, Michael C. and Junge, Till and Sanner, Antoine},
 license = {MIT License},
-title = {{ContactMechanics: v1.8.0}},
-url = {https://github.com/ContactEngineering/ContactMechanics/releases/tag/v1.8.0},
-version = {1.8.0}
+title = {{ContactMechanics: v1.8.1}},
+url = {https://github.com/ContactEngineering/ContactMechanics/releases/tag/v1.8.1},
+version = {1.8.1},
+year = {2026}
 }
