### 2026-04-30 52fe928
MAINT: increase size of residual buffer, enable partial check of buffer
