### 2026-04-30 a51d836
MAINT: add config parameter for acquisition break during cooldown
