### 2026-04-30 hannes-holey/hans#63 @matthewfeickert
Thanks! I'll rerender on https://github.com/openjournals/joss-reviews/issues/9645 and close this out if all looks good. 👍
