### 2026-04-30 hannes-holey/hans#63 @matthewfeickert
Small revisions in https://github.com/hannes-holey/GaPFlow/pull/65#pullrequestreview-4208142672
