### 2026-04-30 4e0f129
JOSS: fix JAX and tinygp bibtex
