### 2026-05-04 33a4ce6
JOSS: fix software urls
