+++ paper/paper.bib
@@ -394,7 +394,7 @@
 author = {Junge, Till and Pastewka, Lars},
 license = {LGPL-3.0 License},
 title = {{muGrid: v0.105.2}},
-url = {https://github.com/muSpectre/muGrid/releases/tag/v0.105.2},
+url = {https://github.com/muSpectre/muGrid/releases/tag/0.105.2},
 version = {0.105.2},
 year = {2026}
 }
@@ -446,7 +446,7 @@ note = {https://github.com/dfm/tinygp/releases/tag/v0.3.1}
 author = {Pastewka, Lars and R{\"o}ttger, Michael C. and Junge, Till and Sanner, Antoine},
 license = {MIT License},
 title = {{ContactMechanics: v1.8.1}},
-url = {https://github.com/ContactEngineering/ContactMechanics/releases/tag/v1.8.1},
+url = {https://github.com/ContactEngineering/ContactMechanics/releases/tag/1.8.1},
 version = {1.8.1},
 year = {2026}
 }
