### 2026-05-04 hannes-holey/hans#63 @matthewfeickert
Resolved by PR https://github.com/hannes-holey/GaPFlow/pull/67. 👍
