### 2026-05-04 #9645 @matthewfeickert
@hannes-holey All that's left to do now before recommending publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the last code release was [GaPFlow `v1.0.2`](https://github.com/hannes-holey/GaPFlow/releases/tag/v1.0.2), as there has been some development since that release the [current `HEAD`](https://github.com/hannes-holey/GaPFlow/commit/ca2c7a2ed8b134b7f848f949202bb8ab6b679dd4) has some [differences from `v1.0.2`](https://github.com/hannes-holey/GaPFlow/compare/v1.0.2...ca2c7a2ed8b134b7f848f949202bb8ab6b679dd4). Can you make a release that corresponds to current state of code reviewed? Read on RE: archival before doing so though:

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Once you make a release of GaPFlow with a DOI for the archive, and have verified all related author tasks listed in https://github.com/openjournals/joss-reviews/issues/9645#issuecomment-4350104307, please just report them here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
