### 2026-05-04 #9645 @hannes-holey
@matthewfeickert I'd like to include hannes-holey/GaPFlow#64 into the new release. These are mostly small changes collected during but unrelated to the JOSS review process. Is it OK to do that right now, or should I wait for a later release?
