### 2026-05-04 #9645 @matthewfeickert
> I'd like to include [hannes-holey/GaPFlow#64](https://github.com/hannes-holey/GaPFlow/pull/64) into the new release.

@hannes-holey From quickly scanning the PR I think that's fine to do. 👍
