### 2026-05-04 #9645 @hannes-holey
I created a new [release v1.1.0](https://github.com/hannes-holey/GaPFlow/releases/tag/v1.1.0).
The corresponding Zenodo archive is https://doi.org/10.5281/zenodo.20021529
