### 2026-05-06 bd551ff
MAINT: apply abs() before floor() --> same behavior for positive and negative rotations
