### 2026-05-07 b86ee09
WIP: working version for FB cavitation, all examples
