+++ GaPFlow/fem_2d/conv_slider_pocket_2D_FB/run.py
@@ -1,5 +1,5 @@
 """
-Twin identical parabolic slider — FB cavitation (DH EOS).
+2D Convergent slider with pocket — FB cavitation (DH EOS).
 
 Run generate_topography.py first:
     python generate_topography.py
@@ -14,7 +14,7 @@ from plot_newton_residual import plot_newton_residual
 
 os.chdir(os.path.dirname(os.path.abspath(__file__)))
 
-problem = GaPFlow.Problem.from_yaml('twin_parabolic_slider_id_FB.yaml')
+problem = GaPFlow.Problem.from_yaml('conv_slider_pocket_2D_FB.yaml')
 problem.run()
 
 plot_newton_residual(problem.solver, problem.outdir)

[…]
