### 2026-05-07 de2f55d
WIP: test uncertainty tolerance protocols
