+++ GaPFlow/models/gp.py
@@ -26,7 +26,7 @@ import abc
 import warnings
 from copy import deepcopy
 from datetime import datetime
-from typing import Tuple
+from typing import Tuple, Deque
 
 import jax
 import jax.numpy as jnp
@@ -40,6 +40,7 @@ with warnings.catch_warnings():
 from tinygp import GaussianProcess, kernels, transforms
 
 from ..logging import get_logger
+from ..utils import above_tolerance
 
 logger = get_logger("gapflow.gp")
 
@@ -64,9 +65,9 @@ class GaussianProcessSurrogate:
     active_dims: list[int]
     use_active_learning: bool
     fix_noise: bool
+    tolerance_protocol: str
     rtol: float
     atol: float
-    tol: str
     max_steps: int
     pause_steps: int
     similarity_check: bool
@@ -282,8 +283,8 @@ class GaussianProcessSurrogate:
             self.history['variance'].append(self.kernel_variance)
             self.history['obs_stddev'].append(self.obs_stddev)
             self.history['maximum_variance'].append(self.maximum_variance)
-            self.history['variance_tol'].append(self.variance_tol)
             self.history['objective'].append(self.objective)
+            self.history['variance_tol'].append(self.variance_tol)
 
             for i, l in enumerate(self.active_dims):
                 self.history[f'lengthscale_{l}'].append(self.kernel_lengthscale[i])
@@ -448,7 +449,9 @@ class GaussianProcessSurrogate:
         if compute_var:
             predictive_mean, self._predictive_var = self._infer_mean_var()
             self.maximum_variance = jnp.max(self._predictive_var)
-            self.variance_tol = self._get_tolerance(predictive_mean)
+
+            # Dummy value overwritten by active learning
+            self.variance_tol = self.maximum_variance
         else:
             predictive_mean = self._infer_mean()
 
@@ -457,12 +460,12 @@ class GaussianProcessSurrogate:
     # ------------------------------------------------------------------
     # Active Learning
     # ------------------------------------------------------------------
-    def _get_tolerance(self, Y):
+    def _get_tolerance(self, m, residual):
         """Compute the variance tolerance based on the current prediction.
 
         Parameters
         ----------
-        Y : jax.Array
+        m : jax.Array
             Predictive mean
 
         Returns
@@ -472,21 +475,23 @@ class GaussianProcessSurrogate:
         """
 
         noise = self.Yerr * self.Yscale
+        atol = self.atol * noise  # "lower bound", multiple of observation noise
 
-        if self.tol == 'delta':
-            Ys = jnp.max(Y) - jnp.min(Y)
-        elif self.tol == 'absmax':
-            Ys = jnp.max(jnp.abs(Y))
-        elif self.tol == 'snr':
-            Ys = jnp.mean(Y) / noise
+        if self.tolerance_protocol == 'atol_rtol_delta':
+            delta = jnp.max(m) - jnp.min(m)
+            rtol = self.rtol * delta
+            std_tol = jnp.maximum(atol, rtol)
+
+        elif self.tolerance_protocol == 'atol_sigmoid':
+            def tolerance(r, rmid=1e-6, t0=10., t1=5., alpha=2.):
+                return -(t1 - t0) / (1. + np.exp(-alpha * (np.log(r) - np.log(rmid)))) + t1
+
+            atol_init = atol
+            atol_final = 0.5 * atol
+            std_tol = tolerance(residual, t0=atol_init, t1=atol_final)
+
         else:
             raise RuntimeError('No tolerance calculation configured.')
-
-        atol = self.atol * noise  # "lower bound", multiple of observation noise
[…]
