### 2026-05-07 e1be574
MAINT: jax numpy
