### 2026-05-15 971f824
ENH: adjustable range for initial database (not only density)
