+++ GaPFlow/io.py
@@ -442,12 +442,13 @@ def sanitize_db(d):
     out['dtool_path'] = d.get('dtool_path', None)
     out['init_size'] = int(d.get('init_size', 5))
     out['init_method'] = str(d.get('init_method', 'lhc'))
-    out['init_width'] = float(d.get('init_width', 1e-2))
-    out['init_seed'] = int(d.get('init_width', 123))
+    out['init_halfwidth'] = list(d.get('init_halfwidth', [1e-2, 0.5, 0.5]))
+    out['init_seed'] = int(d.get('init_seed', 123))
 
     out['normalizer_X'] = d.get('normalizer_X', 'minmax')
     out['normalizer_Y'] = d.get('normalizer_Y', 'standard')
 
+    assert len(out['init_halfwidth']) == 3
     assert out['init_method'] in ['rand', 'lhc', 'sobol']
     assert out['normalizer_X'] in ['max', 'minmax', 'standard', 'none']
     assert out['normalizer_Y'] in ['max', 'minmax', 'standard', 'none']
