+++ examples/config/journal_1d_dowson-higginson_gp.yaml
@@ -50,6 +50,6 @@ db:
     # dtool_path: ... # defaults to options['output']/train   
     init_size: 5
     init_method: lhc
-    init_width: 1.e-6
+    init_halfwidth: [1.e-6, .5, .5]  # training data very close to rho0
     normalizer_X: standard
     normalizer_Y: standard
