### 2026-05-15 bf6cf6e
MAINT: fixed BC key names (W/E, S/N)
