### 2026-05-15 5b2c80d
MAINT: fix logic in topography generation
