### 2026-05-15 2ea12ab
ENH: add 'linear' tolerance protocol and configurable atol_reduction_factor, tol_rmid, tol_alpha
