+++ GaPFlow/models/stress.py
@@ -101,6 +101,9 @@ class WallStress(GaussianProcessSurrogate):
             self.tolerance_protocol = gp['tolerance_protocol']
             self.atol = gp['atol']
             self.rtol = gp['rtol']
+            self.atol_reduction_factor = gp['atol_reduction_factor']
+            self.tol_rmid = gp['tol_rmid']
+            self.tol_alpha = gp['tol_alpha']
             self.max_steps = gp['max_steps']
             self.pause_steps = gp['pause_steps']
             self.use_active_learning = gp['active_learning']
@@ -528,6 +531,9 @@ class Pressure(GaussianProcessSurrogate):
             self.tolerance_protocol = gp['tolerance_protocol']
             self.atol = gp['atol']
             self.rtol = gp['rtol']
+            self.atol_reduction_factor = gp['atol_reduction_factor']
+            self.tol_rmid = gp['tol_rmid']
+            self.tol_alpha = gp['tol_alpha']
             self.max_steps = gp['max_steps']
             self.pause_steps = gp['pause_steps']
             self.use_active_learning = gp['active_learning']
