+++ tests/test_solver_comparison.py
@@ -1,238 +1,0 @@
-"""
-Cross-solver comparison tests.
-
-Compares solutions from different solvers (explicit, fem_1d, fem_2d) on the same
-physical problem to verify consistency.
-
-Problem config files are in /tests/configs/.
-
-Test Cases
-----------
-1. Inclined Slider Bearing (PL EOS, air)
-   - Dirichlet density BC in X, periodic in Y
-   - Linear converging gap
-
-2. Journal Bearing (DH EOS, oil)
-   - Periodic BC in X and Y
-   - Cosine gap profile (journal geometry)
-
-3. Parabolic Slider Bearing (DH EOS, oil)
-   - Dirichlet density BC in X, periodic in Y
-   - Parabolic gap profile with converging/diverging sections
-
-Notes
------
-- Solutions are compared using [:, 0] slice to match 1D/2D array shapes
-- Dirichlet BC cases use inner-region comparison to avoid boundary artifacts
-"""
-import pytest
-import numpy as np
-from pathlib import Path
-
-from GaPFlow.problem import Problem
-
-
-# Path to config templates
-CONFIG_DIR = Path(__file__).parent / "configs"
-
-# =============================================================================
-# Solver Defaults (use-case independent)
-# =============================================================================
-
-SOLVER_DEFAULTS = {
-    'explicit': {
-        'Ny': 1,
-        'solver': 'explicit',
-        'max_it': 50000,
-    },
-    'fem_1d': {
-        'Ny': 1,
-        'solver': 'fem',
-        'dt': 0.1,
-        'max_it': 100,
-    },
-    'fem_2d': {
-        'Ny': 4,
-        'solver': 'fem',
-        'dt': 0.1,
-        'max_it': 100,
-        'pressure_stab_alpha': 1000,
-        'momentum_stab_alpha': 10000,
-    },
-}
-
-# =============================================================================
-# Use-Case Configurations
-# =============================================================================
-
-USE_CASES = {
-    'inclined_slider': {
-        'template': 'inclined_slider',
-        'comparison': 'dirichlet',
-        'explicit': {'dt': 1e-6},
-    },
-    'journal_bearing': {
-        'template': 'journal_bearing',
-        'comparison': 'periodic',
-        'explicit': {'dt': 1e-10, 'adaptive': 1, 'CFL': 0.1},
-    },
-    'parabolic_slider': {
-        'template': 'parabolic_slider',
-        'comparison': 'dirichlet',
-        'rho_bc': 850.0,
-        'explicit': {'dt': 1e-10, 'adaptive': 1, 'CFL': 0.45},
-    },
-}
-
-
-# =============================================================================
-# Helper Functions
-# =============================================================================
-
-def load_template(name: str) -> str:
-    """Load a YAML template from the configs directory."""
-    path = CONFIG_DIR / f"{name}.yaml"
-    with open(path, 'r') as f:
-        return f.read()
-
-
-def build_config(use_case: str, solver: str) -> str:
-    """Build YAML config string by merging solver defaults with use-case overrides."""
-    case = USE_CASES[use_case]
-    template = load_template(case['template'])
-
-    # Start with template defaults
-    params = {
-        'pressure_stab_alpha': 0,
-        'momentum_stab_alpha': 0,
-        'adaptive': 0,
-        'CFL': 0.5,
-        'rho_bc': case.get('rho_bc', 1.0),
-    }
-
-    # Apply solver defaults
-    params.update(SOLVER_DEFAULTS[solver])
-
-    # Apply use-case overrides for this solver
-    params.update(case.get(solver, {}))
-
-    return template.format(**params)
-
-
-def extract_solution(problem) -> tuple[np.ndarray, np.ndarray]:
-    """Extract rho and jx arrays from problem.
-
[…]
