### 2026-05-29 f8efd08
WIP: cleaned assembly, cavitation cases notebook
