+++ GaPFlow/fem_2d_old/grid_index.py
@@ -1,238 +1,0 @@
-#
-# Copyright 2025 Christoph Huber
-#
-# ### MIT License
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-#
-"""Grid index management for FEM assembly.
-
-Handles index masks and boundary condition handling for triangular FEM
-on structured grids with domain decomposition support.
-"""
-from functools import cached_property, lru_cache
-from typing import List, TYPE_CHECKING
-
-import numpy as np
-import numpy.typing as npt
-
-if TYPE_CHECKING:
-    from ..parallel import DomainDecomposition
-
-NDArray = npt.NDArray[np.floating]
-IntArray = npt.NDArray[np.signedinteger]
-
-
-class GridIndexManager:
-    """Manages grid index masks and element connectivity for FEM assembly.
-
-    Handles:
-    - Index masks for inner (residual) and padded (contributor) grids
-    - Square element coordinates and corner connectivity
-    - Local-to-global point mapping
-    - Boundary condition handling (Dirichlet removal, Neumann forwarding)
-    - Does NOT know about triangular elements and quadrature
-    """
-
-    def __init__(self,
-                 decomp: "DomainDecomposition",
-                 variables: List[str],
-                 energy_spec: dict = None):
-        self._decomp = decomp
-        self._variables = variables
-        grid = decomp.grid
-
-        # Grid dimensions from decomp
-        Nx_inner, Ny_inner = decomp.nb_subdomain_grid_pts
-        self.Nx_inner = Nx_inner
-        self.Ny_inner = Ny_inner
-        self.Nx_padded = Nx_inner + 2
-        self.Ny_padded = Ny_inner + 2
-
-        # Determine if subdomain is at domain boundary AND not periodic
-        self.bc_at_W = decomp.is_at_xW and not decomp.periodic_x
-        self.bc_at_E = decomp.is_at_xE and not decomp.periodic_x
-        self.bc_at_S = decomp.is_at_yS and not decomp.periodic_y
-        self.bc_at_N = decomp.is_at_yN and not decomp.periodic_y
-
-        # Build Neumann BC flags from grid config (bool list for each boundary)
-        self._bc_neumann = {
-            'xW': [b == 'N' for b in grid['bc_xW']],
-            'xE': [b == 'N' for b in grid['bc_xE']],
-            'yS': [b == 'N' for b in grid['bc_yS']],
-            'yN': [b == 'N' for b in grid['bc_yN']],
-        }
-        # Append energy BC flags if energy variable present
-        if 'E' in variables:
[…]
