### 2026-06-02 f8c4bbe
WIP: before BC system change
