+++ GaPFlow/fem_2d/tests/interactive_residual.py
@@ -1,215 +1,0 @@
-"""Interactive residual inspector.
-
-Set individual jx nodal values and see the resulting residual.
-
-Usage:
-    python GaPFlow/fem_2d/tests/interactive_residual.py
-
-At the prompt, type:
-    ix,iy,val   — set jx at inner node (ix,iy) to val
-    r           — reset all jx to zero
-    q           — quit
-"""
-
-import numpy as np
-from GaPFlow.problem import Problem
-
-# =============================================================================
-# Config
-# =============================================================================
-
-Nx        = 3
-Ny        = 3
-Lx        = 10
-Ly        = 10
-term_list = ['R23xx']
-bc        = 'periodic_y'
-
-BC_CONFIGS = {
-    'dirichlet': {
-        'xE': "['D', 'N', 'N']", 'xW': "['D', 'N', 'N']",
-        'yS': "['D', 'N', 'N']", 'yN': "['D', 'N', 'N']",
-    },
-    'periodic': {
-        'xE': "['P', 'P', 'P']", 'xW': "['P', 'P', 'P']",
-        'yS': "['P', 'P', 'P']", 'yN': "['P', 'P', 'P']",
-    },
-    'periodic_y': {
-        'xE': "['D', 'N', 'N']", 'xW': "['D', 'N', 'N']",
-        'yS': "['P', 'P', 'P']", 'yN': "['P', 'P', 'P']",
-    },
-}
-
-_CONFIG_TEMPLATE = """
-options:
-    output: /tmp/interactive_res
-    write_freq: 1000
-    silent: True
-grid:
-    Lx: {Lx}
-    Ly: {Ly}
-    Nx: {Nx}
-    Ny: {Ny}
-    xE: {xE}
-    xW: {xW}
-    yS: {yS}
-    yN: {yN}
-    xE_D: 1.1
-    xW_D: 1.0
-    yS_D: 1.05
-    yN_D: 1.05
-geometry:
-    type: inclined
-    hmax: 1e-5
-    hmin: 1e-5
-    U: 0.0
-    V: 0.0
-numerics:
-    solver: fem
-    dt: 1e-3
-    tol: 1e-6
-    max_it: 100
-properties:
-    EOS: PL
-    rho0: 1.0
-    shear: 10
-    bulk: 0.
-    P0: 101325
-    alpha: 0.
-fem_solver:
-    type: newton_alpha
-    equations:
-        energy: False
-        {term_list_entry}
-"""
-
-# =============================================================================
-# Setup
-# =============================================================================
-
-bc_cfg = BC_CONFIGS[bc]
-term_entry = f'term_list: {term_list}' if term_list else ''
-config = _CONFIG_TEMPLATE.format(
-    Nx=Nx, Ny=Ny, Lx=Lx, Ly=Ly,
-    xE=bc_cfg['xE'], xW=bc_cfg['xW'],
-    yS=bc_cfg['yS'], yN=bc_cfg['yN'],
-    term_list_entry=term_entry,
-)
-
-problem = Problem.from_string(config)
-solver = problem.solver
-solver.pre_run()
-
-gi = solver.grid_idx
-Nx_v = gi.Nx_v_inner
-Ny_v = gi.Ny_v_inner
-Nx_v_pad = gi.Nx_v_padded
-Ny_v_pad = gi.Ny_v_padded
-ghost = 2
-
-jx_slice = solver._sol_slices['jx']
-q_base = solver.get_q_nodal().copy()
-
-print(f'\nGrid: {Nx}x{Ny}, Lx={Lx}, Ly={Ly}, bc={bc}, terms={term_list}')
-print(f'P2 inner: {Nx_v}x{Ny_v}, P2 padded: {Nx_v_pad}x{Ny_v_pad}')
[…]
