+++ GaPFlow/fem_2d/tests/test_connectivity.py
@@ -1,134 +1,0 @@
-"""Standalone test for _compute_stencils + apply_stencil."""
-import numpy as np
-from GaPFlow.fem_2d.elements import TaylorHoodP2P1
-from GaPFlow.fem_2d.grid_index import GridIndexManager
-
-BLOCK = (('p', 'p'), ('p', 'v'), ('v', 'p'), ('v', 'v'))
-
-
-class StubDecomp:
-    def __init__(self, Nx_p, Ny_p):
-        Nx_v, Ny_v = 2*Nx_p-1, 2*Ny_p-1
-        self.periodic_x = False; self.periodic_y = False
-        self.is_at_xW = True; self.is_at_xE = True
-        self.is_at_yS = True; self.is_at_yN = True
-        nb_vars = 3
-        self.grid = {'bc_xW': ['D']*nb_vars, 'bc_xE': ['D']*nb_vars,
-                     'bc_yS': ['D']*nb_vars, 'bc_yN': ['D']*nb_vars}
-        self.nb_subdomain_grid_pts = (Nx_p, Ny_p)
-        self.index_mask_padded_global = np.full((Nx_p+2, Ny_p+2), -1, dtype=np.int32)
-        self.index_mask_padded_global[1:-1, 1:-1] = np.arange(Nx_p*Ny_p).reshape(Nx_p, Ny_p, order='F')
-        self.nb_subdomain_grid_pts_v = (Nx_v, Ny_v)
-        self.index_mask_padded_global_v = np.full((Nx_v+4, Ny_v+4), -1, dtype=np.int32)
-        self.index_mask_padded_global_v[2:-2, 2:-2] = np.arange(Nx_v*Ny_v).reshape(Nx_v, Ny_v, order='F')
-
-    @property
-    def bc_at_W(self): return self.is_at_xW and not self.periodic_x
-    @property
-    def bc_at_E(self): return self.is_at_xE and not self.periodic_x
-    @property
-    def bc_at_S(self): return self.is_at_yS and not self.periodic_y
-    @property
-    def bc_at_N(self): return self.is_at_yN and not self.periodic_y
-
-
-def _compute_stencils(element):
-    stencil = {}
-    stencil[3] = {
-        (0, 0): np.array(element.stencil_even_even, dtype=np.int32),
-        (1, 1): np.array(element.stencil_odd_odd,   dtype=np.int32),
-        (0, 1): np.array(element.stencil_even_odd,  dtype=np.int32),
-        (1, 0): np.array(element.stencil_odd_even,  dtype=np.int32),
-    }
-    for idx, combination in enumerate(BLOCK[0:3]):
-        stencil[idx] = {}
-        res, var = combination[0], combination[1]  # BLOCK = (res, var)
-        for origin in stencil[3].keys():
-            if res == 'p' and origin != (0, 0):
-                continue
-            if var == 'v':
-                stencil[idx][origin] = stencil[3][origin]
-                continue
-            points = stencil[3][origin]
-            stencil[idx][origin] = np.empty((0, 2), dtype=np.int32)
-            for point in points:
-                x, y = origin[0] + point[0], origin[1] + point[1]
-                if not (x % 2 == 0 and y % 2 == 0):
-                    continue
-                stencil[idx][origin] = np.vstack([stencil[idx][origin], point])
-    return stencil
-
-
-def apply_stencil(stencil_block, block_type, m_inner_v, m_padded_v, m_padded_p):
-    res_grid, var_grid = block_type
-    inner_pts_2d = np.argwhere(m_inner_v >= 0)
-    inner_idx = m_inner_v[inner_pts_2d[:, 0], inner_pts_2d[:, 1]]
-    il, cl = [], []
-    for origin, offsets in stencil_block.items():
[…]

[…]
