### 2026-06-03 4c57d4f
WIP: BC flexible system
