+++ GaPFlow/parallel.py
@@ -1,6 +1,5 @@
 #
-# Copyright 2025 Hannes Holey
-#           2025 Christoph Huber
+# Copyright 2026 Christoph Huber
 #
 # ### MIT License
 #
@@ -33,15 +32,14 @@ from scipy.ndimage import zoom
 from dataclasses import dataclass
 from functools import cached_property
 
-from typing import TYPE_CHECKING, Callable, List, Tuple
+from typing import TYPE_CHECKING, Tuple
 if TYPE_CHECKING:
     from .problem import Problem
 
 from muGrid import (
     CartesianDecomposition,
     GlobalFieldCollection,
-    Communicator,
-    Field,
+    Communicator
 )
 try:
     from muGrid import FFTEngine
@@ -51,8 +49,6 @@ except ImportError:
     HAS_FFT_ENGINE = False
 
 NDArray = npt.NDArray[np.floating]
-
-BND_IDX = {'N': 0, 'E': 1, 'S': 2, 'W': 3}
 
 @dataclass
 class BCContext:
@@ -834,258 +830,3 @@ class FFTDomainTranslation:
         MPI.Request.Waitall(recv_reqs)
         for buf, y_slice in recv_buffers:
             dst[:, y_slice] = buf
-
-
-# =============================================================================
-# WIP: generalized ghost update (placeholder, not yet wired up)
-# =============================================================================
-
-class BCContext_:
-    """Mutable context object passed to BC functions. Updated in-place per boundary.
-
-    Attributes
-    ----------
-    problem : Problem
-        The GaPFlow Problem instance.
-    required_shape : tuple
-        Shape of the P1 ghost layer at the current boundary.
-    slice_ghost : tuple
-        Slice for the P1 ghost layer (offset 0) at the current boundary.
-    slice_interior : tuple
-        Slice for the P1 interior layer (offset 1) at the current boundary.
-    x_norm : NDArray
-        Normalized x-coordinates at the ghost layer (0 to 1).
-    y_norm : NDArray
-        Normalized y-coordinates at the ghost layer (0 to 1).
-    """
-
-    def __init__(self, problem: "Problem"):
-        self.problem = problem
-        self.required_shape = None
-        self.slice_ghost = None
-        self.slice_interior = None
-        self.x_norm = None
-        self.y_norm = None
-
-class BoundarySpec:
-    """Structured BC specification for a single solution field.
-
-    Attributes
-    ----------
-    field : Field
-        The field this BC spec applies to.
-    grid_type : str
-        'P1' or 'P2', determines ghost depth and interpolation needs.
-    bc_type : List[str]
-        List of BC types for each component ['D', 'N', 'F'] .
-    bc_val : List[float | None]
-        List of BC values for each component if not a function, else None.
-    bc_function : List[Callable | None]
-        List of BC functions for each component if bc_type is 'F', else None.
-    decomp : DomainDecomposition
-        Used for P1-P2 interpolation factors.
-    """
-
-    def __init__(self, 
-                 field: Field,
-                 grid_type: str,
-                 bc_type: List[str],
-                 bc_vals: List[float | None],
-                 bc_functions: List[Callable | None],
-                 decomp: "DomainDecomposition"):
-        
-        self.field = field
-        self.grid_type = grid_type
-        self.bc_type = bc_type
-        self.bc_vals = bc_vals
-        self.bc_functions = bc_functions
-        self.decomp = decomp
-
-        if self.grid_type == 'P2':
-            self._compute_zoom_factors()
[…]
