### 2026-06-04 34c1b9d
WIP: quad field system change part one
