+++ GaPFlow/solver_fem/fieldspec.py
@@ -1,6 +1,31 @@
 from dataclasses import dataclass
 
 
+def resolve_source(problem, source: str):
+    obj = problem
+    for part in source.split('.'):
+        if '[' in part:
+            attr, key = part.split('[')
+            obj = getattr(obj, attr)[key.strip('"]')]
+        else:
+            obj = getattr(obj, part)
+    return obj
+
+
+def categorize_registry_fields():
+    """Split QUAD_FIELD_REGISTRY into nodal, scalar, and computed lists."""
+    nodal, scalar, computed = [], [], []
+    for name, entry in QUAD_FIELD_REGISTRY.items():
+        t = entry['type']
+        if t == 'nodal':
+            nodal.append(name)
+        elif t == 'scalar':
+            scalar.append(name)
+        elif t == 'computed' and entry['source'] is not None:
+            computed.append(name)
+    return nodal, scalar, computed
+
+
 @dataclass
 class FieldSpec:
     """Single source of truth for one active variable or residual."""
@@ -27,39 +52,83 @@ RES_GRID = {
     'R_oss':      'P1',
 }
 
-# Nodal field name lists (P1 may be extended by theta/xi)
-NODAL_P1 = ['rho', 'p', 'h', 'dh_dx', 'dh_dy']
+NODAL_P1 = ['rho', 'p']
 NODAL_P2 = ['jx', 'jy']
 
-# Quadrature field name sets
-BASE_FIELDS = {
-    'rho', 'jx', 'jy',
-    'p', 'h', 'dh_dx', 'dh_dy', 'eta',
-    'U_bot', 'V_bot', 'U_top', 'V_top', 'Ls',
-    'dp_drho', 'drho_dp', 'd2p_drho2',
-    'd_dx_jx', 'd_dy_jy',
-    'p_prev', 'jx_prev', 'jy_prev',
-    'force_x', 'force_y',
-}
 
-STRESS_FIELDS = {
-    'tau_xz', 'dtau_xz_drho', 'dtau_xz_djx',
-    'tau_xz_bot', 'dtau_xz_bot_drho', 'dtau_xz_bot_djx',
-    'tau_yz', 'dtau_yz_drho', 'dtau_yz_djy',
-    'tau_yz_bot', 'dtau_yz_bot_drho', 'dtau_yz_bot_djy',
-}
+# Shared arg lists for groups of computed fields with identical inputs
+_ARGS_XZ  = ['rho', 'jx', 'jy', 'h', 'dh_dx', 'U_bot', 'V_bot', 'U_top', 'V_top', 'Ls', 'theta', 'd_dx_p', 'd_dy_p']
+_ARGS_YZ  = ['rho', 'jx', 'jy', 'h', 'dh_dy', 'U_bot', 'V_bot', 'U_top', 'V_top', 'Ls', 'theta', 'd_dx_p', 'd_dy_p']
+_ARGS_T   = ['rho', 'jx', 'jy', 'E']
+_ARGS_S   = ['h', 'eta', 'rho', 'E', 'jx', 'jy', 'U_bot', 'V_bot', 'Tb_top', 'Tb_bot']
 
-ENERGY_FIELDS = {
-    'E', 'Tb_top', 'Tb_bot',
-    'T', 'dT_drho', 'dT_djx', 'dT_djy', 'dT_dE',
-    'S', 'dS_drho', 'dS_djx', 'dS_djy', 'dS_dE',
-    'E_prev',
-}
+QUAD_FIELD_REGISTRY = {
+    # ------------------------------------------------------------------
+    # nodal: interpolated from nodal field to quad points
+    # ------------------------------------------------------------------
+    'h':      {'type': 'nodal', 'source': 'topo.h'},
+    'dh_dx':  {'type': 'nodal', 'source': 'topo.dh_dx'},
[…]
