+++ GaPFlow/solver_fem/terms.py
@@ -32,7 +32,7 @@ import numpy.typing as npt
 NDArray = npt.NDArray[np.floating]
 
 
-class NonLinearTerm():
+class Term():
     def __init__(self,
                  name: str,
                  description: str,
@@ -93,38 +93,34 @@ class NonLinearTerm():
         return (self.depvar_deriv, self.test_deriv)
 
 
-from .terms_theta import (  # noqa: F401
-    R11x_fb, R11y_fb, R11Sx_fb, R11Sy_fb,
-    R11x_fb_corr, R11y_fb_corr, R11Sx_fb_corr, R11Sy_fb_corr,
-    R_FB, R1STx, R1STy, R24x_fb, R24y_fb,
-)
-from .terms_oss import _OSS_TERMS  # noqa: F401
+from .terms_theta import THETA_TERMS_MASS, R_fb, THETA_TERMS_WALL_STRESS
+from .terms_oss import OSS_TERMS
 
 # -----------------------------------------------------------------------------
 # Mass equation terms (R1*)
 # -----------------------------------------------------------------------------
 
-R11x = NonLinearTerm(
+R11x = Term(
     name='R11x',
     description='flux divergence x',
     res='mass',
     dep_vars=['jx'],
-    dep_vals=['dp_drho'],
+    dep_vals=['dp_drho', 'dx_jx'],
     fun=lambda ctx: lambda jx: -ctx['dp_drho']() * jx,
     der_funs=[lambda ctx: lambda jx: -ctx['dp_drho']()],
     trial_deriv='x')
 
-R11y = NonLinearTerm(
+R11y = Term(
     name='R11y',
     description='flux divergence y',
     res='mass',
     dep_vars=['jy'],
-    dep_vals=['dp_drho'],
+    dep_vals=['dp_drho', 'd_dy_jy'],
     fun=lambda ctx: lambda jy: -ctx['dp_drho']() * jy,
     der_funs=[lambda ctx: lambda jy: -ctx['dp_drho']()],
     trial_deriv='y')
 
-R11Sx = NonLinearTerm(
+R11Sx = Term(
     name='R11Sx',
     description='flux divergence height source x',
     res='mass',
@@ -133,7 +129,7 @@ R11Sx = NonLinearTerm(
     fun=lambda ctx: lambda jx: -ctx['dp_drho']() * (1 / ctx['h']()) * ctx['dh_dx']() * jx,
     der_funs=[lambda ctx: lambda jx: -ctx['dp_drho']() * (1 / ctx['h']()) * ctx['dh_dx']()])
 
-R11Sy = NonLinearTerm(
+R11Sy = Term(
     name='R11Sy',
     description='flux divergence height source y',
     res='mass',
@@ -142,7 +138,7 @@ R11Sy = NonLinearTerm(
     fun=lambda ctx: lambda jy: -ctx['dp_drho']() * (1 / ctx['h']()) * ctx['dh_dy']() * jy,
     der_funs=[lambda ctx: lambda jy: -ctx['dp_drho']() * (1 / ctx['h']()) * ctx['dh_dy']()])
 
-R11x_corr = NonLinearTerm(
+R11x_corr = Term(
     name='R11x_corr',
     description='flux divergence x Jacobian correction (d(dp/drho)/dp)',
     res='mass',
@@ -151,7 +147,7 @@ R11x_corr = NonLinearTerm(
     fun=lambda ctx: lambda p: np.zeros_like(p),
     der_funs=[lambda ctx: lambda p: -ctx['d2p_drho2']() * ctx['drho_dp']() * ctx['d_dx_jx']()])
 
[…]
