### 2026-06-04 d507188
WIP: quad field system refactored and fem test overhaul
