### 2026-06-05 fa03107
WIP: quad field and build_grad refactor done
