+++ GaPFlow/solver_fem/terms.py
@@ -572,6 +572,7 @@ def get_active_terms(fem_solver: dict) -> List[Term]:
     - wall_shear_work:    Wall stress work / shear heating (R34)
     """
     physics = fem_solver['physics']
+    stab = fem_solver['stabilization']
     cavitation = fem_solver['equations']['cavitation']
 
     if cavitation:
@@ -580,7 +581,7 @@ def get_active_terms(fem_solver: dict) -> List[Term]:
         terms = [R11x, R11y, R11Sx, R11Sy, R11x_corr, R11y_corr,
                  R1T, R21x, R21y, R2Tx, R2Ty]
 
-    if cavitation and physics['oss_theta']:
+    if cavitation and stab['oss']:
         terms += OSS_TERMS
 
     if physics['gap_shear']:
