+++ tests/test_fem_analytic.py
@@ -31,7 +31,7 @@ from GaPFlow.problem import Problem
 from GaPFlow.models.pressure import eos_pressure
 
 CONFIG_DIR = Path(__file__).parent / "configs"
-DATA_DIR = Path(__file__).parent / "data"
+DATA_DIR = Path(__file__).parent / "ref_data"
 
 
 # =============================================================================
