### 2026-06-05 8bb2909
MAINT: clean-up and resolve flake8 violations
