+++ GaPFlow/solver_fem/grid_index.py
@@ -57,19 +57,19 @@ class GridIndexManager:
 
         # node counts
         Nx_P1, Ny_P1 = decomp.nb_subdomain_grid_pts
-        self.Nx_P1_inner  = Nx_P1
-        self.Ny_P1_inner  = Ny_P1
+        self.Nx_P1_inner = Nx_P1
+        self.Ny_P1_inner = Ny_P1
         self.Nx_P1_padded = Nx_P1 + 2
         self.Ny_P1_padded = Ny_P1 + 2
 
         Nx_P2, Ny_P2 = decomp.nb_subdomain_grid_pts_P2
-        self.Nx_P2_inner  = Nx_P2
-        self.Ny_P2_inner  = Ny_P2
+        self.Nx_P2_inner = Nx_P2
+        self.Ny_P2_inner = Ny_P2
         self.Nx_P2_padded = Nx_P2 + 4
         self.Ny_P2_padded = Ny_P2 + 4
 
         # square counts
-        self.sq_per_row = self.Nx_P1_padded -1
+        self.sq_per_row = self.Nx_P1_padded - 1
         self.sq_per_col = self.Ny_P1_padded - 1
         self.nb_sq = self.sq_per_row * self.sq_per_col
         sq_idx = np.arange(self.nb_sq)
@@ -114,7 +114,7 @@ class GridIndexManager:
         - serial + periodic: wrap around (both ghost layers)
         - inter-subdomain ghost nodes: new sequential indices
         - Neumann ghost nodes: forward both layers to adjacent inner node
-        
+
         Effectively, only Dirichlet ghost nodes remain -1.
 
         Parameters
@@ -135,10 +135,14 @@ class GridIndexManager:
 
         # fill inter-subdomain ghost nodes with new indices (non-boundary)
         boundary_P1 = np.zeros((self.Nx_P1_padded, self.Ny_P1_padded), dtype=bool)
-        if decomp.bc_at_W: boundary_P1[0,  :] = True
-        if decomp.bc_at_E: boundary_P1[-1, :] = True
-        if decomp.bc_at_S: boundary_P1[:,  0] = True
-        if decomp.bc_at_N: boundary_P1[:, -1] = True
+        if decomp.bc_at_W:
+            boundary_P1[0, :] = True
+        if decomp.bc_at_E:
+            boundary_P1[-1, :] = True
+        if decomp.bc_at_S:
+            boundary_P1[:, 0] = True
+        if decomp.bc_at_N:
+            boundary_P1[:, -1] = True
         ghost_coords = np.argwhere((mask == -1) & ~boundary_P1)
         nb_inner_P1 = self.Nx_P1_inner * self.Ny_P1_inner
         mask[ghost_coords[:, 0], ghost_coords[:, 1]] = np.arange(
@@ -198,7 +202,7 @@ class GridIndexManager:
         - serial + periodic: wrap around (both ghost layers)
         - inter-subdomain ghost nodes: new sequential indices
         - Neumann ghost nodes: forward both layers to adjacent inner node
-        
+
         Effectively, only Dirichlet ghost nodes remain -1.
 
         Parameters
@@ -223,10 +227,14 @@ class GridIndexManager:
 
         # fill inter-subdomain ghost nodes with new indices (non-boundary)
         boundary_P2 = np.zeros((self.Nx_P2_padded, self.Ny_P2_padded), dtype=bool)
-        if decomp.bc_at_W: boundary_P2[:2,  :] = True
-        if decomp.bc_at_E: boundary_P2[-2:, :] = True
-        if decomp.bc_at_S: boundary_P2[:,  :2] = True
-        if decomp.bc_at_N: boundary_P2[:, -2:] = True
+        if decomp.bc_at_W:
+            boundary_P2[:2, :] = True
[…]
