### 2026-06-06 fbd9b3b
FIX: fallback scipy solver
