### 2026-06-10 b68cda4
FIX: lammps licence access failure
