+++ doc/tutorials/D_wish_list.md
@@ -4,6 +4,11 @@
 - custom equation of state
 - custom wall stress box
 
+## Elastic Deformation
+
+- start with dry contact solution
+- sensible limiter for pressure changes (to limit increase of deform and piezo effect)
+
 # Merge
 
 - wrapper()
