### 2026-06-10 e045aea
MAINT: modify density calculation from MD z-profiles
