### 2026-06-11 59a5959
FIX: set submodules to recursive
