+++ .github/workflows/test.yaml
@@ -20,7 +20,7 @@ jobs:
       uses: actions/checkout@v5
       with:
         fetch-depth: 0
-        submodules: 'true'
+        submodules: recursive
 
     - name: Install MPI
       run: |
