### 2026-06-12 7a80f25
ENH: integrated GP-based gradients into fem solver
