+++ GaPFlow/viz/animations.py
@@ -309,11 +309,6 @@ def _create_animation_1d_gp(filename, tol_p=None, tol_t=None):
         ax[1, 1].cla()
         ax[1, 2].cla()
 
-        _tol_t = tol_t[i] if tol_t is not None else None
-        _tol_p = tol_p[i] if tol_p is not None else None
-        _tol_t_max = tol_t_max if tol_t_max is not None else np.sqrt(vt[i, 1:-1, ny // 2]).max()
-        _tol_p_max = tol_p_max if tol_p_max is not None else np.sqrt(vp[i, 1:-1, ny // 2]).max()
-
         # Pressure
         _plot_gp(ax[1, 0], x, p[i, :, ci], vp[i, :, ci], tol=tol_p[i], color=color_p)
 
