### 2026-06-12 30f92b4
WIP: gap height correction rough walls
