### 2026-06-12 1aa2e9e
WIP: modify gap height calculation
