### 2026-06-12 8cea97f
MAINT: fine-tuning MD measurements
